feat: add track groups
