+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tonsolutions/telemetree-pixel",
-  "version": "1.2.6",
+  "version": "1.2.8",
   "scripts": {
     "build": "tsc && vite build",
     "start": "vite",
