+++ telemetree-pixel.js
@@ -1,4 +1,4 @@
-(function(a0,x0){typeof exports=="object"&&typeof module<"u"?module.exports=x0():typeof define=="function"&&define.amd?define(x0):(a0=typeof globalThis<"u"?globalThis:a0||self,a0.telemetree=x0())})(this,function(){"use strict";var Ui=Object.defineProperty;var Vi=(a0,x0,v0)=>x0 in a0?Ui(a0,x0,{enumerable:!0,configurable:!0,writable:!0,value:v0}):a0[x0]=v0;var i0=(a0,x0,v0)=>(Vi(a0,typeof x0!="symbol"?x0+"":x0,v0),v0);var Qe,Ye;const a0="https://ebn.telemetree.io/public-api/v1/client/config",v0="[TS]",hr=" | ";var q=(i=>(i.Click="Click",i.PageView="Pageview",i.Closed="Closed",i.MainButtonPressed="MainButtonPressed",i.SettingsButtonPressed="SettingsButtonPressed",i.InvoiceClosed="InvoiceClosed",i.InvoiceOpened="Invoice opened",i.PopupClosed="PopupClosed",i.ClipboardTextReceived="ClipboardTextReceived",i.WriteAccessRequested="WriteAccessRequested",i.QRTextReceived="QRTextReceived",i.PhoneRequested="PhoneRequested",i.BackButtonPressed="BackButtonPressed",i.SecondaryButtonPressed="SeconadaryButtonPressed",i.PreparedMessageSent="PreparedMessageSent",i.FullScreenChanged="FullScreenChanged",i.HomeScreenAdded="HomeScreenAdded",i.HomeScreenChecked="HomeScreenChecked",i.EmojiStatusSet="EmojiStatusSet",i.LocationChecked="LocationChecked",i.LocationRequested="LocationRequested",i.AccelerometerStarted="AccelerometerStarted",i.AccelerometerStopped="AccelerometerStopped",i.AccelerometerChanged="AccelerometerChanged",i.DeviceOrientationStarted="DeviceOrientationStarted",i.DeviceOrientationStopped="DeviceOrientationStopped",i.DeviceOrientationChanged="DeviceOrientationChanged",i.DeviceOrientationFailed="DeviceOrientationFailed",i.GyroscropeStarted="GyroscropeStarted",i.GyroscropeStopped="GyroscropeStopped",i.GyroscropeChanged="GyroscropeChanged",i.GyroscropeFailed="GyroscropeFailed",i.WebAppRequestFullscreen="Fullscreen on",i.WebAppExitFullscreen="Fullscreen off",i.SwitchInlineQuery="Inline query opened",i.ShareToStory="Story shared",i.SessionStart="Session start",i.SessionEnd="Session end",i.OpenLink="Link opened",i.OpenTgLink="Telegram link opened",i.Wallet="Wallet",i.WalletConnectStarted="Wallet connect started",i.WalletConnectError="Wallet connect error",i.WalletConnectionRestoringStarted="Wallet connection restoring started",i.WalletConnectionRestored="Wallet connection restored",i.WalletConnectionRestoreError="Wallet connection restore error",i.WalletDisconnected="Wallet disconnected",i.TransactionSentForSignature="Transaction sent for signature",i.TransactionSigned="Transaction signed",i.TransactionSigningFailed="Transaction signing failed",i))(q||{}),J=(i=>(i.WalletConnectStarted="connection-started",i.WalletConnectSuccess="connection-completed",i.WalletConnectError="connection-error",i.ConnectionRestoringStarted="connection-restoring-started",i.ConnectionRestoringSuccess="connection-restoring-completed",i.ConnectionRestoringError="connection-restoring-error",i.WalletDisconnect="disconnection",i.TransactionSentForSignature="transaction-sent-for-signature",i.TransactionSigned="transaction-signed",i.TransactionSigningFailed="transaction-signing-failed",i))(J||{});class ur{constructor(e){i0(this,"eventsQueue",[]);i0(this,"isProcessing",!1);this.client=e}push(e){this.eventsQueue.push(e),this.processQueue()}processQueue(){this.isProcessing===!0||this.client.getConfig()===null||(this.isProcessing=!0,this.flush().then(()=>{this.isProcessing=!1,this.eventsQueue.length>0&&this.processQueue()}))}async flush(){const e=this.eventsQueue;if(this.eventsQueue=[],e.length===0)return;const t=e.map(r=>this.client.processEvent.bind(this.client)(r));await Promise.all(t)}}const lr=()=>Math.floor(Date.now()/1e3).toString(),Gt=()=>Date.now().toString();var W=typeof globalThis<"u"?globalThis:typeof window<"u"?window:typeof global<"u"?global:typeof self<"u"?self:{};function dr(i){return i&&i.__esModule&&Object.prototype.hasOwnProperty.call(i,"default")?i.default:i}function vr(i){if(i.__esModule)return i;var e=i.default;if(typeof e=="function"){var t=function r(){return this instanceof r?Reflect.construct(e,arguments,this.constructor):e.apply(this,arguments)};t.prototype=e.prototype}else t={};return Object.defineProperty(t,"__esModule",{value:!0}),Object.keys(i).forEach(function(r){var n=Object.getOwnPropertyDescriptor(i,r);Object.defineProperty(t,r,n.get?n:{enumerable:!0,get:function(){return i[r]}})}),t}var Zt={exports:{}};function pr(i){throw new Error('Could not dynamically require "'+i+'". Please configure the dynamicRequireTargets or/and ignoreDynamicRequires option of @rollup/plugin-commonjs appropriately for this require call to work.')}var et={exports:{}};const gr=vr(Object.freeze(Object.defineProperty({__proto__:null,default:{}},Symbol.toStringTag,{value:"Module"})));var Xt;function V(){return Xt||(Xt=1,function(i,e){(function(t,r){i.exports=r()})(W,function(){var t=t||function(r,n){var s;if(typeof window<"u"&&window.crypto&&(s=window.crypto),typeof self<"u"&&self.crypto&&(s=self.crypto),typeof globalThis<"u"&&globalThis.crypto&&(s=globalThis.crypto),!s&&typeof window<"u"&&window.msCrypto&&(s=window.msCrypto),!s&&typeof W<"u"&&W.crypto&&(s=W.crypto),!s&&typeof pr=="function")try{s=gr}catch{}var x=function(){if(s){if(typeof s.getRandomValues=="function")try{return s.getRandomValues(new Uint32Array(1))[0]}catch{}if(typeof s.randomBytes=="function")try{return s.randomBytes(4).readInt32LE()}catch{}}throw new Error("Native crypto module could not be used to get secure random number.")},o=Object.create||function(){function d(){}return function(g){var B;return d.prototype=g,B=new d,d.prototype=null,B}}(),f={},a=f.lib={},c=a.Base=function(){return{extend:function(d){var g=o(this);return d&&g.mixIn(d),(!g.hasOwnProperty("init")||this.init===g.init)&&(g.init=function(){g.$super.init.apply(this,arguments)}),g.init.prototype=g,g.$super=this,g},create:function(){var d=this.extend();return d.init.apply(d,arguments),d},init:function(){},mixIn:function(d){for(var g in d)d.hasOwnProperty(g)&&(this[g]=d[g]);d.hasOwnProperty("toString")&&(this.toString=d.toString)},clone:function(){return this.init.prototype.extend(this)}}}(),E=a.WordArray=c.extend({init:function(d,g){d=this.words=d||[],g!=n?this.sigBytes=g:this.sigBytes=d.length*4},toString:function(d){return(d||u).stringify(this)},concat:function(d){var g=this.words,B=d.words,F=this.sigBytes,A=d.sigBytes;if(this.clamp(),F%4)for(var _=0;_<A;_++){var b=B[_>>>2]>>>24-_%4*8&255;g[F+_>>>2]|=b<<24-(F+_)%4*8}else for(var P=0;P<A;P+=4)g[F+P>>>2]=B[P>>>2];return this.sigBytes+=A,this},clamp:function(){var d=this.words,g=this.sigBytes;d[g>>>2]&=4294967295<<32-g%4*8,d.length=r.ceil(g/4)},clone:function(){var d=c.clone.call(this);return d.words=this.words.slice(0),d},random:function(d){for(var g=[],B=0;B<d;B+=4)g.push(x());return new E.init(g,d)}}),h=f.enc={},u=h.Hex={stringify:function(d){for(var g=d.words,B=d.sigBytes,F=[],A=0;A<B;A++){var _=g[A>>>2]>>>24-A%4*8&255;F.push((_>>>4).toString(16)),F.push((_&15).toString(16))}return F.join("")},parse:function(d){for(var g=d.length,B=[],F=0;F<g;F+=2)B[F>>>3]|=parseInt(d.substr(F,2),16)<<24-F%8*4;return new E.init(B,g/2)}},v=h.Latin1={stringify:function(d){for(var g=d.words,B=d.sigBytes,F=[],A=0;A<B;A++){var _=g[A>>>2]>>>24-A%4*8&255;F.push(String.fromCharCode(_))}return F.join("")},parse:function(d){for(var g=d.length,B=[],F=0;F<g;F++)B[F>>>2]|=(d.charCodeAt(F)&255)<<24-F%4*8;return new E.init(B,g)}},p=h.Utf8={stringify:function(d){try{return decodeURIComponent(escape(v.stringify(d)))}catch{throw new Error("Malformed UTF-8 data")}},parse:function(d){return v.parse(unescape(encodeURIComponent(d)))}},l=a.BufferedBlockAlgorithm=c.extend({reset:function(){this._data=new E.init,this._nDataBytes=0},_append:function(d){typeof d=="string"&&(d=p.parse(d)),this._data.concat(d),this._nDataBytes+=d.sigBytes},_process:function(d){var g,B=this._data,F=B.words,A=B.sigBytes,_=this.blockSize,b=_*4,P=A/b;d?P=r.ceil(P):P=r.max((P|0)-this._minBufferSize,0);var D=P*_,y=r.min(D*4,A);if(D){for(var T=0;T<D;T+=_)this._doProcessBlock(F,T);g=F.splice(0,D),B.sigBytes-=y}return new E.init(g,y)},clone:function(){var d=c.clone.call(this);return d._data=this._data.clone(),d},_minBufferSize:0});a.Hasher=l.extend({cfg:c.extend(),init:function(d){this.cfg=this.cfg.extend(d),this.reset()},reset:function(){l.reset.call(this),this._doReset()},update:function(d){return this._append(d),this._process(),this},finalize:function(d){d&&this._append(d);var g=this._doFinalize();return g},blockSize:16,_createHelper:function(d){return function(g,B){return new d.init(B).finalize(g)}},_createHmacHelper:function(d){return function(g,B){return new C.HMAC.init(d,B).finalize(g)}}});var C=f.algo={};return f}(Math);return t})}(et)),et.exports}var rt={exports:{}},Qt;function L0(){return Qt||(Qt=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.Base,o=s.WordArray,f=n.x64={};f.Word=x.extend({init:function(a,c){this.high=a,this.low=c}}),f.WordArray=x.extend({init:function(a,c){a=this.words=a||[],c!=r?this.sigBytes=c:this.sigBytes=a.length*8},toX32:function(){for(var a=this.words,c=a.length,E=[],h=0;h<c;h++){var u=a[h];E.push(u.high),E.push(u.low)}return o.create(E,this.sigBytes)},clone:function(){for(var a=x.clone.call(this),c=a.words=this.words.slice(0),E=c.length,h=0;h<E;h++)c[h]=c[h].clone();return a}})}(),t})}(rt)),rt.exports}var it={exports:{}},Yt;function Er(){return Yt||(Yt=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){if(typeof ArrayBuffer=="function"){var r=t,n=r.lib,s=n.WordArray,x=s.init,o=s.init=function(f){if(f instanceof ArrayBuffer&&(f=new Uint8Array(f)),(f instanceof Int8Array||typeof Uint8ClampedArray<"u"&&f instanceof Uint8ClampedArray||f instanceof Int16Array||f instanceof Uint16Array||f instanceof Int32Array||f instanceof Uint32Array||f instanceof Float32Array||f instanceof Float64Array)&&(f=new Uint8Array(f.buffer,f.byteOffset,f.byteLength)),f instanceof Uint8Array){for(var a=f.byteLength,c=[],E=0;E<a;E++)c[E>>>2]|=f[E]<<24-E%4*8;x.call(this,c,a)}else x.apply(this,arguments)};o.prototype=s}}(),t.lib.WordArray})}(it)),it.exports}var nt={exports:{}},Jt;function Br(){return Jt||(Jt=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Utf16=x.Utf16BE={stringify:function(f){for(var a=f.words,c=f.sigBytes,E=[],h=0;h<c;h+=2){var u=a[h>>>2]>>>16-h%4*8&65535;E.push(String.fromCharCode(u))}return E.join("")},parse:function(f){for(var a=f.length,c=[],E=0;E<a;E++)c[E>>>1]|=f.charCodeAt(E)<<16-E%2*16;return s.create(c,a*2)}},x.Utf16LE={stringify:function(f){for(var a=f.words,c=f.sigBytes,E=[],h=0;h<c;h+=2){var u=o(a[h>>>2]>>>16-h%4*8&65535);E.push(String.fromCharCode(u))}return E.join("")},parse:function(f){for(var a=f.length,c=[],E=0;E<a;E++)c[E>>>1]|=o(f.charCodeAt(E)<<16-E%2*16);return s.create(c,a*2)}};function o(f){return f<<8&4278255360|f>>>8&16711935}}(),t.enc.Utf16})}(nt)),nt.exports}var st={exports:{}},te;function D0(){return te||(te=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Base64={stringify:function(f){var a=f.words,c=f.sigBytes,E=this._map;f.clamp();for(var h=[],u=0;u<c;u+=3)for(var v=a[u>>>2]>>>24-u%4*8&255,p=a[u+1>>>2]>>>24-(u+1)%4*8&255,l=a[u+2>>>2]>>>24-(u+2)%4*8&255,C=v<<16|p<<8|l,d=0;d<4&&u+d*.75<c;d++)h.push(E.charAt(C>>>6*(3-d)&63));var g=E.charAt(64);if(g)for(;h.length%4;)h.push(g);return h.join("")},parse:function(f){var a=f.length,c=this._map,E=this._reverseMap;if(!E){E=this._reverseMap=[];for(var h=0;h<c.length;h++)E[c.charCodeAt(h)]=h}var u=c.charAt(64);if(u){var v=f.indexOf(u);v!==-1&&(a=v)}return o(f,a,E)},_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/="};function o(f,a,c){for(var E=[],h=0,u=0;u<a;u++)if(u%4){var v=c[f.charCodeAt(u-1)]<<u%4*2,p=c[f.charCodeAt(u)]>>>6-u%4*2,l=v|p;E[h>>>2]|=l<<24-h%4*8,h++}return s.create(E,h)}}(),t.enc.Base64})}(st)),st.exports}var ot={exports:{}},ee;function Ar(){return ee||(ee=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Base64url={stringify:function(f,a){a===void 0&&(a=!0);var c=f.words,E=f.sigBytes,h=a?this._safe_map:this._map;f.clamp();for(var u=[],v=0;v<E;v+=3)for(var p=c[v>>>2]>>>24-v%4*8&255,l=c[v+1>>>2]>>>24-(v+1)%4*8&255,C=c[v+2>>>2]>>>24-(v+2)%4*8&255,d=p<<16|l<<8|C,g=0;g<4&&v+g*.75<E;g++)u.push(h.charAt(d>>>6*(3-g)&63));var B=h.charAt(64);if(B)for(;u.length%4;)u.push(B);return u.join("")},parse:function(f,a){a===void 0&&(a=!0);var c=f.length,E=a?this._safe_map:this._map,h=this._reverseMap;if(!h){h=this._reverseMap=[];for(var u=0;u<E.length;u++)h[E.charCodeAt(u)]=u}var v=E.charAt(64);if(v){var p=f.indexOf(v);p!==-1&&(c=p)}return o(f,c,h)},_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",_safe_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"};function o(f,a,c){for(var E=[],h=0,u=0;u<a;u++)if(u%4){var v=c[f.charCodeAt(u-1)]<<u%4*2,p=c[f.charCodeAt(u)]>>>6-u%4*2,l=v|p;E[h>>>2]|=l<<24-h%4*8,h++}return s.create(E,h)}}(),t.enc.Base64url})}(ot)),ot.exports}var at={exports:{}},re;function y0(){return re||(re=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=[];(function(){for(var p=0;p<64;p++)a[p]=r.abs(r.sin(p+1))*4294967296|0})();var c=f.MD5=o.extend({_doReset:function(){this._hash=new x.init([1732584193,4023233417,2562383102,271733878])},_doProcessBlock:function(p,l){for(var C=0;C<16;C++){var d=l+C,g=p[d];p[d]=(g<<8|g>>>24)&16711935|(g<<24|g>>>8)&4278255360}var B=this._hash.words,F=p[l+0],A=p[l+1],_=p[l+2],b=p[l+3],P=p[l+4],D=p[l+5],y=p[l+6],T=p[l+7],w=p[l+8],H=p[l+9],N=p[l+10],L=p[l+11],Z=p[l+12],K=p[l+13],j=p[l+14],$=p[l+15],S=B[0],k=B[1],O=B[2],R=B[3];S=E(S,k,O,R,F,7,a[0]),R=E(R,S,k,O,A,12,a[1]),O=E(O,R,S,k,_,17,a[2]),k=E(k,O,R,S,b,22,a[3]),S=E(S,k,O,R,P,7,a[4]),R=E(R,S,k,O,D,12,a[5]),O=E(O,R,S,k,y,17,a[6]),k=E(k,O,R,S,T,22,a[7]),S=E(S,k,O,R,w,7,a[8]),R=E(R,S,k,O,H,12,a[9]),O=E(O,R,S,k,N,17,a[10]),k=E(k,O,R,S,L,22,a[11]),S=E(S,k,O,R,Z,7,a[12]),R=E(R,S,k,O,K,12,a[13]),O=E(O,R,S,k,j,17,a[14]),k=E(k,O,R,S,$,22,a[15]),S=h(S,k,O,R,A,5,a[16]),R=h(R,S,k,O,y,9,a[17]),O=h(O,R,S,k,L,14,a[18]),k=h(k,O,R,S,F,20,a[19]),S=h(S,k,O,R,D,5,a[20]),R=h(R,S,k,O,N,9,a[21]),O=h(O,R,S,k,$,14,a[22]),k=h(k,O,R,S,P,20,a[23]),S=h(S,k,O,R,H,5,a[24]),R=h(R,S,k,O,j,9,a[25]),O=h(O,R,S,k,b,14,a[26]),k=h(k,O,R,S,w,20,a[27]),S=h(S,k,O,R,K,5,a[28]),R=h(R,S,k,O,_,9,a[29]),O=h(O,R,S,k,T,14,a[30]),k=h(k,O,R,S,Z,20,a[31]),S=u(S,k,O,R,D,4,a[32]),R=u(R,S,k,O,w,11,a[33]),O=u(O,R,S,k,L,16,a[34]),k=u(k,O,R,S,j,23,a[35]),S=u(S,k,O,R,A,4,a[36]),R=u(R,S,k,O,P,11,a[37]),O=u(O,R,S,k,T,16,a[38]),k=u(k,O,R,S,N,23,a[39]),S=u(S,k,O,R,K,4,a[40]),R=u(R,S,k,O,F,11,a[41]),O=u(O,R,S,k,b,16,a[42]),k=u(k,O,R,S,y,23,a[43]),S=u(S,k,O,R,H,4,a[44]),R=u(R,S,k,O,Z,11,a[45]),O=u(O,R,S,k,$,16,a[46]),k=u(k,O,R,S,_,23,a[47]),S=v(S,k,O,R,F,6,a[48]),R=v(R,S,k,O,T,10,a[49]),O=v(O,R,S,k,j,15,a[50]),k=v(k,O,R,S,D,21,a[51]),S=v(S,k,O,R,Z,6,a[52]),R=v(R,S,k,O,b,10,a[53]),O=v(O,R,S,k,N,15,a[54]),k=v(k,O,R,S,A,21,a[55]),S=v(S,k,O,R,w,6,a[56]),R=v(R,S,k,O,$,10,a[57]),O=v(O,R,S,k,y,15,a[58]),k=v(k,O,R,S,K,21,a[59]),S=v(S,k,O,R,P,6,a[60]),R=v(R,S,k,O,L,10,a[61]),O=v(O,R,S,k,_,15,a[62]),k=v(k,O,R,S,H,21,a[63]),B[0]=B[0]+S|0,B[1]=B[1]+k|0,B[2]=B[2]+O|0,B[3]=B[3]+R|0},_doFinalize:function(){var p=this._data,l=p.words,C=this._nDataBytes*8,d=p.sigBytes*8;l[d>>>5]|=128<<24-d%32;var g=r.floor(C/4294967296),B=C;l[(d+64>>>9<<4)+15]=(g<<8|g>>>24)&16711935|(g<<24|g>>>8)&4278255360,l[(d+64>>>9<<4)+14]=(B<<8|B>>>24)&16711935|(B<<24|B>>>8)&4278255360,p.sigBytes=(l.length+1)*4,this._process();for(var F=this._hash,A=F.words,_=0;_<4;_++){var b=A[_];A[_]=(b<<8|b>>>24)&16711935|(b<<24|b>>>8)&4278255360}return F},clone:function(){var p=o.clone.call(this);return p._hash=this._hash.clone(),p}});function E(p,l,C,d,g,B,F){var A=p+(l&C|~l&d)+g+F;return(A<<B|A>>>32-B)+l}function h(p,l,C,d,g,B,F){var A=p+(l&d|C&~d)+g+F;return(A<<B|A>>>32-B)+l}function u(p,l,C,d,g,B,F){var A=p+(l^C^d)+g+F;return(A<<B|A>>>32-B)+l}function v(p,l,C,d,g,B,F){var A=p+(C^(l|~d))+g+F;return(A<<B|A>>>32-B)+l}n.MD5=o._createHelper(c),n.HmacMD5=o._createHmacHelper(c)}(Math),t.MD5})}(at)),at.exports}var xt={exports:{}},ie;function ne(){return ie||(ie=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=n.Hasher,o=r.algo,f=[],a=o.SHA1=x.extend({_doReset:function(){this._hash=new s.init([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(c,E){for(var h=this._hash.words,u=h[0],v=h[1],p=h[2],l=h[3],C=h[4],d=0;d<80;d++){if(d<16)f[d]=c[E+d]|0;else{var g=f[d-3]^f[d-8]^f[d-14]^f[d-16];f[d]=g<<1|g>>>31}var B=(u<<5|u>>>27)+C+f[d];d<20?B+=(v&p|~v&l)+1518500249:d<40?B+=(v^p^l)+1859775393:d<60?B+=(v&p|v&l|p&l)-1894007588:B+=(v^p^l)-899497514,C=l,l=p,p=v<<30|v>>>2,v=u,u=B}h[0]=h[0]+u|0,h[1]=h[1]+v|0,h[2]=h[2]+p|0,h[3]=h[3]+l|0,h[4]=h[4]+C|0},_doFinalize:function(){var c=this._data,E=c.words,h=this._nDataBytes*8,u=c.sigBytes*8;return E[u>>>5]|=128<<24-u%32,E[(u+64>>>9<<4)+14]=Math.floor(h/4294967296),E[(u+64>>>9<<4)+15]=h,c.sigBytes=E.length*4,this._process(),this._hash},clone:function(){var c=x.clone.call(this);return c._hash=this._hash.clone(),c}});r.SHA1=x._createHelper(a),r.HmacSHA1=x._createHmacHelper(a)}(),t.SHA1})}(xt)),xt.exports}var ft={exports:{}},se;function ct(){return se||(se=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=[],c=[];(function(){function u(C){for(var d=r.sqrt(C),g=2;g<=d;g++)if(!(C%g))return!1;return!0}function v(C){return(C-(C|0))*4294967296|0}for(var p=2,l=0;l<64;)u(p)&&(l<8&&(a[l]=v(r.pow(p,1/2))),c[l]=v(r.pow(p,1/3)),l++),p++})();var E=[],h=f.SHA256=o.extend({_doReset:function(){this._hash=new x.init(a.slice(0))},_doProcessBlock:function(u,v){for(var p=this._hash.words,l=p[0],C=p[1],d=p[2],g=p[3],B=p[4],F=p[5],A=p[6],_=p[7],b=0;b<64;b++){if(b<16)E[b]=u[v+b]|0;else{var P=E[b-15],D=(P<<25|P>>>7)^(P<<14|P>>>18)^P>>>3,y=E[b-2],T=(y<<15|y>>>17)^(y<<13|y>>>19)^y>>>10;E[b]=D+E[b-7]+T+E[b-16]}var w=B&F^~B&A,H=l&C^l&d^C&d,N=(l<<30|l>>>2)^(l<<19|l>>>13)^(l<<10|l>>>22),L=(B<<26|B>>>6)^(B<<21|B>>>11)^(B<<7|B>>>25),Z=_+L+w+c[b]+E[b],K=N+H;_=A,A=F,F=B,B=g+Z|0,g=d,d=C,C=l,l=Z+K|0}p[0]=p[0]+l|0,p[1]=p[1]+C|0,p[2]=p[2]+d|0,p[3]=p[3]+g|0,p[4]=p[4]+B|0,p[5]=p[5]+F|0,p[6]=p[6]+A|0,p[7]=p[7]+_|0},_doFinalize:function(){var u=this._data,v=u.words,p=this._nDataBytes*8,l=u.sigBytes*8;return v[l>>>5]|=128<<24-l%32,v[(l+64>>>9<<4)+14]=r.floor(p/4294967296),v[(l+64>>>9<<4)+15]=p,u.sigBytes=v.length*4,this._process(),this._hash},clone:function(){var u=o.clone.call(this);return u._hash=this._hash.clone(),u}});n.SHA256=o._createHelper(h),n.HmacSHA256=o._createHmacHelper(h)}(Math),t.SHA256})}(ft)),ft.exports}var ht={exports:{}},oe;function Cr(){return oe||(oe=1,function(i,e){(function(t,r,n){i.exports=r(V(),ct())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.algo,o=x.SHA256,f=x.SHA224=o.extend({_doReset:function(){this._hash=new s.init([3238371032,914150663,812702999,4144912697,4290775857,1750603025,1694076839,3204075428])},_doFinalize:function(){var a=o._doFinalize.call(this);return a.sigBytes-=4,a}});r.SHA224=o._createHelper(f),r.HmacSHA224=o._createHmacHelper(f)}(),t.SHA224})}(ht)),ht.exports}var ut={exports:{}},ae;function xe(){return ae||(ae=1,function(i,e){(function(t,r,n){i.exports=r(V(),L0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Hasher,x=r.x64,o=x.Word,f=x.WordArray,a=r.algo;function c(){return o.create.apply(o,arguments)}var E=[c(1116352408,3609767458),c(1899447441,602891725),c(3049323471,3964484399),c(3921009573,2173295548),c(961987163,4081628472),c(1508970993,3053834265),c(2453635748,2937671579),c(2870763221,3664609560),c(3624381080,2734883394),c(310598401,1164996542),c(607225278,1323610764),c(1426881987,3590304994),c(1925078388,4068182383),c(2162078206,991336113),c(2614888103,633803317),c(3248222580,3479774868),c(3835390401,2666613458),c(4022224774,944711139),c(264347078,2341262773),c(604807628,2007800933),c(770255983,1495990901),c(1249150122,1856431235),c(1555081692,3175218132),c(1996064986,2198950837),c(2554220882,3999719339),c(2821834349,766784016),c(2952996808,2566594879),c(3210313671,3203337956),c(3336571891,1034457026),c(3584528711,2466948901),c(113926993,3758326383),c(338241895,168717936),c(666307205,1188179964),c(773529912,1546045734),c(1294757372,1522805485),c(1396182291,2643833823),c(1695183700,2343527390),c(1986661051,1014477480),c(2177026350,1206759142),c(2456956037,344077627),c(2730485921,1290863460),c(2820302411,3158454273),c(3259730800,3505952657),c(3345764771,106217008),c(3516065817,3606008344),c(3600352804,1432725776),c(4094571909,1467031594),c(275423344,851169720),c(430227734,3100823752),c(506948616,1363258195),c(659060556,3750685593),c(883997877,3785050280),c(958139571,3318307427),c(1322822218,3812723403),c(1537002063,2003034995),c(1747873779,3602036899),c(1955562222,1575990012),c(2024104815,1125592928),c(2227730452,2716904306),c(2361852424,442776044),c(2428436474,593698344),c(2756734187,3733110249),c(3204031479,2999351573),c(3329325298,3815920427),c(3391569614,3928383900),c(3515267271,566280711),c(3940187606,3454069534),c(4118630271,4000239992),c(116418474,1914138554),c(174292421,2731055270),c(289380356,3203993006),c(460393269,320620315),c(685471733,587496836),c(852142971,1086792851),c(1017036298,365543100),c(1126000580,2618297676),c(1288033470,3409855158),c(1501505948,4234509866),c(1607167915,987167468),c(1816402316,1246189591)],h=[];(function(){for(var v=0;v<80;v++)h[v]=c()})();var u=a.SHA512=s.extend({_doReset:function(){this._hash=new f.init([new o.init(1779033703,4089235720),new o.init(3144134277,2227873595),new o.init(1013904242,4271175723),new o.init(2773480762,1595750129),new o.init(1359893119,2917565137),new o.init(2600822924,725511199),new o.init(528734635,4215389547),new o.init(1541459225,327033209)])},_doProcessBlock:function(v,p){for(var l=this._hash.words,C=l[0],d=l[1],g=l[2],B=l[3],F=l[4],A=l[5],_=l[6],b=l[7],P=C.high,D=C.low,y=d.high,T=d.low,w=g.high,H=g.low,N=B.high,L=B.low,Z=F.high,K=F.low,j=A.high,$=A.low,S=_.high,k=_.low,O=b.high,R=b.low,X=P,G=D,r0=y,M=T,O0=w,T0=H,$t=N,I0=L,u0=Z,s0=K,Y0=j,H0=$,J0=S,q0=k,jt=O,N0=R,l0=0;l0<80;l0++){var c0,A0,tt=h[l0];if(l0<16)A0=tt.high=v[p+l0*2]|0,c0=tt.low=v[p+l0*2+1]|0;else{var Je=h[l0-15],R0=Je.high,z0=Je.low,Ti=(R0>>>1|z0<<31)^(R0>>>8|z0<<24)^R0>>>7,tr=(z0>>>1|R0<<31)^(z0>>>8|R0<<24)^(z0>>>7|R0<<25),er=h[l0-2],k0=er.high,W0=er.low,Ri=(k0>>>19|W0<<13)^(k0<<3|W0>>>29)^k0>>>6,rr=(W0>>>19|k0<<13)^(W0<<3|k0>>>29)^(W0>>>6|k0<<26),ir=h[l0-7],ki=ir.high,Pi=ir.low,nr=h[l0-16],Oi=nr.high,sr=nr.low;c0=tr+Pi,A0=Ti+ki+(c0>>>0<tr>>>0?1:0),c0=c0+rr,A0=A0+Ri+(c0>>>0<rr>>>0?1:0),c0=c0+sr,A0=A0+Oi+(c0>>>0<sr>>>0?1:0),tt.high=A0,tt.low=c0}var Ii=u0&Y0^~u0&J0,or=s0&H0^~s0&q0,Hi=X&r0^X&O0^r0&O0,qi=G&M^G&T0^M&T0,Ni=(X>>>28|G<<4)^(X<<30|G>>>2)^(X<<25|G>>>7),ar=(G>>>28|X<<4)^(G<<30|X>>>2)^(G<<25|X>>>7),zi=(u0>>>14|s0<<18)^(u0>>>18|s0<<14)^(u0<<23|s0>>>9),Wi=(s0>>>14|u0<<18)^(s0>>>18|u0<<14)^(s0<<23|u0>>>9),xr=E[l0],Li=xr.high,fr=xr.low,o0=N0+Wi,C0=jt+zi+(o0>>>0<N0>>>0?1:0),o0=o0+or,C0=C0+Ii+(o0>>>0<or>>>0?1:0),o0=o0+fr,C0=C0+Li+(o0>>>0<fr>>>0?1:0),o0=o0+c0,C0=C0+A0+(o0>>>0<c0>>>0?1:0),cr=ar+qi,Mi=Ni+Hi+(cr>>>0<ar>>>0?1:0);jt=J0,N0=q0,J0=Y0,q0=H0,Y0=u0,H0=s0,s0=I0+o0|0,u0=$t+C0+(s0>>>0<I0>>>0?1:0)|0,$t=O0,I0=T0,O0=r0,T0=M,r0=X,M=G,G=o0+cr|0,X=C0+Mi+(G>>>0<o0>>>0?1:0)|0}D=C.low=D+G,C.high=P+X+(D>>>0<G>>>0?1:0),T=d.low=T+M,d.high=y+r0+(T>>>0<M>>>0?1:0),H=g.low=H+T0,g.high=w+O0+(H>>>0<T0>>>0?1:0),L=B.low=L+I0,B.high=N+$t+(L>>>0<I0>>>0?1:0),K=F.low=K+s0,F.high=Z+u0+(K>>>0<s0>>>0?1:0),$=A.low=$+H0,A.high=j+Y0+($>>>0<H0>>>0?1:0),k=_.low=k+q0,_.high=S+J0+(k>>>0<q0>>>0?1:0),R=b.low=R+N0,b.high=O+jt+(R>>>0<N0>>>0?1:0)},_doFinalize:function(){var v=this._data,p=v.words,l=this._nDataBytes*8,C=v.sigBytes*8;p[C>>>5]|=128<<24-C%32,p[(C+128>>>10<<5)+30]=Math.floor(l/4294967296),p[(C+128>>>10<<5)+31]=l,v.sigBytes=p.length*4,this._process();var d=this._hash.toX32();return d},clone:function(){var v=s.clone.call(this);return v._hash=this._hash.clone(),v},blockSize:1024/32});r.SHA512=s._createHelper(u),r.HmacSHA512=s._createHmacHelper(u)}(),t.SHA512})}(ut)),ut.exports}var lt={exports:{}},fe;function Dr(){return fe||(fe=1,function(i,e){(function(t,r,n){i.exports=r(V(),L0(),xe())})(W,function(t){return function(){var r=t,n=r.x64,s=n.Word,x=n.WordArray,o=r.algo,f=o.SHA512,a=o.SHA384=f.extend({_doReset:function(){this._hash=new x.init([new s.init(3418070365,3238371032),new s.init(1654270250,914150663),new s.init(2438529370,812702999),new s.init(355462360,4144912697),new s.init(1731405415,4290775857),new s.init(2394180231,1750603025),new s.init(3675008525,1694076839),new s.init(1203062813,3204075428)])},_doFinalize:function(){var c=f._doFinalize.call(this);return c.sigBytes-=16,c}});r.SHA384=f._createHelper(a),r.HmacSHA384=f._createHmacHelper(a)}(),t.SHA384})}(lt)),lt.exports}var dt={exports:{}},ce;function yr(){return ce||(ce=1,function(i,e){(function(t,r,n){i.exports=r(V(),L0())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.x64,a=f.Word,c=n.algo,E=[],h=[],u=[];(function(){for(var l=1,C=0,d=0;d<24;d++){E[l+5*C]=(d+1)*(d+2)/2%64;var g=C%5,B=(2*l+3*C)%5;l=g,C=B}for(var l=0;l<5;l++)for(var C=0;C<5;C++)h[l+5*C]=C+(2*l+3*C)%5*5;for(var F=1,A=0;A<24;A++){for(var _=0,b=0,P=0;P<7;P++){if(F&1){var D=(1<<P)-1;D<32?b^=1<<D:_^=1<<D-32}F&128?F=F<<1^113:F<<=1}u[A]=a.create(_,b)}})();var v=[];(function(){for(var l=0;l<25;l++)v[l]=a.create()})();var p=c.SHA3=o.extend({cfg:o.cfg.extend({outputLength:512}),_doReset:function(){for(var l=this._state=[],C=0;C<25;C++)l[C]=new a.init;this.blockSize=(1600-2*this.cfg.outputLength)/32},_doProcessBlock:function(l,C){for(var d=this._state,g=this.blockSize/2,B=0;B<g;B++){var F=l[C+2*B],A=l[C+2*B+1];F=(F<<8|F>>>24)&16711935|(F<<24|F>>>8)&4278255360,A=(A<<8|A>>>24)&16711935|(A<<24|A>>>8)&4278255360;var _=d[B];_.high^=A,_.low^=F}for(var b=0;b<24;b++){for(var P=0;P<5;P++){for(var D=0,y=0,T=0;T<5;T++){var _=d[P+5*T];D^=_.high,y^=_.low}var w=v[P];w.high=D,w.low=y}for(var P=0;P<5;P++)for(var H=v[(P+4)%5],N=v[(P+1)%5],L=N.high,Z=N.low,D=H.high^(L<<1|Z>>>31),y=H.low^(Z<<1|L>>>31),T=0;T<5;T++){var _=d[P+5*T];_.high^=D,_.low^=y}for(var K=1;K<25;K++){var D,y,_=d[K],j=_.high,$=_.low,S=E[K];S<32?(D=j<<S|$>>>32-S,y=$<<S|j>>>32-S):(D=$<<S-32|j>>>64-S,y=j<<S-32|$>>>64-S);var k=v[h[K]];k.high=D,k.low=y}var O=v[0],R=d[0];O.high=R.high,O.low=R.low;for(var P=0;P<5;P++)for(var T=0;T<5;T++){var K=P+5*T,_=d[K],X=v[K],G=v[(P+1)%5+5*T],r0=v[(P+2)%5+5*T];_.high=X.high^~G.high&r0.high,_.low=X.low^~G.low&r0.low}var _=d[0],M=u[b];_.high^=M.high,_.low^=M.low}},_doFinalize:function(){var l=this._data,C=l.words;this._nDataBytes*8;var d=l.sigBytes*8,g=this.blockSize*32;C[d>>>5]|=1<<24-d%32,C[(r.ceil((d+1)/g)*g>>>5)-1]|=128,l.sigBytes=C.length*4,this._process();for(var B=this._state,F=this.cfg.outputLength/8,A=F/8,_=[],b=0;b<A;b++){var P=B[b],D=P.high,y=P.low;D=(D<<8|D>>>24)&16711935|(D<<24|D>>>8)&4278255360,y=(y<<8|y>>>24)&16711935|(y<<24|y>>>8)&4278255360,_.push(y),_.push(D)}return new x.init(_,F)},clone:function(){for(var l=o.clone.call(this),C=l._state=this._state.slice(0),d=0;d<25;d++)C[d]=C[d].clone();return l}});n.SHA3=o._createHelper(p),n.HmacSHA3=o._createHmacHelper(p)}(Math),t.SHA3})}(dt)),dt.exports}var vt={exports:{}},he;function Fr(){return he||(he=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){/** @preserve
+(function(a0,x0){typeof exports=="object"&&typeof module<"u"?module.exports=x0():typeof define=="function"&&define.amd?define(x0):(a0=typeof globalThis<"u"?globalThis:a0||self,a0.telemetree=x0())})(this,function(){"use strict";var $i=Object.defineProperty;var ji=(a0,x0,p0)=>x0 in a0?$i(a0,x0,{enumerable:!0,configurable:!0,writable:!0,value:p0}):a0[x0]=p0;var i0=(a0,x0,p0)=>(ji(a0,typeof x0!="symbol"?x0+"":x0,p0),p0);var tr,er;const a0="https://ebn.telemetree.io/public-api/v1/client/config",p0="[TS]",dr=" | ";var q=(i=>(i.Click="Click",i.PageView="Pageview",i.Closed="Closed",i.MainButtonPressed="MainButtonPressed",i.SettingsButtonPressed="SettingsButtonPressed",i.InvoiceClosed="InvoiceClosed",i.InvoiceOpened="Invoice opened",i.PopupClosed="PopupClosed",i.ClipboardTextReceived="ClipboardTextReceived",i.WriteAccessRequested="WriteAccessRequested",i.QRTextReceived="QRTextReceived",i.PhoneRequested="PhoneRequested",i.BackButtonPressed="BackButtonPressed",i.SecondaryButtonPressed="SeconadaryButtonPressed",i.PreparedMessageSent="PreparedMessageSent",i.FullScreenChanged="FullScreenChanged",i.HomeScreenAdded="HomeScreenAdded",i.HomeScreenChecked="HomeScreenChecked",i.EmojiStatusSet="EmojiStatusSet",i.LocationChecked="LocationChecked",i.LocationRequested="LocationRequested",i.AccelerometerStarted="AccelerometerStarted",i.AccelerometerStopped="AccelerometerStopped",i.AccelerometerChanged="AccelerometerChanged",i.DeviceOrientationStarted="DeviceOrientationStarted",i.DeviceOrientationStopped="DeviceOrientationStopped",i.DeviceOrientationChanged="DeviceOrientationChanged",i.DeviceOrientationFailed="DeviceOrientationFailed",i.GyroscropeStarted="GyroscropeStarted",i.GyroscropeStopped="GyroscropeStopped",i.GyroscropeChanged="GyroscropeChanged",i.GyroscropeFailed="GyroscropeFailed",i.WebAppRequestFullscreen="Fullscreen on",i.WebAppExitFullscreen="Fullscreen off",i.SwitchInlineQuery="Inline query opened",i.ShareToStory="Story shared",i.SessionStart="Session start",i.SessionEnd="Session end",i.OpenLink="Link opened",i.OpenTgLink="Telegram link opened",i.Wallet="Wallet",i.WalletConnectStarted="Wallet connect started",i.WalletConnectError="Wallet connect error",i.WalletConnectionRestoringStarted="Wallet connection restoring started",i.WalletConnectionRestored="Wallet connection restored",i.WalletConnectionRestoreError="Wallet connection restore error",i.WalletDisconnected="Wallet disconnected",i.TransactionSentForSignature="Transaction sent for signature",i.TransactionSigned="Transaction signed",i.TransactionSigningFailed="Transaction signing failed",i))(q||{}),J=(i=>(i.WalletConnectStarted="connection-started",i.WalletConnectSuccess="connection-completed",i.WalletConnectError="connection-error",i.ConnectionRestoringStarted="connection-restoring-started",i.ConnectionRestoringSuccess="connection-restoring-completed",i.ConnectionRestoringError="connection-restoring-error",i.WalletDisconnect="disconnection",i.TransactionSentForSignature="transaction-sent-for-signature",i.TransactionSigned="transaction-signed",i.TransactionSigningFailed="transaction-signing-failed",i))(J||{});class vr{constructor(e){i0(this,"eventsQueue",[]);i0(this,"isProcessing",!1);this.client=e}push(e){this.eventsQueue.push(e),this.processQueue()}processQueue(){this.isProcessing===!0||this.client.getConfig()===null||(this.isProcessing=!0,this.flush().then(()=>{this.isProcessing=!1,this.eventsQueue.length>0&&this.processQueue()}))}async flush(){const e=this.eventsQueue;if(this.eventsQueue=[],e.length===0)return;const t=e.map(r=>this.client.processEvent.bind(this.client)(r));await Promise.all(t)}}const pr=()=>Math.floor(Date.now()/1e3).toString(),Zt=()=>Date.now().toString();var W=typeof globalThis<"u"?globalThis:typeof window<"u"?window:typeof global<"u"?global:typeof self<"u"?self:{};function gr(i){return i&&i.__esModule&&Object.prototype.hasOwnProperty.call(i,"default")?i.default:i}function Er(i){if(i.__esModule)return i;var e=i.default;if(typeof e=="function"){var t=function r(){return this instanceof r?Reflect.construct(e,arguments,this.constructor):e.apply(this,arguments)};t.prototype=e.prototype}else t={};return Object.defineProperty(t,"__esModule",{value:!0}),Object.keys(i).forEach(function(r){var n=Object.getOwnPropertyDescriptor(i,r);Object.defineProperty(t,r,n.get?n:{enumerable:!0,get:function(){return i[r]}})}),t}var Xt={exports:{}};function Br(i){throw new Error('Could not dynamically require "'+i+'". Please configure the dynamicRequireTargets or/and ignoreDynamicRequires option of @rollup/plugin-commonjs appropriately for this require call to work.')}var rt={exports:{}};const Cr=Er(Object.freeze(Object.defineProperty({__proto__:null,default:{}},Symbol.toStringTag,{value:"Module"})));var Qt;function V(){return Qt||(Qt=1,function(i,e){(function(t,r){i.exports=r()})(W,function(){var t=t||function(r,n){var s;if(typeof window<"u"&&window.crypto&&(s=window.crypto),typeof self<"u"&&self.crypto&&(s=self.crypto),typeof globalThis<"u"&&globalThis.crypto&&(s=globalThis.crypto),!s&&typeof window<"u"&&window.msCrypto&&(s=window.msCrypto),!s&&typeof W<"u"&&W.crypto&&(s=W.crypto),!s&&typeof Br=="function")try{s=Cr}catch{}var x=function(){if(s){if(typeof s.getRandomValues=="function")try{return s.getRandomValues(new Uint32Array(1))[0]}catch{}if(typeof s.randomBytes=="function")try{return s.randomBytes(4).readInt32LE()}catch{}}throw new Error("Native crypto module could not be used to get secure random number.")},o=Object.create||function(){function d(){}return function(p){var E;return d.prototype=p,E=new d,d.prototype=null,E}}(),f={},a=f.lib={},c=a.Base=function(){return{extend:function(d){var p=o(this);return d&&p.mixIn(d),(!p.hasOwnProperty("init")||this.init===p.init)&&(p.init=function(){p.$super.init.apply(this,arguments)}),p.init.prototype=p,p.$super=this,p},create:function(){var d=this.extend();return d.init.apply(d,arguments),d},init:function(){},mixIn:function(d){for(var p in d)d.hasOwnProperty(p)&&(this[p]=d[p]);d.hasOwnProperty("toString")&&(this.toString=d.toString)},clone:function(){return this.init.prototype.extend(this)}}}(),B=a.WordArray=c.extend({init:function(d,p){d=this.words=d||[],p!=n?this.sigBytes=p:this.sigBytes=d.length*4},toString:function(d){return(d||u).stringify(this)},concat:function(d){var p=this.words,E=d.words,F=this.sigBytes,C=d.sigBytes;if(this.clamp(),F%4)for(var _=0;_<C;_++){var m=E[_>>>2]>>>24-_%4*8&255;p[F+_>>>2]|=m<<24-(F+_)%4*8}else for(var I=0;I<C;I+=4)p[F+I>>>2]=E[I>>>2];return this.sigBytes+=C,this},clamp:function(){var d=this.words,p=this.sigBytes;d[p>>>2]&=4294967295<<32-p%4*8,d.length=r.ceil(p/4)},clone:function(){var d=c.clone.call(this);return d.words=this.words.slice(0),d},random:function(d){for(var p=[],E=0;E<d;E+=4)p.push(x());return new B.init(p,d)}}),h=f.enc={},u=h.Hex={stringify:function(d){for(var p=d.words,E=d.sigBytes,F=[],C=0;C<E;C++){var _=p[C>>>2]>>>24-C%4*8&255;F.push((_>>>4).toString(16)),F.push((_&15).toString(16))}return F.join("")},parse:function(d){for(var p=d.length,E=[],F=0;F<p;F+=2)E[F>>>3]|=parseInt(d.substr(F,2),16)<<24-F%8*4;return new B.init(E,p/2)}},v=h.Latin1={stringify:function(d){for(var p=d.words,E=d.sigBytes,F=[],C=0;C<E;C++){var _=p[C>>>2]>>>24-C%4*8&255;F.push(String.fromCharCode(_))}return F.join("")},parse:function(d){for(var p=d.length,E=[],F=0;F<p;F++)E[F>>>2]|=(d.charCodeAt(F)&255)<<24-F%4*8;return new B.init(E,p)}},g=h.Utf8={stringify:function(d){try{return decodeURIComponent(escape(v.stringify(d)))}catch{throw new Error("Malformed UTF-8 data")}},parse:function(d){return v.parse(unescape(encodeURIComponent(d)))}},l=a.BufferedBlockAlgorithm=c.extend({reset:function(){this._data=new B.init,this._nDataBytes=0},_append:function(d){typeof d=="string"&&(d=g.parse(d)),this._data.concat(d),this._nDataBytes+=d.sigBytes},_process:function(d){var p,E=this._data,F=E.words,C=E.sigBytes,_=this.blockSize,m=_*4,I=C/m;d?I=r.ceil(I):I=r.max((I|0)-this._minBufferSize,0);var D=I*_,y=r.min(D*4,C);if(D){for(var T=0;T<D;T+=_)this._doProcessBlock(F,T);p=F.splice(0,D),E.sigBytes-=y}return new B.init(p,y)},clone:function(){var d=c.clone.call(this);return d._data=this._data.clone(),d},_minBufferSize:0});a.Hasher=l.extend({cfg:c.extend(),init:function(d){this.cfg=this.cfg.extend(d),this.reset()},reset:function(){l.reset.call(this),this._doReset()},update:function(d){return this._append(d),this._process(),this},finalize:function(d){d&&this._append(d);var p=this._doFinalize();return p},blockSize:16,_createHelper:function(d){return function(p,E){return new d.init(E).finalize(p)}},_createHmacHelper:function(d){return function(p,E){return new A.HMAC.init(d,E).finalize(p)}}});var A=f.algo={};return f}(Math);return t})}(rt)),rt.exports}var it={exports:{}},Yt;function M0(){return Yt||(Yt=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.Base,o=s.WordArray,f=n.x64={};f.Word=x.extend({init:function(a,c){this.high=a,this.low=c}}),f.WordArray=x.extend({init:function(a,c){a=this.words=a||[],c!=r?this.sigBytes=c:this.sigBytes=a.length*8},toX32:function(){for(var a=this.words,c=a.length,B=[],h=0;h<c;h++){var u=a[h];B.push(u.high),B.push(u.low)}return o.create(B,this.sigBytes)},clone:function(){for(var a=x.clone.call(this),c=a.words=this.words.slice(0),B=c.length,h=0;h<B;h++)c[h]=c[h].clone();return a}})}(),t})}(it)),it.exports}var nt={exports:{}},Jt;function Dr(){return Jt||(Jt=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){if(typeof ArrayBuffer=="function"){var r=t,n=r.lib,s=n.WordArray,x=s.init,o=s.init=function(f){if(f instanceof ArrayBuffer&&(f=new Uint8Array(f)),(f instanceof Int8Array||typeof Uint8ClampedArray<"u"&&f instanceof Uint8ClampedArray||f instanceof Int16Array||f instanceof Uint16Array||f instanceof Int32Array||f instanceof Uint32Array||f instanceof Float32Array||f instanceof Float64Array)&&(f=new Uint8Array(f.buffer,f.byteOffset,f.byteLength)),f instanceof Uint8Array){for(var a=f.byteLength,c=[],B=0;B<a;B++)c[B>>>2]|=f[B]<<24-B%4*8;x.call(this,c,a)}else x.apply(this,arguments)};o.prototype=s}}(),t.lib.WordArray})}(nt)),nt.exports}var st={exports:{}},te;function Ar(){return te||(te=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Utf16=x.Utf16BE={stringify:function(f){for(var a=f.words,c=f.sigBytes,B=[],h=0;h<c;h+=2){var u=a[h>>>2]>>>16-h%4*8&65535;B.push(String.fromCharCode(u))}return B.join("")},parse:function(f){for(var a=f.length,c=[],B=0;B<a;B++)c[B>>>1]|=f.charCodeAt(B)<<16-B%2*16;return s.create(c,a*2)}},x.Utf16LE={stringify:function(f){for(var a=f.words,c=f.sigBytes,B=[],h=0;h<c;h+=2){var u=o(a[h>>>2]>>>16-h%4*8&65535);B.push(String.fromCharCode(u))}return B.join("")},parse:function(f){for(var a=f.length,c=[],B=0;B<a;B++)c[B>>>1]|=o(f.charCodeAt(B)<<16-B%2*16);return s.create(c,a*2)}};function o(f){return f<<8&4278255360|f>>>8&16711935}}(),t.enc.Utf16})}(st)),st.exports}var ot={exports:{}},ee;function y0(){return ee||(ee=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Base64={stringify:function(f){var a=f.words,c=f.sigBytes,B=this._map;f.clamp();for(var h=[],u=0;u<c;u+=3)for(var v=a[u>>>2]>>>24-u%4*8&255,g=a[u+1>>>2]>>>24-(u+1)%4*8&255,l=a[u+2>>>2]>>>24-(u+2)%4*8&255,A=v<<16|g<<8|l,d=0;d<4&&u+d*.75<c;d++)h.push(B.charAt(A>>>6*(3-d)&63));var p=B.charAt(64);if(p)for(;h.length%4;)h.push(p);return h.join("")},parse:function(f){var a=f.length,c=this._map,B=this._reverseMap;if(!B){B=this._reverseMap=[];for(var h=0;h<c.length;h++)B[c.charCodeAt(h)]=h}var u=c.charAt(64);if(u){var v=f.indexOf(u);v!==-1&&(a=v)}return o(f,a,B)},_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/="};function o(f,a,c){for(var B=[],h=0,u=0;u<a;u++)if(u%4){var v=c[f.charCodeAt(u-1)]<<u%4*2,g=c[f.charCodeAt(u)]>>>6-u%4*2,l=v|g;B[h>>>2]|=l<<24-h%4*8,h++}return s.create(B,h)}}(),t.enc.Base64})}(ot)),ot.exports}var at={exports:{}},re;function yr(){return re||(re=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.enc;x.Base64url={stringify:function(f,a){a===void 0&&(a=!0);var c=f.words,B=f.sigBytes,h=a?this._safe_map:this._map;f.clamp();for(var u=[],v=0;v<B;v+=3)for(var g=c[v>>>2]>>>24-v%4*8&255,l=c[v+1>>>2]>>>24-(v+1)%4*8&255,A=c[v+2>>>2]>>>24-(v+2)%4*8&255,d=g<<16|l<<8|A,p=0;p<4&&v+p*.75<B;p++)u.push(h.charAt(d>>>6*(3-p)&63));var E=h.charAt(64);if(E)for(;u.length%4;)u.push(E);return u.join("")},parse:function(f,a){a===void 0&&(a=!0);var c=f.length,B=a?this._safe_map:this._map,h=this._reverseMap;if(!h){h=this._reverseMap=[];for(var u=0;u<B.length;u++)h[B.charCodeAt(u)]=u}var v=B.charAt(64);if(v){var g=f.indexOf(v);g!==-1&&(c=g)}return o(f,c,h)},_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=",_safe_map:"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"};function o(f,a,c){for(var B=[],h=0,u=0;u<a;u++)if(u%4){var v=c[f.charCodeAt(u-1)]<<u%4*2,g=c[f.charCodeAt(u)]>>>6-u%4*2,l=v|g;B[h>>>2]|=l<<24-h%4*8,h++}return s.create(B,h)}}(),t.enc.Base64url})}(at)),at.exports}var xt={exports:{}},ie;function F0(){return ie||(ie=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=[];(function(){for(var g=0;g<64;g++)a[g]=r.abs(r.sin(g+1))*4294967296|0})();var c=f.MD5=o.extend({_doReset:function(){this._hash=new x.init([1732584193,4023233417,2562383102,271733878])},_doProcessBlock:function(g,l){for(var A=0;A<16;A++){var d=l+A,p=g[d];g[d]=(p<<8|p>>>24)&16711935|(p<<24|p>>>8)&4278255360}var E=this._hash.words,F=g[l+0],C=g[l+1],_=g[l+2],m=g[l+3],I=g[l+4],D=g[l+5],y=g[l+6],T=g[l+7],w=g[l+8],H=g[l+9],N=g[l+10],L=g[l+11],Z=g[l+12],K=g[l+13],$=g[l+14],G=g[l+15],S=E[0],R=E[1],P=E[2],k=E[3];S=B(S,R,P,k,F,7,a[0]),k=B(k,S,R,P,C,12,a[1]),P=B(P,k,S,R,_,17,a[2]),R=B(R,P,k,S,m,22,a[3]),S=B(S,R,P,k,I,7,a[4]),k=B(k,S,R,P,D,12,a[5]),P=B(P,k,S,R,y,17,a[6]),R=B(R,P,k,S,T,22,a[7]),S=B(S,R,P,k,w,7,a[8]),k=B(k,S,R,P,H,12,a[9]),P=B(P,k,S,R,N,17,a[10]),R=B(R,P,k,S,L,22,a[11]),S=B(S,R,P,k,Z,7,a[12]),k=B(k,S,R,P,K,12,a[13]),P=B(P,k,S,R,$,17,a[14]),R=B(R,P,k,S,G,22,a[15]),S=h(S,R,P,k,C,5,a[16]),k=h(k,S,R,P,y,9,a[17]),P=h(P,k,S,R,L,14,a[18]),R=h(R,P,k,S,F,20,a[19]),S=h(S,R,P,k,D,5,a[20]),k=h(k,S,R,P,N,9,a[21]),P=h(P,k,S,R,G,14,a[22]),R=h(R,P,k,S,I,20,a[23]),S=h(S,R,P,k,H,5,a[24]),k=h(k,S,R,P,$,9,a[25]),P=h(P,k,S,R,m,14,a[26]),R=h(R,P,k,S,w,20,a[27]),S=h(S,R,P,k,K,5,a[28]),k=h(k,S,R,P,_,9,a[29]),P=h(P,k,S,R,T,14,a[30]),R=h(R,P,k,S,Z,20,a[31]),S=u(S,R,P,k,D,4,a[32]),k=u(k,S,R,P,w,11,a[33]),P=u(P,k,S,R,L,16,a[34]),R=u(R,P,k,S,$,23,a[35]),S=u(S,R,P,k,C,4,a[36]),k=u(k,S,R,P,I,11,a[37]),P=u(P,k,S,R,T,16,a[38]),R=u(R,P,k,S,N,23,a[39]),S=u(S,R,P,k,K,4,a[40]),k=u(k,S,R,P,F,11,a[41]),P=u(P,k,S,R,m,16,a[42]),R=u(R,P,k,S,y,23,a[43]),S=u(S,R,P,k,H,4,a[44]),k=u(k,S,R,P,Z,11,a[45]),P=u(P,k,S,R,G,16,a[46]),R=u(R,P,k,S,_,23,a[47]),S=v(S,R,P,k,F,6,a[48]),k=v(k,S,R,P,T,10,a[49]),P=v(P,k,S,R,$,15,a[50]),R=v(R,P,k,S,D,21,a[51]),S=v(S,R,P,k,Z,6,a[52]),k=v(k,S,R,P,m,10,a[53]),P=v(P,k,S,R,N,15,a[54]),R=v(R,P,k,S,C,21,a[55]),S=v(S,R,P,k,w,6,a[56]),k=v(k,S,R,P,G,10,a[57]),P=v(P,k,S,R,y,15,a[58]),R=v(R,P,k,S,K,21,a[59]),S=v(S,R,P,k,I,6,a[60]),k=v(k,S,R,P,L,10,a[61]),P=v(P,k,S,R,_,15,a[62]),R=v(R,P,k,S,H,21,a[63]),E[0]=E[0]+S|0,E[1]=E[1]+R|0,E[2]=E[2]+P|0,E[3]=E[3]+k|0},_doFinalize:function(){var g=this._data,l=g.words,A=this._nDataBytes*8,d=g.sigBytes*8;l[d>>>5]|=128<<24-d%32;var p=r.floor(A/4294967296),E=A;l[(d+64>>>9<<4)+15]=(p<<8|p>>>24)&16711935|(p<<24|p>>>8)&4278255360,l[(d+64>>>9<<4)+14]=(E<<8|E>>>24)&16711935|(E<<24|E>>>8)&4278255360,g.sigBytes=(l.length+1)*4,this._process();for(var F=this._hash,C=F.words,_=0;_<4;_++){var m=C[_];C[_]=(m<<8|m>>>24)&16711935|(m<<24|m>>>8)&4278255360}return F},clone:function(){var g=o.clone.call(this);return g._hash=this._hash.clone(),g}});function B(g,l,A,d,p,E,F){var C=g+(l&A|~l&d)+p+F;return(C<<E|C>>>32-E)+l}function h(g,l,A,d,p,E,F){var C=g+(l&d|A&~d)+p+F;return(C<<E|C>>>32-E)+l}function u(g,l,A,d,p,E,F){var C=g+(l^A^d)+p+F;return(C<<E|C>>>32-E)+l}function v(g,l,A,d,p,E,F){var C=g+(A^(l|~d))+p+F;return(C<<E|C>>>32-E)+l}n.MD5=o._createHelper(c),n.HmacMD5=o._createHmacHelper(c)}(Math),t.MD5})}(xt)),xt.exports}var ft={exports:{}},ne;function se(){return ne||(ne=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=n.Hasher,o=r.algo,f=[],a=o.SHA1=x.extend({_doReset:function(){this._hash=new s.init([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(c,B){for(var h=this._hash.words,u=h[0],v=h[1],g=h[2],l=h[3],A=h[4],d=0;d<80;d++){if(d<16)f[d]=c[B+d]|0;else{var p=f[d-3]^f[d-8]^f[d-14]^f[d-16];f[d]=p<<1|p>>>31}var E=(u<<5|u>>>27)+A+f[d];d<20?E+=(v&g|~v&l)+1518500249:d<40?E+=(v^g^l)+1859775393:d<60?E+=(v&g|v&l|g&l)-1894007588:E+=(v^g^l)-899497514,A=l,l=g,g=v<<30|v>>>2,v=u,u=E}h[0]=h[0]+u|0,h[1]=h[1]+v|0,h[2]=h[2]+g|0,h[3]=h[3]+l|0,h[4]=h[4]+A|0},_doFinalize:function(){var c=this._data,B=c.words,h=this._nDataBytes*8,u=c.sigBytes*8;return B[u>>>5]|=128<<24-u%32,B[(u+64>>>9<<4)+14]=Math.floor(h/4294967296),B[(u+64>>>9<<4)+15]=h,c.sigBytes=B.length*4,this._process(),this._hash},clone:function(){var c=x.clone.call(this);return c._hash=this._hash.clone(),c}});r.SHA1=x._createHelper(a),r.HmacSHA1=x._createHmacHelper(a)}(),t.SHA1})}(ft)),ft.exports}var ct={exports:{}},oe;function ht(){return oe||(oe=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=[],c=[];(function(){function u(A){for(var d=r.sqrt(A),p=2;p<=d;p++)if(!(A%p))return!1;return!0}function v(A){return(A-(A|0))*4294967296|0}for(var g=2,l=0;l<64;)u(g)&&(l<8&&(a[l]=v(r.pow(g,1/2))),c[l]=v(r.pow(g,1/3)),l++),g++})();var B=[],h=f.SHA256=o.extend({_doReset:function(){this._hash=new x.init(a.slice(0))},_doProcessBlock:function(u,v){for(var g=this._hash.words,l=g[0],A=g[1],d=g[2],p=g[3],E=g[4],F=g[5],C=g[6],_=g[7],m=0;m<64;m++){if(m<16)B[m]=u[v+m]|0;else{var I=B[m-15],D=(I<<25|I>>>7)^(I<<14|I>>>18)^I>>>3,y=B[m-2],T=(y<<15|y>>>17)^(y<<13|y>>>19)^y>>>10;B[m]=D+B[m-7]+T+B[m-16]}var w=E&F^~E&C,H=l&A^l&d^A&d,N=(l<<30|l>>>2)^(l<<19|l>>>13)^(l<<10|l>>>22),L=(E<<26|E>>>6)^(E<<21|E>>>11)^(E<<7|E>>>25),Z=_+L+w+c[m]+B[m],K=N+H;_=C,C=F,F=E,E=p+Z|0,p=d,d=A,A=l,l=Z+K|0}g[0]=g[0]+l|0,g[1]=g[1]+A|0,g[2]=g[2]+d|0,g[3]=g[3]+p|0,g[4]=g[4]+E|0,g[5]=g[5]+F|0,g[6]=g[6]+C|0,g[7]=g[7]+_|0},_doFinalize:function(){var u=this._data,v=u.words,g=this._nDataBytes*8,l=u.sigBytes*8;return v[l>>>5]|=128<<24-l%32,v[(l+64>>>9<<4)+14]=r.floor(g/4294967296),v[(l+64>>>9<<4)+15]=g,u.sigBytes=v.length*4,this._process(),this._hash},clone:function(){var u=o.clone.call(this);return u._hash=this._hash.clone(),u}});n.SHA256=o._createHelper(h),n.HmacSHA256=o._createHmacHelper(h)}(Math),t.SHA256})}(ct)),ct.exports}var ut={exports:{}},ae;function Fr(){return ae||(ae=1,function(i,e){(function(t,r,n){i.exports=r(V(),ht())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=r.algo,o=x.SHA256,f=x.SHA224=o.extend({_doReset:function(){this._hash=new s.init([3238371032,914150663,812702999,4144912697,4290775857,1750603025,1694076839,3204075428])},_doFinalize:function(){var a=o._doFinalize.call(this);return a.sigBytes-=4,a}});r.SHA224=o._createHelper(f),r.HmacSHA224=o._createHmacHelper(f)}(),t.SHA224})}(ut)),ut.exports}var lt={exports:{}},xe;function fe(){return xe||(xe=1,function(i,e){(function(t,r,n){i.exports=r(V(),M0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Hasher,x=r.x64,o=x.Word,f=x.WordArray,a=r.algo;function c(){return o.create.apply(o,arguments)}var B=[c(1116352408,3609767458),c(1899447441,602891725),c(3049323471,3964484399),c(3921009573,2173295548),c(961987163,4081628472),c(1508970993,3053834265),c(2453635748,2937671579),c(2870763221,3664609560),c(3624381080,2734883394),c(310598401,1164996542),c(607225278,1323610764),c(1426881987,3590304994),c(1925078388,4068182383),c(2162078206,991336113),c(2614888103,633803317),c(3248222580,3479774868),c(3835390401,2666613458),c(4022224774,944711139),c(264347078,2341262773),c(604807628,2007800933),c(770255983,1495990901),c(1249150122,1856431235),c(1555081692,3175218132),c(1996064986,2198950837),c(2554220882,3999719339),c(2821834349,766784016),c(2952996808,2566594879),c(3210313671,3203337956),c(3336571891,1034457026),c(3584528711,2466948901),c(113926993,3758326383),c(338241895,168717936),c(666307205,1188179964),c(773529912,1546045734),c(1294757372,1522805485),c(1396182291,2643833823),c(1695183700,2343527390),c(1986661051,1014477480),c(2177026350,1206759142),c(2456956037,344077627),c(2730485921,1290863460),c(2820302411,3158454273),c(3259730800,3505952657),c(3345764771,106217008),c(3516065817,3606008344),c(3600352804,1432725776),c(4094571909,1467031594),c(275423344,851169720),c(430227734,3100823752),c(506948616,1363258195),c(659060556,3750685593),c(883997877,3785050280),c(958139571,3318307427),c(1322822218,3812723403),c(1537002063,2003034995),c(1747873779,3602036899),c(1955562222,1575990012),c(2024104815,1125592928),c(2227730452,2716904306),c(2361852424,442776044),c(2428436474,593698344),c(2756734187,3733110249),c(3204031479,2999351573),c(3329325298,3815920427),c(3391569614,3928383900),c(3515267271,566280711),c(3940187606,3454069534),c(4118630271,4000239992),c(116418474,1914138554),c(174292421,2731055270),c(289380356,3203993006),c(460393269,320620315),c(685471733,587496836),c(852142971,1086792851),c(1017036298,365543100),c(1126000580,2618297676),c(1288033470,3409855158),c(1501505948,4234509866),c(1607167915,987167468),c(1816402316,1246189591)],h=[];(function(){for(var v=0;v<80;v++)h[v]=c()})();var u=a.SHA512=s.extend({_doReset:function(){this._hash=new f.init([new o.init(1779033703,4089235720),new o.init(3144134277,2227873595),new o.init(1013904242,4271175723),new o.init(2773480762,1595750129),new o.init(1359893119,2917565137),new o.init(2600822924,725511199),new o.init(528734635,4215389547),new o.init(1541459225,327033209)])},_doProcessBlock:function(v,g){for(var l=this._hash.words,A=l[0],d=l[1],p=l[2],E=l[3],F=l[4],C=l[5],_=l[6],m=l[7],I=A.high,D=A.low,y=d.high,T=d.low,w=p.high,H=p.low,N=E.high,L=E.low,Z=F.high,K=F.low,$=C.high,G=C.low,S=_.high,R=_.low,P=m.high,k=m.low,X=I,j=D,r0=y,M=T,O0=w,k0=H,$t=N,H0=L,l0=Z,s0=K,J0=$,q0=G,tt=S,N0=R,jt=P,z0=k,d0=0;d0<80;d0++){var c0,D0,et=h[d0];if(d0<16)D0=et.high=v[g+d0*2]|0,c0=et.low=v[g+d0*2+1]|0;else{var rr=h[d0-15],R0=rr.high,W0=rr.low,Pi=(R0>>>1|W0<<31)^(R0>>>8|W0<<24)^R0>>>7,ir=(W0>>>1|R0<<31)^(W0>>>8|R0<<24)^(W0>>>7|R0<<25),nr=h[d0-2],I0=nr.high,L0=nr.low,Oi=(I0>>>19|L0<<13)^(I0<<3|L0>>>29)^I0>>>6,sr=(L0>>>19|I0<<13)^(L0<<3|I0>>>29)^(L0>>>6|I0<<26),or=h[d0-7],Hi=or.high,qi=or.low,ar=h[d0-16],Ni=ar.high,xr=ar.low;c0=ir+qi,D0=Pi+Hi+(c0>>>0<ir>>>0?1:0),c0=c0+sr,D0=D0+Oi+(c0>>>0<sr>>>0?1:0),c0=c0+xr,D0=D0+Ni+(c0>>>0<xr>>>0?1:0),et.high=D0,et.low=c0}var zi=l0&J0^~l0&tt,fr=s0&q0^~s0&N0,Wi=X&r0^X&O0^r0&O0,Li=j&M^j&k0^M&k0,Mi=(X>>>28|j<<4)^(X<<30|j>>>2)^(X<<25|j>>>7),cr=(j>>>28|X<<4)^(j<<30|X>>>2)^(j<<25|X>>>7),Ui=(l0>>>14|s0<<18)^(l0>>>18|s0<<14)^(l0<<23|s0>>>9),Vi=(s0>>>14|l0<<18)^(s0>>>18|l0<<14)^(s0<<23|l0>>>9),hr=B[d0],Ki=hr.high,ur=hr.low,o0=z0+Vi,A0=jt+Ui+(o0>>>0<z0>>>0?1:0),o0=o0+fr,A0=A0+zi+(o0>>>0<fr>>>0?1:0),o0=o0+ur,A0=A0+Ki+(o0>>>0<ur>>>0?1:0),o0=o0+c0,A0=A0+D0+(o0>>>0<c0>>>0?1:0),lr=cr+Li,Gi=Mi+Wi+(lr>>>0<cr>>>0?1:0);jt=tt,z0=N0,tt=J0,N0=q0,J0=l0,q0=s0,s0=H0+o0|0,l0=$t+A0+(s0>>>0<H0>>>0?1:0)|0,$t=O0,H0=k0,O0=r0,k0=M,r0=X,M=j,j=o0+lr|0,X=A0+Gi+(j>>>0<o0>>>0?1:0)|0}D=A.low=D+j,A.high=I+X+(D>>>0<j>>>0?1:0),T=d.low=T+M,d.high=y+r0+(T>>>0<M>>>0?1:0),H=p.low=H+k0,p.high=w+O0+(H>>>0<k0>>>0?1:0),L=E.low=L+H0,E.high=N+$t+(L>>>0<H0>>>0?1:0),K=F.low=K+s0,F.high=Z+l0+(K>>>0<s0>>>0?1:0),G=C.low=G+q0,C.high=$+J0+(G>>>0<q0>>>0?1:0),R=_.low=R+N0,_.high=S+tt+(R>>>0<N0>>>0?1:0),k=m.low=k+z0,m.high=P+jt+(k>>>0<z0>>>0?1:0)},_doFinalize:function(){var v=this._data,g=v.words,l=this._nDataBytes*8,A=v.sigBytes*8;g[A>>>5]|=128<<24-A%32,g[(A+128>>>10<<5)+30]=Math.floor(l/4294967296),g[(A+128>>>10<<5)+31]=l,v.sigBytes=g.length*4,this._process();var d=this._hash.toX32();return d},clone:function(){var v=s.clone.call(this);return v._hash=this._hash.clone(),v},blockSize:1024/32});r.SHA512=s._createHelper(u),r.HmacSHA512=s._createHmacHelper(u)}(),t.SHA512})}(lt)),lt.exports}var dt={exports:{}},ce;function _r(){return ce||(ce=1,function(i,e){(function(t,r,n){i.exports=r(V(),M0(),fe())})(W,function(t){return function(){var r=t,n=r.x64,s=n.Word,x=n.WordArray,o=r.algo,f=o.SHA512,a=o.SHA384=f.extend({_doReset:function(){this._hash=new x.init([new s.init(3418070365,3238371032),new s.init(1654270250,914150663),new s.init(2438529370,812702999),new s.init(355462360,4144912697),new s.init(1731405415,4290775857),new s.init(2394180231,1750603025),new s.init(3675008525,1694076839),new s.init(1203062813,3204075428)])},_doFinalize:function(){var c=f._doFinalize.call(this);return c.sigBytes-=16,c}});r.SHA384=f._createHelper(a),r.HmacSHA384=f._createHmacHelper(a)}(),t.SHA384})}(dt)),dt.exports}var vt={exports:{}},he;function mr(){return he||(he=1,function(i,e){(function(t,r,n){i.exports=r(V(),M0())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.x64,a=f.Word,c=n.algo,B=[],h=[],u=[];(function(){for(var l=1,A=0,d=0;d<24;d++){B[l+5*A]=(d+1)*(d+2)/2%64;var p=A%5,E=(2*l+3*A)%5;l=p,A=E}for(var l=0;l<5;l++)for(var A=0;A<5;A++)h[l+5*A]=A+(2*l+3*A)%5*5;for(var F=1,C=0;C<24;C++){for(var _=0,m=0,I=0;I<7;I++){if(F&1){var D=(1<<I)-1;D<32?m^=1<<D:_^=1<<D-32}F&128?F=F<<1^113:F<<=1}u[C]=a.create(_,m)}})();var v=[];(function(){for(var l=0;l<25;l++)v[l]=a.create()})();var g=c.SHA3=o.extend({cfg:o.cfg.extend({outputLength:512}),_doReset:function(){for(var l=this._state=[],A=0;A<25;A++)l[A]=new a.init;this.blockSize=(1600-2*this.cfg.outputLength)/32},_doProcessBlock:function(l,A){for(var d=this._state,p=this.blockSize/2,E=0;E<p;E++){var F=l[A+2*E],C=l[A+2*E+1];F=(F<<8|F>>>24)&16711935|(F<<24|F>>>8)&4278255360,C=(C<<8|C>>>24)&16711935|(C<<24|C>>>8)&4278255360;var _=d[E];_.high^=C,_.low^=F}for(var m=0;m<24;m++){for(var I=0;I<5;I++){for(var D=0,y=0,T=0;T<5;T++){var _=d[I+5*T];D^=_.high,y^=_.low}var w=v[I];w.high=D,w.low=y}for(var I=0;I<5;I++)for(var H=v[(I+4)%5],N=v[(I+1)%5],L=N.high,Z=N.low,D=H.high^(L<<1|Z>>>31),y=H.low^(Z<<1|L>>>31),T=0;T<5;T++){var _=d[I+5*T];_.high^=D,_.low^=y}for(var K=1;K<25;K++){var D,y,_=d[K],$=_.high,G=_.low,S=B[K];S<32?(D=$<<S|G>>>32-S,y=G<<S|$>>>32-S):(D=G<<S-32|$>>>64-S,y=$<<S-32|G>>>64-S);var R=v[h[K]];R.high=D,R.low=y}var P=v[0],k=d[0];P.high=k.high,P.low=k.low;for(var I=0;I<5;I++)for(var T=0;T<5;T++){var K=I+5*T,_=d[K],X=v[K],j=v[(I+1)%5+5*T],r0=v[(I+2)%5+5*T];_.high=X.high^~j.high&r0.high,_.low=X.low^~j.low&r0.low}var _=d[0],M=u[m];_.high^=M.high,_.low^=M.low}},_doFinalize:function(){var l=this._data,A=l.words;this._nDataBytes*8;var d=l.sigBytes*8,p=this.blockSize*32;A[d>>>5]|=1<<24-d%32,A[(r.ceil((d+1)/p)*p>>>5)-1]|=128,l.sigBytes=A.length*4,this._process();for(var E=this._state,F=this.cfg.outputLength/8,C=F/8,_=[],m=0;m<C;m++){var I=E[m],D=I.high,y=I.low;D=(D<<8|D>>>24)&16711935|(D<<24|D>>>8)&4278255360,y=(y<<8|y>>>24)&16711935|(y<<24|y>>>8)&4278255360,_.push(y),_.push(D)}return new x.init(_,F)},clone:function(){for(var l=o.clone.call(this),A=l._state=this._state.slice(0),d=0;d<25;d++)A[d]=A[d].clone();return l}});n.SHA3=o._createHelper(g),n.HmacSHA3=o._createHmacHelper(g)}(Math),t.SHA3})}(vt)),vt.exports}var pt={exports:{}},ue;function br(){return ue||(ue=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){/** @preserve
   			(c) 2012 by Cédric Mesnil. All rights reserved.
 
   			Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
@@ -7,18 +7,18 @@
   			    - Redistributions in binary form must reproduce the above copyright notice, this list of conditions and the following disclaimer in the documentation and/or other materials provided with the distribution.
 
   			THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-  			*/return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=x.create([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,7,4,13,1,10,6,15,3,12,0,9,5,2,14,11,8,3,10,14,4,9,15,8,1,2,7,0,6,13,11,5,12,1,9,11,10,0,8,12,4,13,3,7,15,14,5,6,2,4,0,5,9,7,12,2,10,14,1,3,8,11,6,15,13]),c=x.create([5,14,7,0,9,2,11,4,13,6,15,8,1,10,3,12,6,11,3,7,0,13,5,10,14,15,8,12,4,9,1,2,15,5,1,3,7,14,6,9,11,8,12,2,10,0,4,13,8,6,4,1,3,11,15,0,5,12,2,13,9,7,10,14,12,15,10,4,1,5,8,7,6,2,13,14,0,3,9,11]),E=x.create([11,14,15,12,5,8,7,9,11,13,14,15,6,7,9,8,7,6,8,13,11,9,7,15,7,12,15,9,11,7,13,12,11,13,6,7,14,9,13,15,14,8,13,6,5,12,7,5,11,12,14,15,14,15,9,8,9,14,5,6,8,6,5,12,9,15,5,11,6,8,13,12,5,12,13,14,11,8,5,6]),h=x.create([8,9,9,11,13,15,15,5,7,7,8,11,14,14,12,6,9,13,15,7,12,8,9,11,7,7,12,7,6,15,13,11,9,7,15,11,8,6,6,14,12,13,5,14,13,13,7,5,15,5,8,11,14,14,6,14,6,9,12,9,12,5,15,8,8,5,12,9,12,5,14,6,8,13,6,5,15,13,11,11]),u=x.create([0,1518500249,1859775393,2400959708,2840853838]),v=x.create([1352829926,1548603684,1836072691,2053994217,0]),p=f.RIPEMD160=o.extend({_doReset:function(){this._hash=x.create([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(A,_){for(var b=0;b<16;b++){var P=_+b,D=A[P];A[P]=(D<<8|D>>>24)&16711935|(D<<24|D>>>8)&4278255360}var y=this._hash.words,T=u.words,w=v.words,H=a.words,N=c.words,L=E.words,Z=h.words,K,j,$,S,k,O,R,X,G,r0;O=K=y[0],R=j=y[1],X=$=y[2],G=S=y[3],r0=k=y[4];for(var M,b=0;b<80;b+=1)M=K+A[_+H[b]]|0,b<16?M+=l(j,$,S)+T[0]:b<32?M+=C(j,$,S)+T[1]:b<48?M+=d(j,$,S)+T[2]:b<64?M+=g(j,$,S)+T[3]:M+=B(j,$,S)+T[4],M=M|0,M=F(M,L[b]),M=M+k|0,K=k,k=S,S=F($,10),$=j,j=M,M=O+A[_+N[b]]|0,b<16?M+=B(R,X,G)+w[0]:b<32?M+=g(R,X,G)+w[1]:b<48?M+=d(R,X,G)+w[2]:b<64?M+=C(R,X,G)+w[3]:M+=l(R,X,G)+w[4],M=M|0,M=F(M,Z[b]),M=M+r0|0,O=r0,r0=G,G=F(X,10),X=R,R=M;M=y[1]+$+G|0,y[1]=y[2]+S+r0|0,y[2]=y[3]+k+O|0,y[3]=y[4]+K+R|0,y[4]=y[0]+j+X|0,y[0]=M},_doFinalize:function(){var A=this._data,_=A.words,b=this._nDataBytes*8,P=A.sigBytes*8;_[P>>>5]|=128<<24-P%32,_[(P+64>>>9<<4)+14]=(b<<8|b>>>24)&16711935|(b<<24|b>>>8)&4278255360,A.sigBytes=(_.length+1)*4,this._process();for(var D=this._hash,y=D.words,T=0;T<5;T++){var w=y[T];y[T]=(w<<8|w>>>24)&16711935|(w<<24|w>>>8)&4278255360}return D},clone:function(){var A=o.clone.call(this);return A._hash=this._hash.clone(),A}});function l(A,_,b){return A^_^b}function C(A,_,b){return A&_|~A&b}function d(A,_,b){return(A|~_)^b}function g(A,_,b){return A&b|_&~b}function B(A,_,b){return A^(_|~b)}function F(A,_){return A<<_|A>>>32-_}n.RIPEMD160=o._createHelper(p),n.HmacRIPEMD160=o._createHmacHelper(p)}(),t.RIPEMD160})}(vt)),vt.exports}var pt={exports:{}},ue;function gt(){return ue||(ue=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){(function(){var r=t,n=r.lib,s=n.Base,x=r.enc,o=x.Utf8,f=r.algo;f.HMAC=s.extend({init:function(a,c){a=this._hasher=new a.init,typeof c=="string"&&(c=o.parse(c));var E=a.blockSize,h=E*4;c.sigBytes>h&&(c=a.finalize(c)),c.clamp();for(var u=this._oKey=c.clone(),v=this._iKey=c.clone(),p=u.words,l=v.words,C=0;C<E;C++)p[C]^=1549556828,l[C]^=909522486;u.sigBytes=v.sigBytes=h,this.reset()},reset:function(){var a=this._hasher;a.reset(),a.update(this._iKey)},update:function(a){return this._hasher.update(a),this},finalize:function(a){var c=this._hasher,E=c.finalize(a);c.reset();var h=c.finalize(this._oKey.clone().concat(E));return h}})})()})}(pt)),pt.exports}var Et={exports:{}},le;function _r(){return le||(le=1,function(i,e){(function(t,r,n){i.exports=r(V(),ct(),gt())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Base,x=n.WordArray,o=r.algo,f=o.SHA256,a=o.HMAC,c=o.PBKDF2=s.extend({cfg:s.extend({keySize:128/32,hasher:f,iterations:25e4}),init:function(E){this.cfg=this.cfg.extend(E)},compute:function(E,h){for(var u=this.cfg,v=a.create(u.hasher,E),p=x.create(),l=x.create([1]),C=p.words,d=l.words,g=u.keySize,B=u.iterations;C.length<g;){var F=v.update(h).finalize(l);v.reset();for(var A=F.words,_=A.length,b=F,P=1;P<B;P++){b=v.finalize(b),v.reset();for(var D=b.words,y=0;y<_;y++)A[y]^=D[y]}p.concat(F),d[0]++}return p.sigBytes=g*4,p}});r.PBKDF2=function(E,h,u){return c.create(u).compute(E,h)}}(),t.PBKDF2})}(Et)),Et.exports}var Bt={exports:{}},de;function p0(){return de||(de=1,function(i,e){(function(t,r,n){i.exports=r(V(),ne(),gt())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Base,x=n.WordArray,o=r.algo,f=o.MD5,a=o.EvpKDF=s.extend({cfg:s.extend({keySize:128/32,hasher:f,iterations:1}),init:function(c){this.cfg=this.cfg.extend(c)},compute:function(c,E){for(var h,u=this.cfg,v=u.hasher.create(),p=x.create(),l=p.words,C=u.keySize,d=u.iterations;l.length<C;){h&&v.update(h),h=v.update(c).finalize(E),v.reset();for(var g=1;g<d;g++)h=v.finalize(h),v.reset();p.concat(h)}return p.sigBytes=C*4,p}});r.EvpKDF=function(c,E,h){return a.create(h).compute(c,E)}}(),t.EvpKDF})}(Bt)),Bt.exports}var At={exports:{}},ve;function t0(){return ve||(ve=1,function(i,e){(function(t,r,n){i.exports=r(V(),p0())})(W,function(t){t.lib.Cipher||function(r){var n=t,s=n.lib,x=s.Base,o=s.WordArray,f=s.BufferedBlockAlgorithm,a=n.enc;a.Utf8;var c=a.Base64,E=n.algo,h=E.EvpKDF,u=s.Cipher=f.extend({cfg:x.extend(),createEncryptor:function(D,y){return this.create(this._ENC_XFORM_MODE,D,y)},createDecryptor:function(D,y){return this.create(this._DEC_XFORM_MODE,D,y)},init:function(D,y,T){this.cfg=this.cfg.extend(T),this._xformMode=D,this._key=y,this.reset()},reset:function(){f.reset.call(this),this._doReset()},process:function(D){return this._append(D),this._process()},finalize:function(D){D&&this._append(D);var y=this._doFinalize();return y},keySize:128/32,ivSize:128/32,_ENC_XFORM_MODE:1,_DEC_XFORM_MODE:2,_createHelper:function(){function D(y){return typeof y=="string"?P:A}return function(y){return{encrypt:function(T,w,H){return D(w).encrypt(y,T,w,H)},decrypt:function(T,w,H){return D(w).decrypt(y,T,w,H)}}}}()});s.StreamCipher=u.extend({_doFinalize:function(){var D=this._process(!0);return D},blockSize:1});var v=n.mode={},p=s.BlockCipherMode=x.extend({createEncryptor:function(D,y){return this.Encryptor.create(D,y)},createDecryptor:function(D,y){return this.Decryptor.create(D,y)},init:function(D,y){this._cipher=D,this._iv=y}}),l=v.CBC=function(){var D=p.extend();D.Encryptor=D.extend({processBlock:function(T,w){var H=this._cipher,N=H.blockSize;y.call(this,T,w,N),H.encryptBlock(T,w),this._prevBlock=T.slice(w,w+N)}}),D.Decryptor=D.extend({processBlock:function(T,w){var H=this._cipher,N=H.blockSize,L=T.slice(w,w+N);H.decryptBlock(T,w),y.call(this,T,w,N),this._prevBlock=L}});function y(T,w,H){var N,L=this._iv;L?(N=L,this._iv=r):N=this._prevBlock;for(var Z=0;Z<H;Z++)T[w+Z]^=N[Z]}return D}(),C=n.pad={},d=C.Pkcs7={pad:function(D,y){for(var T=y*4,w=T-D.sigBytes%T,H=w<<24|w<<16|w<<8|w,N=[],L=0;L<w;L+=4)N.push(H);var Z=o.create(N,w);D.concat(Z)},unpad:function(D){var y=D.words[D.sigBytes-1>>>2]&255;D.sigBytes-=y}};s.BlockCipher=u.extend({cfg:u.cfg.extend({mode:l,padding:d}),reset:function(){var D;u.reset.call(this);var y=this.cfg,T=y.iv,w=y.mode;this._xformMode==this._ENC_XFORM_MODE?D=w.createEncryptor:(D=w.createDecryptor,this._minBufferSize=1),this._mode&&this._mode.__creator==D?this._mode.init(this,T&&T.words):(this._mode=D.call(w,this,T&&T.words),this._mode.__creator=D)},_doProcessBlock:function(D,y){this._mode.processBlock(D,y)},_doFinalize:function(){var D,y=this.cfg.padding;return this._xformMode==this._ENC_XFORM_MODE?(y.pad(this._data,this.blockSize),D=this._process(!0)):(D=this._process(!0),y.unpad(D)),D},blockSize:128/32});var g=s.CipherParams=x.extend({init:function(D){this.mixIn(D)},toString:function(D){return(D||this.formatter).stringify(this)}}),B=n.format={},F=B.OpenSSL={stringify:function(D){var y,T=D.ciphertext,w=D.salt;return w?y=o.create([1398893684,1701076831]).concat(w).concat(T):y=T,y.toString(c)},parse:function(D){var y,T=c.parse(D),w=T.words;return w[0]==1398893684&&w[1]==1701076831&&(y=o.create(w.slice(2,4)),w.splice(0,4),T.sigBytes-=16),g.create({ciphertext:T,salt:y})}},A=s.SerializableCipher=x.extend({cfg:x.extend({format:F}),encrypt:function(D,y,T,w){w=this.cfg.extend(w);var H=D.createEncryptor(T,w),N=H.finalize(y),L=H.cfg;return g.create({ciphertext:N,key:T,iv:L.iv,algorithm:D,mode:L.mode,padding:L.padding,blockSize:D.blockSize,formatter:w.format})},decrypt:function(D,y,T,w){w=this.cfg.extend(w),y=this._parse(y,w.format);var H=D.createDecryptor(T,w).finalize(y.ciphertext);return H},_parse:function(D,y){return typeof D=="string"?y.parse(D,this):D}}),_=n.kdf={},b=_.OpenSSL={execute:function(D,y,T,w,H){if(w||(w=o.random(64/8)),H)var N=h.create({keySize:y+T,hasher:H}).compute(D,w);else var N=h.create({keySize:y+T}).compute(D,w);var L=o.create(N.words.slice(y),T*4);return N.sigBytes=y*4,g.create({key:N,iv:L,salt:w})}},P=s.PasswordBasedCipher=A.extend({cfg:A.cfg.extend({kdf:b}),encrypt:function(D,y,T,w){w=this.cfg.extend(w);var H=w.kdf.execute(T,D.keySize,D.ivSize,w.salt,w.hasher);w.iv=H.iv;var N=A.encrypt.call(this,D,y,H.key,w);return N.mixIn(H),N},decrypt:function(D,y,T,w){w=this.cfg.extend(w),y=this._parse(y,w.format);var H=w.kdf.execute(T,D.keySize,D.ivSize,y.salt,w.hasher);w.iv=H.iv;var N=A.decrypt.call(this,D,y,H.key,w);return N}})}()})}(At)),At.exports}var Ct={exports:{}},pe;function br(){return pe||(pe=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.CFB=function(){var r=t.lib.BlockCipherMode.extend();r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize;n.call(this,s,x,f,o),this._prevBlock=s.slice(x,x+f)}}),r.Decryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=s.slice(x,x+f);n.call(this,s,x,f,o),this._prevBlock=a}});function n(s,x,o,f){var a,c=this._iv;c?(a=c.slice(0),this._iv=void 0):a=this._prevBlock,f.encryptBlock(a,0);for(var E=0;E<o;E++)s[x+E]^=a[E]}return r}(),t.mode.CFB})}(Ct)),Ct.exports}var Dt={exports:{}},ge;function mr(){return ge||(ge=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.CTR=function(){var r=t.lib.BlockCipherMode.extend(),n=r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=this._iv,c=this._counter;a&&(c=this._counter=a.slice(0),this._iv=void 0);var E=c.slice(0);o.encryptBlock(E,0),c[f-1]=c[f-1]+1|0;for(var h=0;h<f;h++)s[x+h]^=E[h]}});return r.Decryptor=n,r}(),t.mode.CTR})}(Dt)),Dt.exports}var yt={exports:{}},Ee;function Sr(){return Ee||(Ee=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){/** @preserve
+  			*/return function(r){var n=t,s=n.lib,x=s.WordArray,o=s.Hasher,f=n.algo,a=x.create([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,7,4,13,1,10,6,15,3,12,0,9,5,2,14,11,8,3,10,14,4,9,15,8,1,2,7,0,6,13,11,5,12,1,9,11,10,0,8,12,4,13,3,7,15,14,5,6,2,4,0,5,9,7,12,2,10,14,1,3,8,11,6,15,13]),c=x.create([5,14,7,0,9,2,11,4,13,6,15,8,1,10,3,12,6,11,3,7,0,13,5,10,14,15,8,12,4,9,1,2,15,5,1,3,7,14,6,9,11,8,12,2,10,0,4,13,8,6,4,1,3,11,15,0,5,12,2,13,9,7,10,14,12,15,10,4,1,5,8,7,6,2,13,14,0,3,9,11]),B=x.create([11,14,15,12,5,8,7,9,11,13,14,15,6,7,9,8,7,6,8,13,11,9,7,15,7,12,15,9,11,7,13,12,11,13,6,7,14,9,13,15,14,8,13,6,5,12,7,5,11,12,14,15,14,15,9,8,9,14,5,6,8,6,5,12,9,15,5,11,6,8,13,12,5,12,13,14,11,8,5,6]),h=x.create([8,9,9,11,13,15,15,5,7,7,8,11,14,14,12,6,9,13,15,7,12,8,9,11,7,7,12,7,6,15,13,11,9,7,15,11,8,6,6,14,12,13,5,14,13,13,7,5,15,5,8,11,14,14,6,14,6,9,12,9,12,5,15,8,8,5,12,9,12,5,14,6,8,13,6,5,15,13,11,11]),u=x.create([0,1518500249,1859775393,2400959708,2840853838]),v=x.create([1352829926,1548603684,1836072691,2053994217,0]),g=f.RIPEMD160=o.extend({_doReset:function(){this._hash=x.create([1732584193,4023233417,2562383102,271733878,3285377520])},_doProcessBlock:function(C,_){for(var m=0;m<16;m++){var I=_+m,D=C[I];C[I]=(D<<8|D>>>24)&16711935|(D<<24|D>>>8)&4278255360}var y=this._hash.words,T=u.words,w=v.words,H=a.words,N=c.words,L=B.words,Z=h.words,K,$,G,S,R,P,k,X,j,r0;P=K=y[0],k=$=y[1],X=G=y[2],j=S=y[3],r0=R=y[4];for(var M,m=0;m<80;m+=1)M=K+C[_+H[m]]|0,m<16?M+=l($,G,S)+T[0]:m<32?M+=A($,G,S)+T[1]:m<48?M+=d($,G,S)+T[2]:m<64?M+=p($,G,S)+T[3]:M+=E($,G,S)+T[4],M=M|0,M=F(M,L[m]),M=M+R|0,K=R,R=S,S=F(G,10),G=$,$=M,M=P+C[_+N[m]]|0,m<16?M+=E(k,X,j)+w[0]:m<32?M+=p(k,X,j)+w[1]:m<48?M+=d(k,X,j)+w[2]:m<64?M+=A(k,X,j)+w[3]:M+=l(k,X,j)+w[4],M=M|0,M=F(M,Z[m]),M=M+r0|0,P=r0,r0=j,j=F(X,10),X=k,k=M;M=y[1]+G+j|0,y[1]=y[2]+S+r0|0,y[2]=y[3]+R+P|0,y[3]=y[4]+K+k|0,y[4]=y[0]+$+X|0,y[0]=M},_doFinalize:function(){var C=this._data,_=C.words,m=this._nDataBytes*8,I=C.sigBytes*8;_[I>>>5]|=128<<24-I%32,_[(I+64>>>9<<4)+14]=(m<<8|m>>>24)&16711935|(m<<24|m>>>8)&4278255360,C.sigBytes=(_.length+1)*4,this._process();for(var D=this._hash,y=D.words,T=0;T<5;T++){var w=y[T];y[T]=(w<<8|w>>>24)&16711935|(w<<24|w>>>8)&4278255360}return D},clone:function(){var C=o.clone.call(this);return C._hash=this._hash.clone(),C}});function l(C,_,m){return C^_^m}function A(C,_,m){return C&_|~C&m}function d(C,_,m){return(C|~_)^m}function p(C,_,m){return C&m|_&~m}function E(C,_,m){return C^(_|~m)}function F(C,_){return C<<_|C>>>32-_}n.RIPEMD160=o._createHelper(g),n.HmacRIPEMD160=o._createHmacHelper(g)}(),t.RIPEMD160})}(pt)),pt.exports}var gt={exports:{}},le;function Et(){return le||(le=1,function(i,e){(function(t,r){i.exports=r(V())})(W,function(t){(function(){var r=t,n=r.lib,s=n.Base,x=r.enc,o=x.Utf8,f=r.algo;f.HMAC=s.extend({init:function(a,c){a=this._hasher=new a.init,typeof c=="string"&&(c=o.parse(c));var B=a.blockSize,h=B*4;c.sigBytes>h&&(c=a.finalize(c)),c.clamp();for(var u=this._oKey=c.clone(),v=this._iKey=c.clone(),g=u.words,l=v.words,A=0;A<B;A++)g[A]^=1549556828,l[A]^=909522486;u.sigBytes=v.sigBytes=h,this.reset()},reset:function(){var a=this._hasher;a.reset(),a.update(this._iKey)},update:function(a){return this._hasher.update(a),this},finalize:function(a){var c=this._hasher,B=c.finalize(a);c.reset();var h=c.finalize(this._oKey.clone().concat(B));return h}})})()})}(gt)),gt.exports}var Bt={exports:{}},de;function Sr(){return de||(de=1,function(i,e){(function(t,r,n){i.exports=r(V(),ht(),Et())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Base,x=n.WordArray,o=r.algo,f=o.SHA256,a=o.HMAC,c=o.PBKDF2=s.extend({cfg:s.extend({keySize:128/32,hasher:f,iterations:25e4}),init:function(B){this.cfg=this.cfg.extend(B)},compute:function(B,h){for(var u=this.cfg,v=a.create(u.hasher,B),g=x.create(),l=x.create([1]),A=g.words,d=l.words,p=u.keySize,E=u.iterations;A.length<p;){var F=v.update(h).finalize(l);v.reset();for(var C=F.words,_=C.length,m=F,I=1;I<E;I++){m=v.finalize(m),v.reset();for(var D=m.words,y=0;y<_;y++)C[y]^=D[y]}g.concat(F),d[0]++}return g.sigBytes=p*4,g}});r.PBKDF2=function(B,h,u){return c.create(u).compute(B,h)}}(),t.PBKDF2})}(Bt)),Bt.exports}var Ct={exports:{}},ve;function g0(){return ve||(ve=1,function(i,e){(function(t,r,n){i.exports=r(V(),se(),Et())})(W,function(t){return function(){var r=t,n=r.lib,s=n.Base,x=n.WordArray,o=r.algo,f=o.MD5,a=o.EvpKDF=s.extend({cfg:s.extend({keySize:128/32,hasher:f,iterations:1}),init:function(c){this.cfg=this.cfg.extend(c)},compute:function(c,B){for(var h,u=this.cfg,v=u.hasher.create(),g=x.create(),l=g.words,A=u.keySize,d=u.iterations;l.length<A;){h&&v.update(h),h=v.update(c).finalize(B),v.reset();for(var p=1;p<d;p++)h=v.finalize(h),v.reset();g.concat(h)}return g.sigBytes=A*4,g}});r.EvpKDF=function(c,B,h){return a.create(h).compute(c,B)}}(),t.EvpKDF})}(Ct)),Ct.exports}var Dt={exports:{}},pe;function t0(){return pe||(pe=1,function(i,e){(function(t,r,n){i.exports=r(V(),g0())})(W,function(t){t.lib.Cipher||function(r){var n=t,s=n.lib,x=s.Base,o=s.WordArray,f=s.BufferedBlockAlgorithm,a=n.enc;a.Utf8;var c=a.Base64,B=n.algo,h=B.EvpKDF,u=s.Cipher=f.extend({cfg:x.extend(),createEncryptor:function(D,y){return this.create(this._ENC_XFORM_MODE,D,y)},createDecryptor:function(D,y){return this.create(this._DEC_XFORM_MODE,D,y)},init:function(D,y,T){this.cfg=this.cfg.extend(T),this._xformMode=D,this._key=y,this.reset()},reset:function(){f.reset.call(this),this._doReset()},process:function(D){return this._append(D),this._process()},finalize:function(D){D&&this._append(D);var y=this._doFinalize();return y},keySize:128/32,ivSize:128/32,_ENC_XFORM_MODE:1,_DEC_XFORM_MODE:2,_createHelper:function(){function D(y){return typeof y=="string"?I:C}return function(y){return{encrypt:function(T,w,H){return D(w).encrypt(y,T,w,H)},decrypt:function(T,w,H){return D(w).decrypt(y,T,w,H)}}}}()});s.StreamCipher=u.extend({_doFinalize:function(){var D=this._process(!0);return D},blockSize:1});var v=n.mode={},g=s.BlockCipherMode=x.extend({createEncryptor:function(D,y){return this.Encryptor.create(D,y)},createDecryptor:function(D,y){return this.Decryptor.create(D,y)},init:function(D,y){this._cipher=D,this._iv=y}}),l=v.CBC=function(){var D=g.extend();D.Encryptor=D.extend({processBlock:function(T,w){var H=this._cipher,N=H.blockSize;y.call(this,T,w,N),H.encryptBlock(T,w),this._prevBlock=T.slice(w,w+N)}}),D.Decryptor=D.extend({processBlock:function(T,w){var H=this._cipher,N=H.blockSize,L=T.slice(w,w+N);H.decryptBlock(T,w),y.call(this,T,w,N),this._prevBlock=L}});function y(T,w,H){var N,L=this._iv;L?(N=L,this._iv=r):N=this._prevBlock;for(var Z=0;Z<H;Z++)T[w+Z]^=N[Z]}return D}(),A=n.pad={},d=A.Pkcs7={pad:function(D,y){for(var T=y*4,w=T-D.sigBytes%T,H=w<<24|w<<16|w<<8|w,N=[],L=0;L<w;L+=4)N.push(H);var Z=o.create(N,w);D.concat(Z)},unpad:function(D){var y=D.words[D.sigBytes-1>>>2]&255;D.sigBytes-=y}};s.BlockCipher=u.extend({cfg:u.cfg.extend({mode:l,padding:d}),reset:function(){var D;u.reset.call(this);var y=this.cfg,T=y.iv,w=y.mode;this._xformMode==this._ENC_XFORM_MODE?D=w.createEncryptor:(D=w.createDecryptor,this._minBufferSize=1),this._mode&&this._mode.__creator==D?this._mode.init(this,T&&T.words):(this._mode=D.call(w,this,T&&T.words),this._mode.__creator=D)},_doProcessBlock:function(D,y){this._mode.processBlock(D,y)},_doFinalize:function(){var D,y=this.cfg.padding;return this._xformMode==this._ENC_XFORM_MODE?(y.pad(this._data,this.blockSize),D=this._process(!0)):(D=this._process(!0),y.unpad(D)),D},blockSize:128/32});var p=s.CipherParams=x.extend({init:function(D){this.mixIn(D)},toString:function(D){return(D||this.formatter).stringify(this)}}),E=n.format={},F=E.OpenSSL={stringify:function(D){var y,T=D.ciphertext,w=D.salt;return w?y=o.create([1398893684,1701076831]).concat(w).concat(T):y=T,y.toString(c)},parse:function(D){var y,T=c.parse(D),w=T.words;return w[0]==1398893684&&w[1]==1701076831&&(y=o.create(w.slice(2,4)),w.splice(0,4),T.sigBytes-=16),p.create({ciphertext:T,salt:y})}},C=s.SerializableCipher=x.extend({cfg:x.extend({format:F}),encrypt:function(D,y,T,w){w=this.cfg.extend(w);var H=D.createEncryptor(T,w),N=H.finalize(y),L=H.cfg;return p.create({ciphertext:N,key:T,iv:L.iv,algorithm:D,mode:L.mode,padding:L.padding,blockSize:D.blockSize,formatter:w.format})},decrypt:function(D,y,T,w){w=this.cfg.extend(w),y=this._parse(y,w.format);var H=D.createDecryptor(T,w).finalize(y.ciphertext);return H},_parse:function(D,y){return typeof D=="string"?y.parse(D,this):D}}),_=n.kdf={},m=_.OpenSSL={execute:function(D,y,T,w,H){if(w||(w=o.random(64/8)),H)var N=h.create({keySize:y+T,hasher:H}).compute(D,w);else var N=h.create({keySize:y+T}).compute(D,w);var L=o.create(N.words.slice(y),T*4);return N.sigBytes=y*4,p.create({key:N,iv:L,salt:w})}},I=s.PasswordBasedCipher=C.extend({cfg:C.cfg.extend({kdf:m}),encrypt:function(D,y,T,w){w=this.cfg.extend(w);var H=w.kdf.execute(T,D.keySize,D.ivSize,w.salt,w.hasher);w.iv=H.iv;var N=C.encrypt.call(this,D,y,H.key,w);return N.mixIn(H),N},decrypt:function(D,y,T,w){w=this.cfg.extend(w),y=this._parse(y,w.format);var H=w.kdf.execute(T,D.keySize,D.ivSize,y.salt,w.hasher);w.iv=H.iv;var N=C.decrypt.call(this,D,y,H.key,w);return N}})}()})}(Dt)),Dt.exports}var At={exports:{}},ge;function wr(){return ge||(ge=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.CFB=function(){var r=t.lib.BlockCipherMode.extend();r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize;n.call(this,s,x,f,o),this._prevBlock=s.slice(x,x+f)}}),r.Decryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=s.slice(x,x+f);n.call(this,s,x,f,o),this._prevBlock=a}});function n(s,x,o,f){var a,c=this._iv;c?(a=c.slice(0),this._iv=void 0):a=this._prevBlock,f.encryptBlock(a,0);for(var B=0;B<o;B++)s[x+B]^=a[B]}return r}(),t.mode.CFB})}(At)),At.exports}var yt={exports:{}},Ee;function Tr(){return Ee||(Ee=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.CTR=function(){var r=t.lib.BlockCipherMode.extend(),n=r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=this._iv,c=this._counter;a&&(c=this._counter=a.slice(0),this._iv=void 0);var B=c.slice(0);o.encryptBlock(B,0),c[f-1]=c[f-1]+1|0;for(var h=0;h<f;h++)s[x+h]^=B[h]}});return r.Decryptor=n,r}(),t.mode.CTR})}(yt)),yt.exports}var Ft={exports:{}},Be;function kr(){return Be||(Be=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){/** @preserve
  * Counter block mode compatible with  Dr Brian Gladman fileenc.c
  * derived from CryptoJS.mode.CTR
  * Jan Hruby jhruby.web@gmail.com
- */return t.mode.CTRGladman=function(){var r=t.lib.BlockCipherMode.extend();function n(o){if((o>>24&255)===255){var f=o>>16&255,a=o>>8&255,c=o&255;f===255?(f=0,a===255?(a=0,c===255?c=0:++c):++a):++f,o=0,o+=f<<16,o+=a<<8,o+=c}else o+=1<<24;return o}function s(o){return(o[0]=n(o[0]))===0&&(o[1]=n(o[1])),o}var x=r.Encryptor=r.extend({processBlock:function(o,f){var a=this._cipher,c=a.blockSize,E=this._iv,h=this._counter;E&&(h=this._counter=E.slice(0),this._iv=void 0),s(h);var u=h.slice(0);a.encryptBlock(u,0);for(var v=0;v<c;v++)o[f+v]^=u[v]}});return r.Decryptor=x,r}(),t.mode.CTRGladman})}(yt)),yt.exports}var Ft={exports:{}},Be;function wr(){return Be||(Be=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.OFB=function(){var r=t.lib.BlockCipherMode.extend(),n=r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=this._iv,c=this._keystream;a&&(c=this._keystream=a.slice(0),this._iv=void 0),o.encryptBlock(c,0);for(var E=0;E<f;E++)s[x+E]^=c[E]}});return r.Decryptor=n,r}(),t.mode.OFB})}(Ft)),Ft.exports}var _t={exports:{}},Ae;function Tr(){return Ae||(Ae=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.ECB=function(){var r=t.lib.BlockCipherMode.extend();return r.Encryptor=r.extend({processBlock:function(n,s){this._cipher.encryptBlock(n,s)}}),r.Decryptor=r.extend({processBlock:function(n,s){this._cipher.decryptBlock(n,s)}}),r}(),t.mode.ECB})}(_t)),_t.exports}var bt={exports:{}},Ce;function Rr(){return Ce||(Ce=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.AnsiX923={pad:function(r,n){var s=r.sigBytes,x=n*4,o=x-s%x,f=s+o-1;r.clamp(),r.words[f>>>2]|=o<<24-f%4*8,r.sigBytes+=o},unpad:function(r){var n=r.words[r.sigBytes-1>>>2]&255;r.sigBytes-=n}},t.pad.Ansix923})}(bt)),bt.exports}var mt={exports:{}},De;function kr(){return De||(De=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.Iso10126={pad:function(r,n){var s=n*4,x=s-r.sigBytes%s;r.concat(t.lib.WordArray.random(x-1)).concat(t.lib.WordArray.create([x<<24],1))},unpad:function(r){var n=r.words[r.sigBytes-1>>>2]&255;r.sigBytes-=n}},t.pad.Iso10126})}(mt)),mt.exports}var St={exports:{}},ye;function Pr(){return ye||(ye=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.Iso97971={pad:function(r,n){r.concat(t.lib.WordArray.create([2147483648],1)),t.pad.ZeroPadding.pad(r,n)},unpad:function(r){t.pad.ZeroPadding.unpad(r),r.sigBytes--}},t.pad.Iso97971})}(St)),St.exports}var wt={exports:{}},Fe;function Or(){return Fe||(Fe=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.ZeroPadding={pad:function(r,n){var s=n*4;r.clamp(),r.sigBytes+=s-(r.sigBytes%s||s)},unpad:function(r){for(var n=r.words,s=r.sigBytes-1,s=r.sigBytes-1;s>=0;s--)if(n[s>>>2]>>>24-s%4*8&255){r.sigBytes=s+1;break}}},t.pad.ZeroPadding})}(wt)),wt.exports}var Tt={exports:{}},_e;function Ir(){return _e||(_e=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.NoPadding={pad:function(){},unpad:function(){}},t.pad.NoPadding})}(Tt)),Tt.exports}var Rt={exports:{}},be;function Hr(){return be||(be=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.CipherParams,o=n.enc,f=o.Hex,a=n.format;a.Hex={stringify:function(c){return c.ciphertext.toString(f)},parse:function(c){var E=f.parse(c);return x.create({ciphertext:E})}}}(),t.format.Hex})}(Rt)),Rt.exports}var kt={exports:{}},me;function qr(){return me||(me=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.BlockCipher,x=r.algo,o=[],f=[],a=[],c=[],E=[],h=[],u=[],v=[],p=[],l=[];(function(){for(var g=[],B=0;B<256;B++)B<128?g[B]=B<<1:g[B]=B<<1^283;for(var F=0,A=0,B=0;B<256;B++){var _=A^A<<1^A<<2^A<<3^A<<4;_=_>>>8^_&255^99,o[F]=_,f[_]=F;var b=g[F],P=g[b],D=g[P],y=g[_]*257^_*16843008;a[F]=y<<24|y>>>8,c[F]=y<<16|y>>>16,E[F]=y<<8|y>>>24,h[F]=y;var y=D*16843009^P*65537^b*257^F*16843008;u[_]=y<<24|y>>>8,v[_]=y<<16|y>>>16,p[_]=y<<8|y>>>24,l[_]=y,F?(F=b^g[g[g[D^b]]],A^=g[g[A]]):F=A=1}})();var C=[0,1,2,4,8,16,32,64,128,27,54],d=x.AES=s.extend({_doReset:function(){var g;if(!(this._nRounds&&this._keyPriorReset===this._key)){for(var B=this._keyPriorReset=this._key,F=B.words,A=B.sigBytes/4,_=this._nRounds=A+6,b=(_+1)*4,P=this._keySchedule=[],D=0;D<b;D++)D<A?P[D]=F[D]:(g=P[D-1],D%A?A>6&&D%A==4&&(g=o[g>>>24]<<24|o[g>>>16&255]<<16|o[g>>>8&255]<<8|o[g&255]):(g=g<<8|g>>>24,g=o[g>>>24]<<24|o[g>>>16&255]<<16|o[g>>>8&255]<<8|o[g&255],g^=C[D/A|0]<<24),P[D]=P[D-A]^g);for(var y=this._invKeySchedule=[],T=0;T<b;T++){var D=b-T;if(T%4)var g=P[D];else var g=P[D-4];T<4||D<=4?y[T]=g:y[T]=u[o[g>>>24]]^v[o[g>>>16&255]]^p[o[g>>>8&255]]^l[o[g&255]]}}},encryptBlock:function(g,B){this._doCryptBlock(g,B,this._keySchedule,a,c,E,h,o)},decryptBlock:function(g,B){var F=g[B+1];g[B+1]=g[B+3],g[B+3]=F,this._doCryptBlock(g,B,this._invKeySchedule,u,v,p,l,f);var F=g[B+1];g[B+1]=g[B+3],g[B+3]=F},_doCryptBlock:function(g,B,F,A,_,b,P,D){for(var y=this._nRounds,T=g[B]^F[0],w=g[B+1]^F[1],H=g[B+2]^F[2],N=g[B+3]^F[3],L=4,Z=1;Z<y;Z++){var K=A[T>>>24]^_[w>>>16&255]^b[H>>>8&255]^P[N&255]^F[L++],j=A[w>>>24]^_[H>>>16&255]^b[N>>>8&255]^P[T&255]^F[L++],$=A[H>>>24]^_[N>>>16&255]^b[T>>>8&255]^P[w&255]^F[L++],S=A[N>>>24]^_[T>>>16&255]^b[w>>>8&255]^P[H&255]^F[L++];T=K,w=j,H=$,N=S}var K=(D[T>>>24]<<24|D[w>>>16&255]<<16|D[H>>>8&255]<<8|D[N&255])^F[L++],j=(D[w>>>24]<<24|D[H>>>16&255]<<16|D[N>>>8&255]<<8|D[T&255])^F[L++],$=(D[H>>>24]<<24|D[N>>>16&255]<<16|D[T>>>8&255]<<8|D[w&255])^F[L++],S=(D[N>>>24]<<24|D[T>>>16&255]<<16|D[w>>>8&255]<<8|D[H&255])^F[L++];g[B]=K,g[B+1]=j,g[B+2]=$,g[B+3]=S},keySize:256/32});r.AES=s._createHelper(d)}(),t.AES})}(kt)),kt.exports}var Pt={exports:{}},Se;function Nr(){return Se||(Se=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=n.BlockCipher,o=r.algo,f=[57,49,41,33,25,17,9,1,58,50,42,34,26,18,10,2,59,51,43,35,27,19,11,3,60,52,44,36,63,55,47,39,31,23,15,7,62,54,46,38,30,22,14,6,61,53,45,37,29,21,13,5,28,20,12,4],a=[14,17,11,24,1,5,3,28,15,6,21,10,23,19,12,4,26,8,16,7,27,20,13,2,41,52,31,37,47,55,30,40,51,45,33,48,44,49,39,56,34,53,46,42,50,36,29,32],c=[1,2,4,6,8,10,12,14,15,17,19,21,23,25,27,28],E=[{0:8421888,268435456:32768,536870912:8421378,805306368:2,1073741824:512,1342177280:8421890,1610612736:8389122,1879048192:8388608,2147483648:514,2415919104:8389120,2684354560:33280,2952790016:8421376,3221225472:32770,3489660928:8388610,3758096384:0,4026531840:33282,134217728:0,402653184:8421890,671088640:33282,939524096:32768,1207959552:8421888,1476395008:512,1744830464:8421378,2013265920:2,2281701376:8389120,2550136832:33280,2818572288:8421376,3087007744:8389122,3355443200:8388610,3623878656:32770,3892314112:514,4160749568:8388608,1:32768,268435457:2,536870913:8421888,805306369:8388608,1073741825:8421378,1342177281:33280,1610612737:512,1879048193:8389122,2147483649:8421890,2415919105:8421376,2684354561:8388610,2952790017:33282,3221225473:514,3489660929:8389120,3758096385:32770,4026531841:0,134217729:8421890,402653185:8421376,671088641:8388608,939524097:512,1207959553:32768,1476395009:8388610,1744830465:2,2013265921:33282,2281701377:32770,2550136833:8389122,2818572289:514,3087007745:8421888,3355443201:8389120,3623878657:0,3892314113:33280,4160749569:8421378},{0:1074282512,16777216:16384,33554432:524288,50331648:1074266128,67108864:1073741840,83886080:1074282496,100663296:1073758208,117440512:16,134217728:540672,150994944:1073758224,167772160:1073741824,184549376:540688,201326592:524304,218103808:0,234881024:16400,251658240:1074266112,8388608:1073758208,25165824:540688,41943040:16,58720256:1073758224,75497472:1074282512,92274688:1073741824,109051904:524288,125829120:1074266128,142606336:524304,159383552:0,176160768:16384,192937984:1074266112,209715200:1073741840,226492416:540672,243269632:1074282496,260046848:16400,268435456:0,285212672:1074266128,301989888:1073758224,318767104:1074282496,335544320:1074266112,352321536:16,369098752:540688,385875968:16384,402653184:16400,419430400:524288,436207616:524304,452984832:1073741840,469762048:540672,486539264:1073758208,503316480:1073741824,520093696:1074282512,276824064:540688,293601280:524288,310378496:1074266112,327155712:16384,343932928:1073758208,360710144:1074282512,377487360:16,394264576:1073741824,411041792:1074282496,427819008:1073741840,444596224:1073758224,461373440:524304,478150656:0,494927872:16400,511705088:1074266128,528482304:540672},{0:260,1048576:0,2097152:67109120,3145728:65796,4194304:65540,5242880:67108868,6291456:67174660,7340032:67174400,8388608:67108864,9437184:67174656,10485760:65792,11534336:67174404,12582912:67109124,13631488:65536,14680064:4,15728640:256,524288:67174656,1572864:67174404,2621440:0,3670016:67109120,4718592:67108868,5767168:65536,6815744:65540,7864320:260,8912896:4,9961472:256,11010048:67174400,12058624:65796,13107200:65792,14155776:67109124,15204352:67174660,16252928:67108864,16777216:67174656,17825792:65540,18874368:65536,19922944:67109120,20971520:256,22020096:67174660,23068672:67108868,24117248:0,25165824:67109124,26214400:67108864,27262976:4,28311552:65792,29360128:67174400,30408704:260,31457280:65796,32505856:67174404,17301504:67108864,18350080:260,19398656:67174656,20447232:0,21495808:65540,22544384:67109120,23592960:256,24641536:67174404,25690112:65536,26738688:67174660,27787264:65796,28835840:67108868,29884416:67109124,30932992:67174400,31981568:4,33030144:65792},{0:2151682048,65536:2147487808,131072:4198464,196608:2151677952,262144:0,327680:4198400,393216:2147483712,458752:4194368,524288:2147483648,589824:4194304,655360:64,720896:2147487744,786432:2151678016,851968:4160,917504:4096,983040:2151682112,32768:2147487808,98304:64,163840:2151678016,229376:2147487744,294912:4198400,360448:2151682112,425984:0,491520:2151677952,557056:4096,622592:2151682048,688128:4194304,753664:4160,819200:2147483648,884736:4194368,950272:4198464,1015808:2147483712,1048576:4194368,1114112:4198400,1179648:2147483712,1245184:0,1310720:4160,1376256:2151678016,1441792:2151682048,1507328:2147487808,1572864:2151682112,1638400:2147483648,1703936:2151677952,1769472:4198464,1835008:2147487744,1900544:4194304,1966080:64,2031616:4096,1081344:2151677952,1146880:2151682112,1212416:0,1277952:4198400,1343488:4194368,1409024:2147483648,1474560:2147487808,1540096:64,1605632:2147483712,1671168:4096,1736704:2147487744,1802240:2151678016,1867776:4160,1933312:2151682048,1998848:4194304,2064384:4198464},{0:128,4096:17039360,8192:262144,12288:536870912,16384:537133184,20480:16777344,24576:553648256,28672:262272,32768:16777216,36864:537133056,40960:536871040,45056:553910400,49152:553910272,53248:0,57344:17039488,61440:553648128,2048:17039488,6144:553648256,10240:128,14336:17039360,18432:262144,22528:537133184,26624:553910272,30720:536870912,34816:537133056,38912:0,43008:553910400,47104:16777344,51200:536871040,55296:553648128,59392:16777216,63488:262272,65536:262144,69632:128,73728:536870912,77824:553648256,81920:16777344,86016:553910272,90112:537133184,94208:16777216,98304:553910400,102400:553648128,106496:17039360,110592:537133056,114688:262272,118784:536871040,122880:0,126976:17039488,67584:553648256,71680:16777216,75776:17039360,79872:537133184,83968:536870912,88064:17039488,92160:128,96256:553910272,100352:262272,104448:553910400,108544:0,112640:553648128,116736:16777344,120832:262144,124928:537133056,129024:536871040},{0:268435464,256:8192,512:270532608,768:270540808,1024:268443648,1280:2097152,1536:2097160,1792:268435456,2048:0,2304:268443656,2560:2105344,2816:8,3072:270532616,3328:2105352,3584:8200,3840:270540800,128:270532608,384:270540808,640:8,896:2097152,1152:2105352,1408:268435464,1664:268443648,1920:8200,2176:2097160,2432:8192,2688:268443656,2944:270532616,3200:0,3456:270540800,3712:2105344,3968:268435456,4096:268443648,4352:270532616,4608:270540808,4864:8200,5120:2097152,5376:268435456,5632:268435464,5888:2105344,6144:2105352,6400:0,6656:8,6912:270532608,7168:8192,7424:268443656,7680:270540800,7936:2097160,4224:8,4480:2105344,4736:2097152,4992:268435464,5248:268443648,5504:8200,5760:270540808,6016:270532608,6272:270540800,6528:270532616,6784:8192,7040:2105352,7296:2097160,7552:0,7808:268435456,8064:268443656},{0:1048576,16:33555457,32:1024,48:1049601,64:34604033,80:0,96:1,112:34603009,128:33555456,144:1048577,160:33554433,176:34604032,192:34603008,208:1025,224:1049600,240:33554432,8:34603009,24:0,40:33555457,56:34604032,72:1048576,88:33554433,104:33554432,120:1025,136:1049601,152:33555456,168:34603008,184:1048577,200:1024,216:34604033,232:1,248:1049600,256:33554432,272:1048576,288:33555457,304:34603009,320:1048577,336:33555456,352:34604032,368:1049601,384:1025,400:34604033,416:1049600,432:1,448:0,464:34603008,480:33554433,496:1024,264:1049600,280:33555457,296:34603009,312:1,328:33554432,344:1048576,360:1025,376:34604032,392:33554433,408:34603008,424:0,440:34604033,456:1049601,472:1024,488:33555456,504:1048577},{0:134219808,1:131072,2:134217728,3:32,4:131104,5:134350880,6:134350848,7:2048,8:134348800,9:134219776,10:133120,11:134348832,12:2080,13:0,14:134217760,15:133152,2147483648:2048,2147483649:134350880,2147483650:134219808,2147483651:134217728,2147483652:134348800,2147483653:133120,2147483654:133152,2147483655:32,2147483656:134217760,2147483657:2080,2147483658:131104,2147483659:134350848,2147483660:0,2147483661:134348832,2147483662:134219776,2147483663:131072,16:133152,17:134350848,18:32,19:2048,20:134219776,21:134217760,22:134348832,23:131072,24:0,25:131104,26:134348800,27:134219808,28:134350880,29:133120,30:2080,31:134217728,2147483664:131072,2147483665:2048,2147483666:134348832,2147483667:133152,2147483668:32,2147483669:134348800,2147483670:134217728,2147483671:134219808,2147483672:134350880,2147483673:134217760,2147483674:134219776,2147483675:0,2147483676:133120,2147483677:2080,2147483678:131104,2147483679:134350848}],h=[4160749569,528482304,33030144,2064384,129024,8064,504,2147483679],u=o.DES=x.extend({_doReset:function(){for(var C=this._key,d=C.words,g=[],B=0;B<56;B++){var F=f[B]-1;g[B]=d[F>>>5]>>>31-F%32&1}for(var A=this._subKeys=[],_=0;_<16;_++){for(var b=A[_]=[],P=c[_],B=0;B<24;B++)b[B/6|0]|=g[(a[B]-1+P)%28]<<31-B%6,b[4+(B/6|0)]|=g[28+(a[B+24]-1+P)%28]<<31-B%6;b[0]=b[0]<<1|b[0]>>>31;for(var B=1;B<7;B++)b[B]=b[B]>>>(B-1)*4+3;b[7]=b[7]<<5|b[7]>>>27}for(var D=this._invSubKeys=[],B=0;B<16;B++)D[B]=A[15-B]},encryptBlock:function(C,d){this._doCryptBlock(C,d,this._subKeys)},decryptBlock:function(C,d){this._doCryptBlock(C,d,this._invSubKeys)},_doCryptBlock:function(C,d,g){this._lBlock=C[d],this._rBlock=C[d+1],v.call(this,4,252645135),v.call(this,16,65535),p.call(this,2,858993459),p.call(this,8,16711935),v.call(this,1,1431655765);for(var B=0;B<16;B++){for(var F=g[B],A=this._lBlock,_=this._rBlock,b=0,P=0;P<8;P++)b|=E[P][((_^F[P])&h[P])>>>0];this._lBlock=_,this._rBlock=A^b}var D=this._lBlock;this._lBlock=this._rBlock,this._rBlock=D,v.call(this,1,1431655765),p.call(this,8,16711935),p.call(this,2,858993459),v.call(this,16,65535),v.call(this,4,252645135),C[d]=this._lBlock,C[d+1]=this._rBlock},keySize:64/32,ivSize:64/32,blockSize:64/32});function v(C,d){var g=(this._lBlock>>>C^this._rBlock)&d;this._rBlock^=g,this._lBlock^=g<<C}function p(C,d){var g=(this._rBlock>>>C^this._lBlock)&d;this._lBlock^=g,this._rBlock^=g<<C}r.DES=x._createHelper(u);var l=o.TripleDES=x.extend({_doReset:function(){var C=this._key,d=C.words;if(d.length!==2&&d.length!==4&&d.length<6)throw new Error("Invalid key length - 3DES requires the key length to be 64, 128, 192 or >192.");var g=d.slice(0,2),B=d.length<4?d.slice(0,2):d.slice(2,4),F=d.length<6?d.slice(0,2):d.slice(4,6);this._des1=u.createEncryptor(s.create(g)),this._des2=u.createEncryptor(s.create(B)),this._des3=u.createEncryptor(s.create(F))},encryptBlock:function(C,d){this._des1.encryptBlock(C,d),this._des2.decryptBlock(C,d),this._des3.encryptBlock(C,d)},decryptBlock:function(C,d){this._des3.decryptBlock(C,d),this._des2.encryptBlock(C,d),this._des1.decryptBlock(C,d)},keySize:192/32,ivSize:64/32,blockSize:64/32});r.TripleDES=x._createHelper(l)}(),t.TripleDES})}(Pt)),Pt.exports}var Ot={exports:{}},we;function zr(){return we||(we=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=x.RC4=s.extend({_doReset:function(){for(var c=this._key,E=c.words,h=c.sigBytes,u=this._S=[],v=0;v<256;v++)u[v]=v;for(var v=0,p=0;v<256;v++){var l=v%h,C=E[l>>>2]>>>24-l%4*8&255;p=(p+u[v]+C)%256;var d=u[v];u[v]=u[p],u[p]=d}this._i=this._j=0},_doProcessBlock:function(c,E){c[E]^=f.call(this)},keySize:256/32,ivSize:0});function f(){for(var c=this._S,E=this._i,h=this._j,u=0,v=0;v<4;v++){E=(E+1)%256,h=(h+c[E])%256;var p=c[E];c[E]=c[h],c[h]=p,u|=c[(c[E]+c[h])%256]<<24-v*8}return this._i=E,this._j=h,u}r.RC4=s._createHelper(o);var a=x.RC4Drop=o.extend({cfg:o.cfg.extend({drop:192}),_doReset:function(){o._doReset.call(this);for(var c=this.cfg.drop;c>0;c--)f.call(this)}});r.RC4Drop=s._createHelper(a)}(),t.RC4})}(Ot)),Ot.exports}var It={exports:{}},Te;function Wr(){return Te||(Te=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=[],f=[],a=[],c=x.Rabbit=s.extend({_doReset:function(){for(var h=this._key.words,u=this.cfg.iv,v=0;v<4;v++)h[v]=(h[v]<<8|h[v]>>>24)&16711935|(h[v]<<24|h[v]>>>8)&4278255360;var p=this._X=[h[0],h[3]<<16|h[2]>>>16,h[1],h[0]<<16|h[3]>>>16,h[2],h[1]<<16|h[0]>>>16,h[3],h[2]<<16|h[1]>>>16],l=this._C=[h[2]<<16|h[2]>>>16,h[0]&4294901760|h[1]&65535,h[3]<<16|h[3]>>>16,h[1]&4294901760|h[2]&65535,h[0]<<16|h[0]>>>16,h[2]&4294901760|h[3]&65535,h[1]<<16|h[1]>>>16,h[3]&4294901760|h[0]&65535];this._b=0;for(var v=0;v<4;v++)E.call(this);for(var v=0;v<8;v++)l[v]^=p[v+4&7];if(u){var C=u.words,d=C[0],g=C[1],B=(d<<8|d>>>24)&16711935|(d<<24|d>>>8)&4278255360,F=(g<<8|g>>>24)&16711935|(g<<24|g>>>8)&4278255360,A=B>>>16|F&4294901760,_=F<<16|B&65535;l[0]^=B,l[1]^=A,l[2]^=F,l[3]^=_,l[4]^=B,l[5]^=A,l[6]^=F,l[7]^=_;for(var v=0;v<4;v++)E.call(this)}},_doProcessBlock:function(h,u){var v=this._X;E.call(this),o[0]=v[0]^v[5]>>>16^v[3]<<16,o[1]=v[2]^v[7]>>>16^v[5]<<16,o[2]=v[4]^v[1]>>>16^v[7]<<16,o[3]=v[6]^v[3]>>>16^v[1]<<16;for(var p=0;p<4;p++)o[p]=(o[p]<<8|o[p]>>>24)&16711935|(o[p]<<24|o[p]>>>8)&4278255360,h[u+p]^=o[p]},blockSize:128/32,ivSize:64/32});function E(){for(var h=this._X,u=this._C,v=0;v<8;v++)f[v]=u[v];u[0]=u[0]+1295307597+this._b|0,u[1]=u[1]+3545052371+(u[0]>>>0<f[0]>>>0?1:0)|0,u[2]=u[2]+886263092+(u[1]>>>0<f[1]>>>0?1:0)|0,u[3]=u[3]+1295307597+(u[2]>>>0<f[2]>>>0?1:0)|0,u[4]=u[4]+3545052371+(u[3]>>>0<f[3]>>>0?1:0)|0,u[5]=u[5]+886263092+(u[4]>>>0<f[4]>>>0?1:0)|0,u[6]=u[6]+1295307597+(u[5]>>>0<f[5]>>>0?1:0)|0,u[7]=u[7]+3545052371+(u[6]>>>0<f[6]>>>0?1:0)|0,this._b=u[7]>>>0<f[7]>>>0?1:0;for(var v=0;v<8;v++){var p=h[v]+u[v],l=p&65535,C=p>>>16,d=((l*l>>>17)+l*C>>>15)+C*C,g=((p&4294901760)*p|0)+((p&65535)*p|0);a[v]=d^g}h[0]=a[0]+(a[7]<<16|a[7]>>>16)+(a[6]<<16|a[6]>>>16)|0,h[1]=a[1]+(a[0]<<8|a[0]>>>24)+a[7]|0,h[2]=a[2]+(a[1]<<16|a[1]>>>16)+(a[0]<<16|a[0]>>>16)|0,h[3]=a[3]+(a[2]<<8|a[2]>>>24)+a[1]|0,h[4]=a[4]+(a[3]<<16|a[3]>>>16)+(a[2]<<16|a[2]>>>16)|0,h[5]=a[5]+(a[4]<<8|a[4]>>>24)+a[3]|0,h[6]=a[6]+(a[5]<<16|a[5]>>>16)+(a[4]<<16|a[4]>>>16)|0,h[7]=a[7]+(a[6]<<8|a[6]>>>24)+a[5]|0}r.Rabbit=s._createHelper(c)}(),t.Rabbit})}(It)),It.exports}var Ht={exports:{}},Re;function Lr(){return Re||(Re=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=[],f=[],a=[],c=x.RabbitLegacy=s.extend({_doReset:function(){var h=this._key.words,u=this.cfg.iv,v=this._X=[h[0],h[3]<<16|h[2]>>>16,h[1],h[0]<<16|h[3]>>>16,h[2],h[1]<<16|h[0]>>>16,h[3],h[2]<<16|h[1]>>>16],p=this._C=[h[2]<<16|h[2]>>>16,h[0]&4294901760|h[1]&65535,h[3]<<16|h[3]>>>16,h[1]&4294901760|h[2]&65535,h[0]<<16|h[0]>>>16,h[2]&4294901760|h[3]&65535,h[1]<<16|h[1]>>>16,h[3]&4294901760|h[0]&65535];this._b=0;for(var l=0;l<4;l++)E.call(this);for(var l=0;l<8;l++)p[l]^=v[l+4&7];if(u){var C=u.words,d=C[0],g=C[1],B=(d<<8|d>>>24)&16711935|(d<<24|d>>>8)&4278255360,F=(g<<8|g>>>24)&16711935|(g<<24|g>>>8)&4278255360,A=B>>>16|F&4294901760,_=F<<16|B&65535;p[0]^=B,p[1]^=A,p[2]^=F,p[3]^=_,p[4]^=B,p[5]^=A,p[6]^=F,p[7]^=_;for(var l=0;l<4;l++)E.call(this)}},_doProcessBlock:function(h,u){var v=this._X;E.call(this),o[0]=v[0]^v[5]>>>16^v[3]<<16,o[1]=v[2]^v[7]>>>16^v[5]<<16,o[2]=v[4]^v[1]>>>16^v[7]<<16,o[3]=v[6]^v[3]>>>16^v[1]<<16;for(var p=0;p<4;p++)o[p]=(o[p]<<8|o[p]>>>24)&16711935|(o[p]<<24|o[p]>>>8)&4278255360,h[u+p]^=o[p]},blockSize:128/32,ivSize:64/32});function E(){for(var h=this._X,u=this._C,v=0;v<8;v++)f[v]=u[v];u[0]=u[0]+1295307597+this._b|0,u[1]=u[1]+3545052371+(u[0]>>>0<f[0]>>>0?1:0)|0,u[2]=u[2]+886263092+(u[1]>>>0<f[1]>>>0?1:0)|0,u[3]=u[3]+1295307597+(u[2]>>>0<f[2]>>>0?1:0)|0,u[4]=u[4]+3545052371+(u[3]>>>0<f[3]>>>0?1:0)|0,u[5]=u[5]+886263092+(u[4]>>>0<f[4]>>>0?1:0)|0,u[6]=u[6]+1295307597+(u[5]>>>0<f[5]>>>0?1:0)|0,u[7]=u[7]+3545052371+(u[6]>>>0<f[6]>>>0?1:0)|0,this._b=u[7]>>>0<f[7]>>>0?1:0;for(var v=0;v<8;v++){var p=h[v]+u[v],l=p&65535,C=p>>>16,d=((l*l>>>17)+l*C>>>15)+C*C,g=((p&4294901760)*p|0)+((p&65535)*p|0);a[v]=d^g}h[0]=a[0]+(a[7]<<16|a[7]>>>16)+(a[6]<<16|a[6]>>>16)|0,h[1]=a[1]+(a[0]<<8|a[0]>>>24)+a[7]|0,h[2]=a[2]+(a[1]<<16|a[1]>>>16)+(a[0]<<16|a[0]>>>16)|0,h[3]=a[3]+(a[2]<<8|a[2]>>>24)+a[1]|0,h[4]=a[4]+(a[3]<<16|a[3]>>>16)+(a[2]<<16|a[2]>>>16)|0,h[5]=a[5]+(a[4]<<8|a[4]>>>24)+a[3]|0,h[6]=a[6]+(a[5]<<16|a[5]>>>16)+(a[4]<<16|a[4]>>>16)|0,h[7]=a[7]+(a[6]<<8|a[6]>>>24)+a[5]|0}r.RabbitLegacy=s._createHelper(c)}(),t.RabbitLegacy})}(Ht)),Ht.exports}var qt={exports:{}},ke;function Mr(){return ke||(ke=1,function(i,e){(function(t,r,n){i.exports=r(V(),D0(),y0(),p0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.BlockCipher,x=r.algo;const o=16,f=[608135816,2242054355,320440878,57701188,2752067618,698298832,137296536,3964562569,1160258022,953160567,3193202383,887688300,3232508343,3380367581,1065670069,3041331479,2450970073,2306472731],a=[[3509652390,2564797868,805139163,3491422135,3101798381,1780907670,3128725573,4046225305,614570311,3012652279,134345442,2240740374,1667834072,1901547113,2757295779,4103290238,227898511,1921955416,1904987480,2182433518,2069144605,3260701109,2620446009,720527379,3318853667,677414384,3393288472,3101374703,2390351024,1614419982,1822297739,2954791486,3608508353,3174124327,2024746970,1432378464,3864339955,2857741204,1464375394,1676153920,1439316330,715854006,3033291828,289532110,2706671279,2087905683,3018724369,1668267050,732546397,1947742710,3462151702,2609353502,2950085171,1814351708,2050118529,680887927,999245976,1800124847,3300911131,1713906067,1641548236,4213287313,1216130144,1575780402,4018429277,3917837745,3693486850,3949271944,596196993,3549867205,258830323,2213823033,772490370,2760122372,1774776394,2652871518,566650946,4142492826,1728879713,2882767088,1783734482,3629395816,2517608232,2874225571,1861159788,326777828,3124490320,2130389656,2716951837,967770486,1724537150,2185432712,2364442137,1164943284,2105845187,998989502,3765401048,2244026483,1075463327,1455516326,1322494562,910128902,469688178,1117454909,936433444,3490320968,3675253459,1240580251,122909385,2157517691,634681816,4142456567,3825094682,3061402683,2540495037,79693498,3249098678,1084186820,1583128258,426386531,1761308591,1047286709,322548459,995290223,1845252383,2603652396,3431023940,2942221577,3202600964,3727903485,1712269319,422464435,3234572375,1170764815,3523960633,3117677531,1434042557,442511882,3600875718,1076654713,1738483198,4213154764,2393238008,3677496056,1014306527,4251020053,793779912,2902807211,842905082,4246964064,1395751752,1040244610,2656851899,3396308128,445077038,3742853595,3577915638,679411651,2892444358,2354009459,1767581616,3150600392,3791627101,3102740896,284835224,4246832056,1258075500,768725851,2589189241,3069724005,3532540348,1274779536,3789419226,2764799539,1660621633,3471099624,4011903706,913787905,3497959166,737222580,2514213453,2928710040,3937242737,1804850592,3499020752,2949064160,2386320175,2390070455,2415321851,4061277028,2290661394,2416832540,1336762016,1754252060,3520065937,3014181293,791618072,3188594551,3933548030,2332172193,3852520463,3043980520,413987798,3465142937,3030929376,4245938359,2093235073,3534596313,375366246,2157278981,2479649556,555357303,3870105701,2008414854,3344188149,4221384143,3956125452,2067696032,3594591187,2921233993,2428461,544322398,577241275,1471733935,610547355,4027169054,1432588573,1507829418,2025931657,3646575487,545086370,48609733,2200306550,1653985193,298326376,1316178497,3007786442,2064951626,458293330,2589141269,3591329599,3164325604,727753846,2179363840,146436021,1461446943,4069977195,705550613,3059967265,3887724982,4281599278,3313849956,1404054877,2845806497,146425753,1854211946],[1266315497,3048417604,3681880366,3289982499,290971e4,1235738493,2632868024,2414719590,3970600049,1771706367,1449415276,3266420449,422970021,1963543593,2690192192,3826793022,1062508698,1531092325,1804592342,2583117782,2714934279,4024971509,1294809318,4028980673,1289560198,2221992742,1669523910,35572830,157838143,1052438473,1016535060,1802137761,1753167236,1386275462,3080475397,2857371447,1040679964,2145300060,2390574316,1461121720,2956646967,4031777805,4028374788,33600511,2920084762,1018524850,629373528,3691585981,3515945977,2091462646,2486323059,586499841,988145025,935516892,3367335476,2599673255,2839830854,265290510,3972581182,2759138881,3795373465,1005194799,847297441,406762289,1314163512,1332590856,1866599683,4127851711,750260880,613907577,1450815602,3165620655,3734664991,3650291728,3012275730,3704569646,1427272223,778793252,1343938022,2676280711,2052605720,1946737175,3164576444,3914038668,3967478842,3682934266,1661551462,3294938066,4011595847,840292616,3712170807,616741398,312560963,711312465,1351876610,322626781,1910503582,271666773,2175563734,1594956187,70604529,3617834859,1007753275,1495573769,4069517037,2549218298,2663038764,504708206,2263041392,3941167025,2249088522,1514023603,1998579484,1312622330,694541497,2582060303,2151582166,1382467621,776784248,2618340202,3323268794,2497899128,2784771155,503983604,4076293799,907881277,423175695,432175456,1378068232,4145222326,3954048622,3938656102,3820766613,2793130115,2977904593,26017576,3274890735,3194772133,1700274565,1756076034,4006520079,3677328699,720338349,1533947780,354530856,688349552,3973924725,1637815568,332179504,3949051286,53804574,2852348879,3044236432,1282449977,3583942155,3416972820,4006381244,1617046695,2628476075,3002303598,1686838959,431878346,2686675385,1700445008,1080580658,1009431731,832498133,3223435511,2605976345,2271191193,2516031870,1648197032,4164389018,2548247927,300782431,375919233,238389289,3353747414,2531188641,2019080857,1475708069,455242339,2609103871,448939670,3451063019,1395535956,2413381860,1841049896,1491858159,885456874,4264095073,4001119347,1565136089,3898914787,1108368660,540939232,1173283510,2745871338,3681308437,4207628240,3343053890,4016749493,1699691293,1103962373,3625875870,2256883143,3830138730,1031889488,3479347698,1535977030,4236805024,3251091107,2132092099,1774941330,1199868427,1452454533,157007616,2904115357,342012276,595725824,1480756522,206960106,497939518,591360097,863170706,2375253569,3596610801,1814182875,2094937945,3421402208,1082520231,3463918190,2785509508,435703966,3908032597,1641649973,2842273706,3305899714,1510255612,2148256476,2655287854,3276092548,4258621189,236887753,3681803219,274041037,1734335097,3815195456,3317970021,1899903192,1026095262,4050517792,356393447,2410691914,3873677099,3682840055],[3913112168,2491498743,4132185628,2489919796,1091903735,1979897079,3170134830,3567386728,3557303409,857797738,1136121015,1342202287,507115054,2535736646,337727348,3213592640,1301675037,2528481711,1895095763,1721773893,3216771564,62756741,2142006736,835421444,2531993523,1442658625,3659876326,2882144922,676362277,1392781812,170690266,3921047035,1759253602,3611846912,1745797284,664899054,1329594018,3901205900,3045908486,2062866102,2865634940,3543621612,3464012697,1080764994,553557557,3656615353,3996768171,991055499,499776247,1265440854,648242737,3940784050,980351604,3713745714,1749149687,3396870395,4211799374,3640570775,1161844396,3125318951,1431517754,545492359,4268468663,3499529547,1437099964,2702547544,3433638243,2581715763,2787789398,1060185593,1593081372,2418618748,4260947970,69676912,2159744348,86519011,2512459080,3838209314,1220612927,3339683548,133810670,1090789135,1078426020,1569222167,845107691,3583754449,4072456591,1091646820,628848692,1613405280,3757631651,526609435,236106946,48312990,2942717905,3402727701,1797494240,859738849,992217954,4005476642,2243076622,3870952857,3732016268,765654824,3490871365,2511836413,1685915746,3888969200,1414112111,2273134842,3281911079,4080962846,172450625,2569994100,980381355,4109958455,2819808352,2716589560,2568741196,3681446669,3329971472,1835478071,660984891,3704678404,4045999559,3422617507,3040415634,1762651403,1719377915,3470491036,2693910283,3642056355,3138596744,1364962596,2073328063,1983633131,926494387,3423689081,2150032023,4096667949,1749200295,3328846651,309677260,2016342300,1779581495,3079819751,111262694,1274766160,443224088,298511866,1025883608,3806446537,1145181785,168956806,3641502830,3584813610,1689216846,3666258015,3200248200,1692713982,2646376535,4042768518,1618508792,1610833997,3523052358,4130873264,2001055236,3610705100,2202168115,4028541809,2961195399,1006657119,2006996926,3186142756,1430667929,3210227297,1314452623,4074634658,4101304120,2273951170,1399257539,3367210612,3027628629,1190975929,2062231137,2333990788,2221543033,2438960610,1181637006,548689776,2362791313,3372408396,3104550113,3145860560,296247880,1970579870,3078560182,3769228297,1714227617,3291629107,3898220290,166772364,1251581989,493813264,448347421,195405023,2709975567,677966185,3703036547,1463355134,2715995803,1338867538,1343315457,2802222074,2684532164,233230375,2599980071,2000651841,3277868038,1638401717,4028070440,3237316320,6314154,819756386,300326615,590932579,1405279636,3267499572,3150704214,2428286686,3959192993,3461946742,1862657033,1266418056,963775037,2089974820,2263052895,1917689273,448879540,3550394620,3981727096,150775221,3627908307,1303187396,508620638,2975983352,2726630617,1817252668,1876281319,1457606340,908771278,3720792119,3617206836,2455994898,1729034894,1080033504],[976866871,3556439503,2881648439,1522871579,1555064734,1336096578,3548522304,2579274686,3574697629,3205460757,3593280638,3338716283,3079412587,564236357,2993598910,1781952180,1464380207,3163844217,3332601554,1699332808,1393555694,1183702653,3581086237,1288719814,691649499,2847557200,2895455976,3193889540,2717570544,1781354906,1676643554,2592534050,3230253752,1126444790,2770207658,2633158820,2210423226,2615765581,2414155088,3127139286,673620729,2805611233,1269405062,4015350505,3341807571,4149409754,1057255273,2012875353,2162469141,2276492801,2601117357,993977747,3918593370,2654263191,753973209,36408145,2530585658,25011837,3520020182,2088578344,530523599,2918365339,1524020338,1518925132,3760827505,3759777254,1202760957,3985898139,3906192525,674977740,4174734889,2031300136,2019492241,3983892565,4153806404,3822280332,352677332,2297720250,60907813,90501309,3286998549,1016092578,2535922412,2839152426,457141659,509813237,4120667899,652014361,1966332200,2975202805,55981186,2327461051,676427537,3255491064,2882294119,3433927263,1307055953,942726286,933058658,2468411793,3933900994,4215176142,1361170020,2001714738,2830558078,3274259782,1222529897,1679025792,2729314320,3714953764,1770335741,151462246,3013232138,1682292957,1483529935,471910574,1539241949,458788160,3436315007,1807016891,3718408830,978976581,1043663428,3165965781,1927990952,4200891579,2372276910,3208408903,3533431907,1412390302,2931980059,4132332400,1947078029,3881505623,4168226417,2941484381,1077988104,1320477388,886195818,18198404,3786409e3,2509781533,112762804,3463356488,1866414978,891333506,18488651,661792760,1628790961,3885187036,3141171499,876946877,2693282273,1372485963,791857591,2686433993,3759982718,3167212022,3472953795,2716379847,445679433,3561995674,3504004811,3574258232,54117162,3331405415,2381918588,3769707343,4154350007,1140177722,4074052095,668550556,3214352940,367459370,261225585,2610173221,4209349473,3468074219,3265815641,314222801,3066103646,3808782860,282218597,3406013506,3773591054,379116347,1285071038,846784868,2669647154,3771962079,3550491691,2305946142,453669953,1268987020,3317592352,3279303384,3744833421,2610507566,3859509063,266596637,3847019092,517658769,3462560207,3443424879,370717030,4247526661,2224018117,4143653529,4112773975,2788324899,2477274417,1456262402,2901442914,1517677493,1846949527,2295493580,3734397586,2176403920,1280348187,1908823572,3871786941,846861322,1172426758,3287448474,3383383037,1655181056,3139813346,901632758,1897031941,2986607138,3066810236,3447102507,1393639104,373351379,950779232,625454576,3124240540,4148612726,2007998917,544563296,2244738638,2330496472,2058025392,1291430526,424198748,50039436,29584100,3605783033,2429876329,2791104160,1057563949,3255363231,3075367218,3463963227,1469046755,985887462]];var c={pbox:[],sbox:[]};function E(l,C){let d=C>>24&255,g=C>>16&255,B=C>>8&255,F=C&255,A=l.sbox[0][d]+l.sbox[1][g];return A=A^l.sbox[2][B],A=A+l.sbox[3][F],A}function h(l,C,d){let g=C,B=d,F;for(let A=0;A<o;++A)g=g^l.pbox[A],B=E(l,g)^B,F=g,g=B,B=F;return F=g,g=B,B=F,B=B^l.pbox[o],g=g^l.pbox[o+1],{left:g,right:B}}function u(l,C,d){let g=C,B=d,F;for(let A=o+1;A>1;--A)g=g^l.pbox[A],B=E(l,g)^B,F=g,g=B,B=F;return F=g,g=B,B=F,B=B^l.pbox[1],g=g^l.pbox[0],{left:g,right:B}}function v(l,C,d){for(let _=0;_<4;_++){l.sbox[_]=[];for(let b=0;b<256;b++)l.sbox[_][b]=a[_][b]}let g=0;for(let _=0;_<o+2;_++)l.pbox[_]=f[_]^C[g],g++,g>=d&&(g=0);let B=0,F=0,A=0;for(let _=0;_<o+2;_+=2)A=h(l,B,F),B=A.left,F=A.right,l.pbox[_]=B,l.pbox[_+1]=F;for(let _=0;_<4;_++)for(let b=0;b<256;b+=2)A=h(l,B,F),B=A.left,F=A.right,l.sbox[_][b]=B,l.sbox[_][b+1]=F;return!0}var p=x.Blowfish=s.extend({_doReset:function(){if(this._keyPriorReset!==this._key){var l=this._keyPriorReset=this._key,C=l.words,d=l.sigBytes/4;v(c,C,d)}},encryptBlock:function(l,C){var d=h(c,l[C],l[C+1]);l[C]=d.left,l[C+1]=d.right},decryptBlock:function(l,C){var d=u(c,l[C],l[C+1]);l[C]=d.left,l[C+1]=d.right},blockSize:64/32,keySize:128/32,ivSize:64/32});r.Blowfish=s._createHelper(p)}(),t.Blowfish})}(qt)),qt.exports}(function(i,e){(function(t,r,n){i.exports=r(V(),L0(),Er(),Br(),D0(),Ar(),y0(),ne(),ct(),Cr(),xe(),Dr(),yr(),Fr(),gt(),_r(),p0(),t0(),br(),mr(),Sr(),wr(),Tr(),Rr(),kr(),Pr(),Or(),Ir(),Hr(),qr(),Nr(),zr(),Wr(),Lr(),Mr())})(W,function(t){return t})})(Zt);var Ur=Zt.exports;const Nt=dr(Ur);var Vr="0123456789abcdefghijklmnopqrstuvwxyz";function d0(i){return Vr.charAt(i)}function Kr(i,e){return i&e}function M0(i,e){return i|e}function Pe(i,e){return i^e}function Oe(i,e){return i&~e}function $r(i){if(i==0)return-1;var e=0;return i&65535||(i>>=16,e+=16),i&255||(i>>=8,e+=8),i&15||(i>>=4,e+=4),i&3||(i>>=2,e+=2),i&1||++e,e}function jr(i){for(var e=0;i!=0;)i&=i-1,++e;return e}var _0="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",Ie="=";function U0(i){var e,t,r="";for(e=0;e+3<=i.length;e+=3)t=parseInt(i.substring(e,e+3),16),r+=_0.charAt(t>>6)+_0.charAt(t&63);for(e+1==i.length?(t=parseInt(i.substring(e,e+1),16),r+=_0.charAt(t<<2)):e+2==i.length&&(t=parseInt(i.substring(e,e+2),16),r+=_0.charAt(t>>2)+_0.charAt((t&3)<<4));(r.length&3)>0;)r+=Ie;return r}function He(i){var e="",t,r=0,n=0;for(t=0;t<i.length&&i.charAt(t)!=Ie;++t){var s=_0.indexOf(i.charAt(t));s<0||(r==0?(e+=d0(s>>2),n=s&3,r=1):r==1?(e+=d0(n<<2|s>>4),n=s&15,r=2):r==2?(e+=d0(n),e+=d0(s>>2),n=s&3,r=3):(e+=d0(n<<2|s>>4),e+=d0(s&15),r=0))}return r==1&&(e+=d0(n<<2)),e}var b0,Gr={decode:function(i){var e;if(b0===void 0){var t="0123456789ABCDEF",r=` \f
-\r	 \u2028\u2029`;for(b0={},e=0;e<16;++e)b0[t.charAt(e)]=e;for(t=t.toLowerCase(),e=10;e<16;++e)b0[t.charAt(e)]=e;for(e=0;e<r.length;++e)b0[r.charAt(e)]=-1}var n=[],s=0,x=0;for(e=0;e<i.length;++e){var o=i.charAt(e);if(o=="=")break;if(o=b0[o],o!=-1){if(o===void 0)throw new Error("Illegal character at offset "+e);s|=o,++x>=2?(n[n.length]=s,s=0,x=0):s<<=4}}if(x)throw new Error("Hex encoding incomplete: 4 bits missing");return n}},F0,zt={decode:function(i){var e;if(F0===void 0){var t="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",r=`= \f
-\r	 \u2028\u2029`;for(F0=Object.create(null),e=0;e<64;++e)F0[t.charAt(e)]=e;for(F0["-"]=62,F0._=63,e=0;e<r.length;++e)F0[r.charAt(e)]=-1}var n=[],s=0,x=0;for(e=0;e<i.length;++e){var o=i.charAt(e);if(o=="=")break;if(o=F0[o],o!=-1){if(o===void 0)throw new Error("Illegal character at offset "+e);s|=o,++x>=4?(n[n.length]=s>>16,n[n.length]=s>>8&255,n[n.length]=s&255,s=0,x=0):s<<=6}}switch(x){case 1:throw new Error("Base64 encoding incomplete: at least 2 bits missing");case 2:n[n.length]=s>>10;break;case 3:n[n.length]=s>>16,n[n.length]=s>>8&255;break}return n},re:/-----BEGIN [^-]+-----([A-Za-z0-9+\/=\s]+)-----END [^-]+-----|begin-base64[^\n]+\n([A-Za-z0-9+\/=\s]+)====/,unarmor:function(i){var e=zt.re.exec(i);if(e)if(e[1])i=e[1];else if(e[2])i=e[2];else throw new Error("RegExp out of sync");return zt.decode(i)}},m0=1e13,P0=function(){function i(e){this.buf=[+e||0]}return i.prototype.mulAdd=function(e,t){var r=this.buf,n=r.length,s,x;for(s=0;s<n;++s)x=r[s]*e+t,x<m0?t=0:(t=0|x/m0,x-=t*m0),r[s]=x;t>0&&(r[s]=t)},i.prototype.sub=function(e){var t=this.buf,r=t.length,n,s;for(n=0;n<r;++n)s=t[n]-e,s<0?(s+=m0,e=1):e=0,t[n]=s;for(;t[t.length-1]===0;)t.pop()},i.prototype.toString=function(e){if((e||10)!=10)throw new Error("only base 10 is supported");for(var t=this.buf,r=t[t.length-1].toString(),n=t.length-2;n>=0;--n)r+=(m0+t[n]).toString().substring(1);return r},i.prototype.valueOf=function(){for(var e=this.buf,t=0,r=e.length-1;r>=0;--r)t=t*m0+e[r];return t},i.prototype.simplify=function(){var e=this.buf;return e.length==1?e[0]:this},i}(),qe="…",Zr=/^(\d\d)(0[1-9]|1[0-2])(0[1-9]|[12]\d|3[01])([01]\d|2[0-3])(?:([0-5]\d)(?:([0-5]\d)(?:[.,](\d{1,3}))?)?)?(Z|[-+](?:[0]\d|1[0-2])([0-5]\d)?)?$/,Xr=/^(\d\d\d\d)(0[1-9]|1[0-2])(0[1-9]|[12]\d|3[01])([01]\d|2[0-3])(?:([0-5]\d)(?:([0-5]\d)(?:[.,](\d{1,3}))?)?)?(Z|[-+](?:[0]\d|1[0-2])([0-5]\d)?)?$/;function S0(i,e){return i.length>e&&(i=i.substring(0,e)+qe),i}var Wt=function(){function i(e,t){this.hexDigits="0123456789ABCDEF",e instanceof i?(this.enc=e.enc,this.pos=e.pos):(this.enc=e,this.pos=t)}return i.prototype.get=function(e){if(e===void 0&&(e=this.pos++),e>=this.enc.length)throw new Error("Requesting byte offset ".concat(e," on a stream of length ").concat(this.enc.length));return typeof this.enc=="string"?this.enc.charCodeAt(e):this.enc[e]},i.prototype.hexByte=function(e){return this.hexDigits.charAt(e>>4&15)+this.hexDigits.charAt(e&15)},i.prototype.hexDump=function(e,t,r){for(var n="",s=e;s<t;++s)if(n+=this.hexByte(this.get(s)),r!==!0)switch(s&15){case 7:n+="  ";break;case 15:n+=`
-`;break;default:n+=" "}return n},i.prototype.isASCII=function(e,t){for(var r=e;r<t;++r){var n=this.get(r);if(n<32||n>176)return!1}return!0},i.prototype.parseStringISO=function(e,t){for(var r="",n=e;n<t;++n)r+=String.fromCharCode(this.get(n));return r},i.prototype.parseStringUTF=function(e,t){for(var r="",n=e;n<t;){var s=this.get(n++);s<128?r+=String.fromCharCode(s):s>191&&s<224?r+=String.fromCharCode((s&31)<<6|this.get(n++)&63):r+=String.fromCharCode((s&15)<<12|(this.get(n++)&63)<<6|this.get(n++)&63)}return r},i.prototype.parseStringBMP=function(e,t){for(var r="",n,s,x=e;x<t;)n=this.get(x++),s=this.get(x++),r+=String.fromCharCode(n<<8|s);return r},i.prototype.parseTime=function(e,t,r){var n=this.parseStringISO(e,t),s=(r?Zr:Xr).exec(n);return s?(r&&(s[1]=+s[1],s[1]+=+s[1]<70?2e3:1900),n=s[1]+"-"+s[2]+"-"+s[3]+" "+s[4],s[5]&&(n+=":"+s[5],s[6]&&(n+=":"+s[6],s[7]&&(n+="."+s[7]))),s[8]&&(n+=" UTC",s[8]!="Z"&&(n+=s[8],s[9]&&(n+=":"+s[9]))),n):"Unrecognized time: "+n},i.prototype.parseInteger=function(e,t){for(var r=this.get(e),n=r>127,s=n?255:0,x,o="";r==s&&++e<t;)r=this.get(e);if(x=t-e,x===0)return n?-1:0;if(x>4){for(o=r,x<<=3;!((+o^s)&128);)o=+o<<1,--x;o="("+x+` bit)
+ */return t.mode.CTRGladman=function(){var r=t.lib.BlockCipherMode.extend();function n(o){if((o>>24&255)===255){var f=o>>16&255,a=o>>8&255,c=o&255;f===255?(f=0,a===255?(a=0,c===255?c=0:++c):++a):++f,o=0,o+=f<<16,o+=a<<8,o+=c}else o+=1<<24;return o}function s(o){return(o[0]=n(o[0]))===0&&(o[1]=n(o[1])),o}var x=r.Encryptor=r.extend({processBlock:function(o,f){var a=this._cipher,c=a.blockSize,B=this._iv,h=this._counter;B&&(h=this._counter=B.slice(0),this._iv=void 0),s(h);var u=h.slice(0);a.encryptBlock(u,0);for(var v=0;v<c;v++)o[f+v]^=u[v]}});return r.Decryptor=x,r}(),t.mode.CTRGladman})}(Ft)),Ft.exports}var _t={exports:{}},Ce;function Rr(){return Ce||(Ce=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.OFB=function(){var r=t.lib.BlockCipherMode.extend(),n=r.Encryptor=r.extend({processBlock:function(s,x){var o=this._cipher,f=o.blockSize,a=this._iv,c=this._keystream;a&&(c=this._keystream=a.slice(0),this._iv=void 0),o.encryptBlock(c,0);for(var B=0;B<f;B++)s[x+B]^=c[B]}});return r.Decryptor=n,r}(),t.mode.OFB})}(_t)),_t.exports}var mt={exports:{}},De;function Ir(){return De||(De=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.mode.ECB=function(){var r=t.lib.BlockCipherMode.extend();return r.Encryptor=r.extend({processBlock:function(n,s){this._cipher.encryptBlock(n,s)}}),r.Decryptor=r.extend({processBlock:function(n,s){this._cipher.decryptBlock(n,s)}}),r}(),t.mode.ECB})}(mt)),mt.exports}var bt={exports:{}},Ae;function Pr(){return Ae||(Ae=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.AnsiX923={pad:function(r,n){var s=r.sigBytes,x=n*4,o=x-s%x,f=s+o-1;r.clamp(),r.words[f>>>2]|=o<<24-f%4*8,r.sigBytes+=o},unpad:function(r){var n=r.words[r.sigBytes-1>>>2]&255;r.sigBytes-=n}},t.pad.Ansix923})}(bt)),bt.exports}var St={exports:{}},ye;function Or(){return ye||(ye=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.Iso10126={pad:function(r,n){var s=n*4,x=s-r.sigBytes%s;r.concat(t.lib.WordArray.random(x-1)).concat(t.lib.WordArray.create([x<<24],1))},unpad:function(r){var n=r.words[r.sigBytes-1>>>2]&255;r.sigBytes-=n}},t.pad.Iso10126})}(St)),St.exports}var wt={exports:{}},Fe;function Hr(){return Fe||(Fe=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.Iso97971={pad:function(r,n){r.concat(t.lib.WordArray.create([2147483648],1)),t.pad.ZeroPadding.pad(r,n)},unpad:function(r){t.pad.ZeroPadding.unpad(r),r.sigBytes--}},t.pad.Iso97971})}(wt)),wt.exports}var Tt={exports:{}},_e;function qr(){return _e||(_e=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.ZeroPadding={pad:function(r,n){var s=n*4;r.clamp(),r.sigBytes+=s-(r.sigBytes%s||s)},unpad:function(r){for(var n=r.words,s=r.sigBytes-1,s=r.sigBytes-1;s>=0;s--)if(n[s>>>2]>>>24-s%4*8&255){r.sigBytes=s+1;break}}},t.pad.ZeroPadding})}(Tt)),Tt.exports}var kt={exports:{}},me;function Nr(){return me||(me=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return t.pad.NoPadding={pad:function(){},unpad:function(){}},t.pad.NoPadding})}(kt)),kt.exports}var Rt={exports:{}},be;function zr(){return be||(be=1,function(i,e){(function(t,r,n){i.exports=r(V(),t0())})(W,function(t){return function(r){var n=t,s=n.lib,x=s.CipherParams,o=n.enc,f=o.Hex,a=n.format;a.Hex={stringify:function(c){return c.ciphertext.toString(f)},parse:function(c){var B=f.parse(c);return x.create({ciphertext:B})}}}(),t.format.Hex})}(Rt)),Rt.exports}var It={exports:{}},Se;function Wr(){return Se||(Se=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.BlockCipher,x=r.algo,o=[],f=[],a=[],c=[],B=[],h=[],u=[],v=[],g=[],l=[];(function(){for(var p=[],E=0;E<256;E++)E<128?p[E]=E<<1:p[E]=E<<1^283;for(var F=0,C=0,E=0;E<256;E++){var _=C^C<<1^C<<2^C<<3^C<<4;_=_>>>8^_&255^99,o[F]=_,f[_]=F;var m=p[F],I=p[m],D=p[I],y=p[_]*257^_*16843008;a[F]=y<<24|y>>>8,c[F]=y<<16|y>>>16,B[F]=y<<8|y>>>24,h[F]=y;var y=D*16843009^I*65537^m*257^F*16843008;u[_]=y<<24|y>>>8,v[_]=y<<16|y>>>16,g[_]=y<<8|y>>>24,l[_]=y,F?(F=m^p[p[p[D^m]]],C^=p[p[C]]):F=C=1}})();var A=[0,1,2,4,8,16,32,64,128,27,54],d=x.AES=s.extend({_doReset:function(){var p;if(!(this._nRounds&&this._keyPriorReset===this._key)){for(var E=this._keyPriorReset=this._key,F=E.words,C=E.sigBytes/4,_=this._nRounds=C+6,m=(_+1)*4,I=this._keySchedule=[],D=0;D<m;D++)D<C?I[D]=F[D]:(p=I[D-1],D%C?C>6&&D%C==4&&(p=o[p>>>24]<<24|o[p>>>16&255]<<16|o[p>>>8&255]<<8|o[p&255]):(p=p<<8|p>>>24,p=o[p>>>24]<<24|o[p>>>16&255]<<16|o[p>>>8&255]<<8|o[p&255],p^=A[D/C|0]<<24),I[D]=I[D-C]^p);for(var y=this._invKeySchedule=[],T=0;T<m;T++){var D=m-T;if(T%4)var p=I[D];else var p=I[D-4];T<4||D<=4?y[T]=p:y[T]=u[o[p>>>24]]^v[o[p>>>16&255]]^g[o[p>>>8&255]]^l[o[p&255]]}}},encryptBlock:function(p,E){this._doCryptBlock(p,E,this._keySchedule,a,c,B,h,o)},decryptBlock:function(p,E){var F=p[E+1];p[E+1]=p[E+3],p[E+3]=F,this._doCryptBlock(p,E,this._invKeySchedule,u,v,g,l,f);var F=p[E+1];p[E+1]=p[E+3],p[E+3]=F},_doCryptBlock:function(p,E,F,C,_,m,I,D){for(var y=this._nRounds,T=p[E]^F[0],w=p[E+1]^F[1],H=p[E+2]^F[2],N=p[E+3]^F[3],L=4,Z=1;Z<y;Z++){var K=C[T>>>24]^_[w>>>16&255]^m[H>>>8&255]^I[N&255]^F[L++],$=C[w>>>24]^_[H>>>16&255]^m[N>>>8&255]^I[T&255]^F[L++],G=C[H>>>24]^_[N>>>16&255]^m[T>>>8&255]^I[w&255]^F[L++],S=C[N>>>24]^_[T>>>16&255]^m[w>>>8&255]^I[H&255]^F[L++];T=K,w=$,H=G,N=S}var K=(D[T>>>24]<<24|D[w>>>16&255]<<16|D[H>>>8&255]<<8|D[N&255])^F[L++],$=(D[w>>>24]<<24|D[H>>>16&255]<<16|D[N>>>8&255]<<8|D[T&255])^F[L++],G=(D[H>>>24]<<24|D[N>>>16&255]<<16|D[T>>>8&255]<<8|D[w&255])^F[L++],S=(D[N>>>24]<<24|D[T>>>16&255]<<16|D[w>>>8&255]<<8|D[H&255])^F[L++];p[E]=K,p[E+1]=$,p[E+2]=G,p[E+3]=S},keySize:256/32});r.AES=s._createHelper(d)}(),t.AES})}(It)),It.exports}var Pt={exports:{}},we;function Lr(){return we||(we=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.WordArray,x=n.BlockCipher,o=r.algo,f=[57,49,41,33,25,17,9,1,58,50,42,34,26,18,10,2,59,51,43,35,27,19,11,3,60,52,44,36,63,55,47,39,31,23,15,7,62,54,46,38,30,22,14,6,61,53,45,37,29,21,13,5,28,20,12,4],a=[14,17,11,24,1,5,3,28,15,6,21,10,23,19,12,4,26,8,16,7,27,20,13,2,41,52,31,37,47,55,30,40,51,45,33,48,44,49,39,56,34,53,46,42,50,36,29,32],c=[1,2,4,6,8,10,12,14,15,17,19,21,23,25,27,28],B=[{0:8421888,268435456:32768,536870912:8421378,805306368:2,1073741824:512,1342177280:8421890,1610612736:8389122,1879048192:8388608,2147483648:514,2415919104:8389120,2684354560:33280,2952790016:8421376,3221225472:32770,3489660928:8388610,3758096384:0,4026531840:33282,134217728:0,402653184:8421890,671088640:33282,939524096:32768,1207959552:8421888,1476395008:512,1744830464:8421378,2013265920:2,2281701376:8389120,2550136832:33280,2818572288:8421376,3087007744:8389122,3355443200:8388610,3623878656:32770,3892314112:514,4160749568:8388608,1:32768,268435457:2,536870913:8421888,805306369:8388608,1073741825:8421378,1342177281:33280,1610612737:512,1879048193:8389122,2147483649:8421890,2415919105:8421376,2684354561:8388610,2952790017:33282,3221225473:514,3489660929:8389120,3758096385:32770,4026531841:0,134217729:8421890,402653185:8421376,671088641:8388608,939524097:512,1207959553:32768,1476395009:8388610,1744830465:2,2013265921:33282,2281701377:32770,2550136833:8389122,2818572289:514,3087007745:8421888,3355443201:8389120,3623878657:0,3892314113:33280,4160749569:8421378},{0:1074282512,16777216:16384,33554432:524288,50331648:1074266128,67108864:1073741840,83886080:1074282496,100663296:1073758208,117440512:16,134217728:540672,150994944:1073758224,167772160:1073741824,184549376:540688,201326592:524304,218103808:0,234881024:16400,251658240:1074266112,8388608:1073758208,25165824:540688,41943040:16,58720256:1073758224,75497472:1074282512,92274688:1073741824,109051904:524288,125829120:1074266128,142606336:524304,159383552:0,176160768:16384,192937984:1074266112,209715200:1073741840,226492416:540672,243269632:1074282496,260046848:16400,268435456:0,285212672:1074266128,301989888:1073758224,318767104:1074282496,335544320:1074266112,352321536:16,369098752:540688,385875968:16384,402653184:16400,419430400:524288,436207616:524304,452984832:1073741840,469762048:540672,486539264:1073758208,503316480:1073741824,520093696:1074282512,276824064:540688,293601280:524288,310378496:1074266112,327155712:16384,343932928:1073758208,360710144:1074282512,377487360:16,394264576:1073741824,411041792:1074282496,427819008:1073741840,444596224:1073758224,461373440:524304,478150656:0,494927872:16400,511705088:1074266128,528482304:540672},{0:260,1048576:0,2097152:67109120,3145728:65796,4194304:65540,5242880:67108868,6291456:67174660,7340032:67174400,8388608:67108864,9437184:67174656,10485760:65792,11534336:67174404,12582912:67109124,13631488:65536,14680064:4,15728640:256,524288:67174656,1572864:67174404,2621440:0,3670016:67109120,4718592:67108868,5767168:65536,6815744:65540,7864320:260,8912896:4,9961472:256,11010048:67174400,12058624:65796,13107200:65792,14155776:67109124,15204352:67174660,16252928:67108864,16777216:67174656,17825792:65540,18874368:65536,19922944:67109120,20971520:256,22020096:67174660,23068672:67108868,24117248:0,25165824:67109124,26214400:67108864,27262976:4,28311552:65792,29360128:67174400,30408704:260,31457280:65796,32505856:67174404,17301504:67108864,18350080:260,19398656:67174656,20447232:0,21495808:65540,22544384:67109120,23592960:256,24641536:67174404,25690112:65536,26738688:67174660,27787264:65796,28835840:67108868,29884416:67109124,30932992:67174400,31981568:4,33030144:65792},{0:2151682048,65536:2147487808,131072:4198464,196608:2151677952,262144:0,327680:4198400,393216:2147483712,458752:4194368,524288:2147483648,589824:4194304,655360:64,720896:2147487744,786432:2151678016,851968:4160,917504:4096,983040:2151682112,32768:2147487808,98304:64,163840:2151678016,229376:2147487744,294912:4198400,360448:2151682112,425984:0,491520:2151677952,557056:4096,622592:2151682048,688128:4194304,753664:4160,819200:2147483648,884736:4194368,950272:4198464,1015808:2147483712,1048576:4194368,1114112:4198400,1179648:2147483712,1245184:0,1310720:4160,1376256:2151678016,1441792:2151682048,1507328:2147487808,1572864:2151682112,1638400:2147483648,1703936:2151677952,1769472:4198464,1835008:2147487744,1900544:4194304,1966080:64,2031616:4096,1081344:2151677952,1146880:2151682112,1212416:0,1277952:4198400,1343488:4194368,1409024:2147483648,1474560:2147487808,1540096:64,1605632:2147483712,1671168:4096,1736704:2147487744,1802240:2151678016,1867776:4160,1933312:2151682048,1998848:4194304,2064384:4198464},{0:128,4096:17039360,8192:262144,12288:536870912,16384:537133184,20480:16777344,24576:553648256,28672:262272,32768:16777216,36864:537133056,40960:536871040,45056:553910400,49152:553910272,53248:0,57344:17039488,61440:553648128,2048:17039488,6144:553648256,10240:128,14336:17039360,18432:262144,22528:537133184,26624:553910272,30720:536870912,34816:537133056,38912:0,43008:553910400,47104:16777344,51200:536871040,55296:553648128,59392:16777216,63488:262272,65536:262144,69632:128,73728:536870912,77824:553648256,81920:16777344,86016:553910272,90112:537133184,94208:16777216,98304:553910400,102400:553648128,106496:17039360,110592:537133056,114688:262272,118784:536871040,122880:0,126976:17039488,67584:553648256,71680:16777216,75776:17039360,79872:537133184,83968:536870912,88064:17039488,92160:128,96256:553910272,100352:262272,104448:553910400,108544:0,112640:553648128,116736:16777344,120832:262144,124928:537133056,129024:536871040},{0:268435464,256:8192,512:270532608,768:270540808,1024:268443648,1280:2097152,1536:2097160,1792:268435456,2048:0,2304:268443656,2560:2105344,2816:8,3072:270532616,3328:2105352,3584:8200,3840:270540800,128:270532608,384:270540808,640:8,896:2097152,1152:2105352,1408:268435464,1664:268443648,1920:8200,2176:2097160,2432:8192,2688:268443656,2944:270532616,3200:0,3456:270540800,3712:2105344,3968:268435456,4096:268443648,4352:270532616,4608:270540808,4864:8200,5120:2097152,5376:268435456,5632:268435464,5888:2105344,6144:2105352,6400:0,6656:8,6912:270532608,7168:8192,7424:268443656,7680:270540800,7936:2097160,4224:8,4480:2105344,4736:2097152,4992:268435464,5248:268443648,5504:8200,5760:270540808,6016:270532608,6272:270540800,6528:270532616,6784:8192,7040:2105352,7296:2097160,7552:0,7808:268435456,8064:268443656},{0:1048576,16:33555457,32:1024,48:1049601,64:34604033,80:0,96:1,112:34603009,128:33555456,144:1048577,160:33554433,176:34604032,192:34603008,208:1025,224:1049600,240:33554432,8:34603009,24:0,40:33555457,56:34604032,72:1048576,88:33554433,104:33554432,120:1025,136:1049601,152:33555456,168:34603008,184:1048577,200:1024,216:34604033,232:1,248:1049600,256:33554432,272:1048576,288:33555457,304:34603009,320:1048577,336:33555456,352:34604032,368:1049601,384:1025,400:34604033,416:1049600,432:1,448:0,464:34603008,480:33554433,496:1024,264:1049600,280:33555457,296:34603009,312:1,328:33554432,344:1048576,360:1025,376:34604032,392:33554433,408:34603008,424:0,440:34604033,456:1049601,472:1024,488:33555456,504:1048577},{0:134219808,1:131072,2:134217728,3:32,4:131104,5:134350880,6:134350848,7:2048,8:134348800,9:134219776,10:133120,11:134348832,12:2080,13:0,14:134217760,15:133152,2147483648:2048,2147483649:134350880,2147483650:134219808,2147483651:134217728,2147483652:134348800,2147483653:133120,2147483654:133152,2147483655:32,2147483656:134217760,2147483657:2080,2147483658:131104,2147483659:134350848,2147483660:0,2147483661:134348832,2147483662:134219776,2147483663:131072,16:133152,17:134350848,18:32,19:2048,20:134219776,21:134217760,22:134348832,23:131072,24:0,25:131104,26:134348800,27:134219808,28:134350880,29:133120,30:2080,31:134217728,2147483664:131072,2147483665:2048,2147483666:134348832,2147483667:133152,2147483668:32,2147483669:134348800,2147483670:134217728,2147483671:134219808,2147483672:134350880,2147483673:134217760,2147483674:134219776,2147483675:0,2147483676:133120,2147483677:2080,2147483678:131104,2147483679:134350848}],h=[4160749569,528482304,33030144,2064384,129024,8064,504,2147483679],u=o.DES=x.extend({_doReset:function(){for(var A=this._key,d=A.words,p=[],E=0;E<56;E++){var F=f[E]-1;p[E]=d[F>>>5]>>>31-F%32&1}for(var C=this._subKeys=[],_=0;_<16;_++){for(var m=C[_]=[],I=c[_],E=0;E<24;E++)m[E/6|0]|=p[(a[E]-1+I)%28]<<31-E%6,m[4+(E/6|0)]|=p[28+(a[E+24]-1+I)%28]<<31-E%6;m[0]=m[0]<<1|m[0]>>>31;for(var E=1;E<7;E++)m[E]=m[E]>>>(E-1)*4+3;m[7]=m[7]<<5|m[7]>>>27}for(var D=this._invSubKeys=[],E=0;E<16;E++)D[E]=C[15-E]},encryptBlock:function(A,d){this._doCryptBlock(A,d,this._subKeys)},decryptBlock:function(A,d){this._doCryptBlock(A,d,this._invSubKeys)},_doCryptBlock:function(A,d,p){this._lBlock=A[d],this._rBlock=A[d+1],v.call(this,4,252645135),v.call(this,16,65535),g.call(this,2,858993459),g.call(this,8,16711935),v.call(this,1,1431655765);for(var E=0;E<16;E++){for(var F=p[E],C=this._lBlock,_=this._rBlock,m=0,I=0;I<8;I++)m|=B[I][((_^F[I])&h[I])>>>0];this._lBlock=_,this._rBlock=C^m}var D=this._lBlock;this._lBlock=this._rBlock,this._rBlock=D,v.call(this,1,1431655765),g.call(this,8,16711935),g.call(this,2,858993459),v.call(this,16,65535),v.call(this,4,252645135),A[d]=this._lBlock,A[d+1]=this._rBlock},keySize:64/32,ivSize:64/32,blockSize:64/32});function v(A,d){var p=(this._lBlock>>>A^this._rBlock)&d;this._rBlock^=p,this._lBlock^=p<<A}function g(A,d){var p=(this._rBlock>>>A^this._lBlock)&d;this._lBlock^=p,this._rBlock^=p<<A}r.DES=x._createHelper(u);var l=o.TripleDES=x.extend({_doReset:function(){var A=this._key,d=A.words;if(d.length!==2&&d.length!==4&&d.length<6)throw new Error("Invalid key length - 3DES requires the key length to be 64, 128, 192 or >192.");var p=d.slice(0,2),E=d.length<4?d.slice(0,2):d.slice(2,4),F=d.length<6?d.slice(0,2):d.slice(4,6);this._des1=u.createEncryptor(s.create(p)),this._des2=u.createEncryptor(s.create(E)),this._des3=u.createEncryptor(s.create(F))},encryptBlock:function(A,d){this._des1.encryptBlock(A,d),this._des2.decryptBlock(A,d),this._des3.encryptBlock(A,d)},decryptBlock:function(A,d){this._des3.decryptBlock(A,d),this._des2.encryptBlock(A,d),this._des1.decryptBlock(A,d)},keySize:192/32,ivSize:64/32,blockSize:64/32});r.TripleDES=x._createHelper(l)}(),t.TripleDES})}(Pt)),Pt.exports}var Ot={exports:{}},Te;function Mr(){return Te||(Te=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=x.RC4=s.extend({_doReset:function(){for(var c=this._key,B=c.words,h=c.sigBytes,u=this._S=[],v=0;v<256;v++)u[v]=v;for(var v=0,g=0;v<256;v++){var l=v%h,A=B[l>>>2]>>>24-l%4*8&255;g=(g+u[v]+A)%256;var d=u[v];u[v]=u[g],u[g]=d}this._i=this._j=0},_doProcessBlock:function(c,B){c[B]^=f.call(this)},keySize:256/32,ivSize:0});function f(){for(var c=this._S,B=this._i,h=this._j,u=0,v=0;v<4;v++){B=(B+1)%256,h=(h+c[B])%256;var g=c[B];c[B]=c[h],c[h]=g,u|=c[(c[B]+c[h])%256]<<24-v*8}return this._i=B,this._j=h,u}r.RC4=s._createHelper(o);var a=x.RC4Drop=o.extend({cfg:o.cfg.extend({drop:192}),_doReset:function(){o._doReset.call(this);for(var c=this.cfg.drop;c>0;c--)f.call(this)}});r.RC4Drop=s._createHelper(a)}(),t.RC4})}(Ot)),Ot.exports}var Ht={exports:{}},ke;function Ur(){return ke||(ke=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=[],f=[],a=[],c=x.Rabbit=s.extend({_doReset:function(){for(var h=this._key.words,u=this.cfg.iv,v=0;v<4;v++)h[v]=(h[v]<<8|h[v]>>>24)&16711935|(h[v]<<24|h[v]>>>8)&4278255360;var g=this._X=[h[0],h[3]<<16|h[2]>>>16,h[1],h[0]<<16|h[3]>>>16,h[2],h[1]<<16|h[0]>>>16,h[3],h[2]<<16|h[1]>>>16],l=this._C=[h[2]<<16|h[2]>>>16,h[0]&4294901760|h[1]&65535,h[3]<<16|h[3]>>>16,h[1]&4294901760|h[2]&65535,h[0]<<16|h[0]>>>16,h[2]&4294901760|h[3]&65535,h[1]<<16|h[1]>>>16,h[3]&4294901760|h[0]&65535];this._b=0;for(var v=0;v<4;v++)B.call(this);for(var v=0;v<8;v++)l[v]^=g[v+4&7];if(u){var A=u.words,d=A[0],p=A[1],E=(d<<8|d>>>24)&16711935|(d<<24|d>>>8)&4278255360,F=(p<<8|p>>>24)&16711935|(p<<24|p>>>8)&4278255360,C=E>>>16|F&4294901760,_=F<<16|E&65535;l[0]^=E,l[1]^=C,l[2]^=F,l[3]^=_,l[4]^=E,l[5]^=C,l[6]^=F,l[7]^=_;for(var v=0;v<4;v++)B.call(this)}},_doProcessBlock:function(h,u){var v=this._X;B.call(this),o[0]=v[0]^v[5]>>>16^v[3]<<16,o[1]=v[2]^v[7]>>>16^v[5]<<16,o[2]=v[4]^v[1]>>>16^v[7]<<16,o[3]=v[6]^v[3]>>>16^v[1]<<16;for(var g=0;g<4;g++)o[g]=(o[g]<<8|o[g]>>>24)&16711935|(o[g]<<24|o[g]>>>8)&4278255360,h[u+g]^=o[g]},blockSize:128/32,ivSize:64/32});function B(){for(var h=this._X,u=this._C,v=0;v<8;v++)f[v]=u[v];u[0]=u[0]+1295307597+this._b|0,u[1]=u[1]+3545052371+(u[0]>>>0<f[0]>>>0?1:0)|0,u[2]=u[2]+886263092+(u[1]>>>0<f[1]>>>0?1:0)|0,u[3]=u[3]+1295307597+(u[2]>>>0<f[2]>>>0?1:0)|0,u[4]=u[4]+3545052371+(u[3]>>>0<f[3]>>>0?1:0)|0,u[5]=u[5]+886263092+(u[4]>>>0<f[4]>>>0?1:0)|0,u[6]=u[6]+1295307597+(u[5]>>>0<f[5]>>>0?1:0)|0,u[7]=u[7]+3545052371+(u[6]>>>0<f[6]>>>0?1:0)|0,this._b=u[7]>>>0<f[7]>>>0?1:0;for(var v=0;v<8;v++){var g=h[v]+u[v],l=g&65535,A=g>>>16,d=((l*l>>>17)+l*A>>>15)+A*A,p=((g&4294901760)*g|0)+((g&65535)*g|0);a[v]=d^p}h[0]=a[0]+(a[7]<<16|a[7]>>>16)+(a[6]<<16|a[6]>>>16)|0,h[1]=a[1]+(a[0]<<8|a[0]>>>24)+a[7]|0,h[2]=a[2]+(a[1]<<16|a[1]>>>16)+(a[0]<<16|a[0]>>>16)|0,h[3]=a[3]+(a[2]<<8|a[2]>>>24)+a[1]|0,h[4]=a[4]+(a[3]<<16|a[3]>>>16)+(a[2]<<16|a[2]>>>16)|0,h[5]=a[5]+(a[4]<<8|a[4]>>>24)+a[3]|0,h[6]=a[6]+(a[5]<<16|a[5]>>>16)+(a[4]<<16|a[4]>>>16)|0,h[7]=a[7]+(a[6]<<8|a[6]>>>24)+a[5]|0}r.Rabbit=s._createHelper(c)}(),t.Rabbit})}(Ht)),Ht.exports}var qt={exports:{}},Re;function Vr(){return Re||(Re=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.StreamCipher,x=r.algo,o=[],f=[],a=[],c=x.RabbitLegacy=s.extend({_doReset:function(){var h=this._key.words,u=this.cfg.iv,v=this._X=[h[0],h[3]<<16|h[2]>>>16,h[1],h[0]<<16|h[3]>>>16,h[2],h[1]<<16|h[0]>>>16,h[3],h[2]<<16|h[1]>>>16],g=this._C=[h[2]<<16|h[2]>>>16,h[0]&4294901760|h[1]&65535,h[3]<<16|h[3]>>>16,h[1]&4294901760|h[2]&65535,h[0]<<16|h[0]>>>16,h[2]&4294901760|h[3]&65535,h[1]<<16|h[1]>>>16,h[3]&4294901760|h[0]&65535];this._b=0;for(var l=0;l<4;l++)B.call(this);for(var l=0;l<8;l++)g[l]^=v[l+4&7];if(u){var A=u.words,d=A[0],p=A[1],E=(d<<8|d>>>24)&16711935|(d<<24|d>>>8)&4278255360,F=(p<<8|p>>>24)&16711935|(p<<24|p>>>8)&4278255360,C=E>>>16|F&4294901760,_=F<<16|E&65535;g[0]^=E,g[1]^=C,g[2]^=F,g[3]^=_,g[4]^=E,g[5]^=C,g[6]^=F,g[7]^=_;for(var l=0;l<4;l++)B.call(this)}},_doProcessBlock:function(h,u){var v=this._X;B.call(this),o[0]=v[0]^v[5]>>>16^v[3]<<16,o[1]=v[2]^v[7]>>>16^v[5]<<16,o[2]=v[4]^v[1]>>>16^v[7]<<16,o[3]=v[6]^v[3]>>>16^v[1]<<16;for(var g=0;g<4;g++)o[g]=(o[g]<<8|o[g]>>>24)&16711935|(o[g]<<24|o[g]>>>8)&4278255360,h[u+g]^=o[g]},blockSize:128/32,ivSize:64/32});function B(){for(var h=this._X,u=this._C,v=0;v<8;v++)f[v]=u[v];u[0]=u[0]+1295307597+this._b|0,u[1]=u[1]+3545052371+(u[0]>>>0<f[0]>>>0?1:0)|0,u[2]=u[2]+886263092+(u[1]>>>0<f[1]>>>0?1:0)|0,u[3]=u[3]+1295307597+(u[2]>>>0<f[2]>>>0?1:0)|0,u[4]=u[4]+3545052371+(u[3]>>>0<f[3]>>>0?1:0)|0,u[5]=u[5]+886263092+(u[4]>>>0<f[4]>>>0?1:0)|0,u[6]=u[6]+1295307597+(u[5]>>>0<f[5]>>>0?1:0)|0,u[7]=u[7]+3545052371+(u[6]>>>0<f[6]>>>0?1:0)|0,this._b=u[7]>>>0<f[7]>>>0?1:0;for(var v=0;v<8;v++){var g=h[v]+u[v],l=g&65535,A=g>>>16,d=((l*l>>>17)+l*A>>>15)+A*A,p=((g&4294901760)*g|0)+((g&65535)*g|0);a[v]=d^p}h[0]=a[0]+(a[7]<<16|a[7]>>>16)+(a[6]<<16|a[6]>>>16)|0,h[1]=a[1]+(a[0]<<8|a[0]>>>24)+a[7]|0,h[2]=a[2]+(a[1]<<16|a[1]>>>16)+(a[0]<<16|a[0]>>>16)|0,h[3]=a[3]+(a[2]<<8|a[2]>>>24)+a[1]|0,h[4]=a[4]+(a[3]<<16|a[3]>>>16)+(a[2]<<16|a[2]>>>16)|0,h[5]=a[5]+(a[4]<<8|a[4]>>>24)+a[3]|0,h[6]=a[6]+(a[5]<<16|a[5]>>>16)+(a[4]<<16|a[4]>>>16)|0,h[7]=a[7]+(a[6]<<8|a[6]>>>24)+a[5]|0}r.RabbitLegacy=s._createHelper(c)}(),t.RabbitLegacy})}(qt)),qt.exports}var Nt={exports:{}},Ie;function Kr(){return Ie||(Ie=1,function(i,e){(function(t,r,n){i.exports=r(V(),y0(),F0(),g0(),t0())})(W,function(t){return function(){var r=t,n=r.lib,s=n.BlockCipher,x=r.algo;const o=16,f=[608135816,2242054355,320440878,57701188,2752067618,698298832,137296536,3964562569,1160258022,953160567,3193202383,887688300,3232508343,3380367581,1065670069,3041331479,2450970073,2306472731],a=[[3509652390,2564797868,805139163,3491422135,3101798381,1780907670,3128725573,4046225305,614570311,3012652279,134345442,2240740374,1667834072,1901547113,2757295779,4103290238,227898511,1921955416,1904987480,2182433518,2069144605,3260701109,2620446009,720527379,3318853667,677414384,3393288472,3101374703,2390351024,1614419982,1822297739,2954791486,3608508353,3174124327,2024746970,1432378464,3864339955,2857741204,1464375394,1676153920,1439316330,715854006,3033291828,289532110,2706671279,2087905683,3018724369,1668267050,732546397,1947742710,3462151702,2609353502,2950085171,1814351708,2050118529,680887927,999245976,1800124847,3300911131,1713906067,1641548236,4213287313,1216130144,1575780402,4018429277,3917837745,3693486850,3949271944,596196993,3549867205,258830323,2213823033,772490370,2760122372,1774776394,2652871518,566650946,4142492826,1728879713,2882767088,1783734482,3629395816,2517608232,2874225571,1861159788,326777828,3124490320,2130389656,2716951837,967770486,1724537150,2185432712,2364442137,1164943284,2105845187,998989502,3765401048,2244026483,1075463327,1455516326,1322494562,910128902,469688178,1117454909,936433444,3490320968,3675253459,1240580251,122909385,2157517691,634681816,4142456567,3825094682,3061402683,2540495037,79693498,3249098678,1084186820,1583128258,426386531,1761308591,1047286709,322548459,995290223,1845252383,2603652396,3431023940,2942221577,3202600964,3727903485,1712269319,422464435,3234572375,1170764815,3523960633,3117677531,1434042557,442511882,3600875718,1076654713,1738483198,4213154764,2393238008,3677496056,1014306527,4251020053,793779912,2902807211,842905082,4246964064,1395751752,1040244610,2656851899,3396308128,445077038,3742853595,3577915638,679411651,2892444358,2354009459,1767581616,3150600392,3791627101,3102740896,284835224,4246832056,1258075500,768725851,2589189241,3069724005,3532540348,1274779536,3789419226,2764799539,1660621633,3471099624,4011903706,913787905,3497959166,737222580,2514213453,2928710040,3937242737,1804850592,3499020752,2949064160,2386320175,2390070455,2415321851,4061277028,2290661394,2416832540,1336762016,1754252060,3520065937,3014181293,791618072,3188594551,3933548030,2332172193,3852520463,3043980520,413987798,3465142937,3030929376,4245938359,2093235073,3534596313,375366246,2157278981,2479649556,555357303,3870105701,2008414854,3344188149,4221384143,3956125452,2067696032,3594591187,2921233993,2428461,544322398,577241275,1471733935,610547355,4027169054,1432588573,1507829418,2025931657,3646575487,545086370,48609733,2200306550,1653985193,298326376,1316178497,3007786442,2064951626,458293330,2589141269,3591329599,3164325604,727753846,2179363840,146436021,1461446943,4069977195,705550613,3059967265,3887724982,4281599278,3313849956,1404054877,2845806497,146425753,1854211946],[1266315497,3048417604,3681880366,3289982499,290971e4,1235738493,2632868024,2414719590,3970600049,1771706367,1449415276,3266420449,422970021,1963543593,2690192192,3826793022,1062508698,1531092325,1804592342,2583117782,2714934279,4024971509,1294809318,4028980673,1289560198,2221992742,1669523910,35572830,157838143,1052438473,1016535060,1802137761,1753167236,1386275462,3080475397,2857371447,1040679964,2145300060,2390574316,1461121720,2956646967,4031777805,4028374788,33600511,2920084762,1018524850,629373528,3691585981,3515945977,2091462646,2486323059,586499841,988145025,935516892,3367335476,2599673255,2839830854,265290510,3972581182,2759138881,3795373465,1005194799,847297441,406762289,1314163512,1332590856,1866599683,4127851711,750260880,613907577,1450815602,3165620655,3734664991,3650291728,3012275730,3704569646,1427272223,778793252,1343938022,2676280711,2052605720,1946737175,3164576444,3914038668,3967478842,3682934266,1661551462,3294938066,4011595847,840292616,3712170807,616741398,312560963,711312465,1351876610,322626781,1910503582,271666773,2175563734,1594956187,70604529,3617834859,1007753275,1495573769,4069517037,2549218298,2663038764,504708206,2263041392,3941167025,2249088522,1514023603,1998579484,1312622330,694541497,2582060303,2151582166,1382467621,776784248,2618340202,3323268794,2497899128,2784771155,503983604,4076293799,907881277,423175695,432175456,1378068232,4145222326,3954048622,3938656102,3820766613,2793130115,2977904593,26017576,3274890735,3194772133,1700274565,1756076034,4006520079,3677328699,720338349,1533947780,354530856,688349552,3973924725,1637815568,332179504,3949051286,53804574,2852348879,3044236432,1282449977,3583942155,3416972820,4006381244,1617046695,2628476075,3002303598,1686838959,431878346,2686675385,1700445008,1080580658,1009431731,832498133,3223435511,2605976345,2271191193,2516031870,1648197032,4164389018,2548247927,300782431,375919233,238389289,3353747414,2531188641,2019080857,1475708069,455242339,2609103871,448939670,3451063019,1395535956,2413381860,1841049896,1491858159,885456874,4264095073,4001119347,1565136089,3898914787,1108368660,540939232,1173283510,2745871338,3681308437,4207628240,3343053890,4016749493,1699691293,1103962373,3625875870,2256883143,3830138730,1031889488,3479347698,1535977030,4236805024,3251091107,2132092099,1774941330,1199868427,1452454533,157007616,2904115357,342012276,595725824,1480756522,206960106,497939518,591360097,863170706,2375253569,3596610801,1814182875,2094937945,3421402208,1082520231,3463918190,2785509508,435703966,3908032597,1641649973,2842273706,3305899714,1510255612,2148256476,2655287854,3276092548,4258621189,236887753,3681803219,274041037,1734335097,3815195456,3317970021,1899903192,1026095262,4050517792,356393447,2410691914,3873677099,3682840055],[3913112168,2491498743,4132185628,2489919796,1091903735,1979897079,3170134830,3567386728,3557303409,857797738,1136121015,1342202287,507115054,2535736646,337727348,3213592640,1301675037,2528481711,1895095763,1721773893,3216771564,62756741,2142006736,835421444,2531993523,1442658625,3659876326,2882144922,676362277,1392781812,170690266,3921047035,1759253602,3611846912,1745797284,664899054,1329594018,3901205900,3045908486,2062866102,2865634940,3543621612,3464012697,1080764994,553557557,3656615353,3996768171,991055499,499776247,1265440854,648242737,3940784050,980351604,3713745714,1749149687,3396870395,4211799374,3640570775,1161844396,3125318951,1431517754,545492359,4268468663,3499529547,1437099964,2702547544,3433638243,2581715763,2787789398,1060185593,1593081372,2418618748,4260947970,69676912,2159744348,86519011,2512459080,3838209314,1220612927,3339683548,133810670,1090789135,1078426020,1569222167,845107691,3583754449,4072456591,1091646820,628848692,1613405280,3757631651,526609435,236106946,48312990,2942717905,3402727701,1797494240,859738849,992217954,4005476642,2243076622,3870952857,3732016268,765654824,3490871365,2511836413,1685915746,3888969200,1414112111,2273134842,3281911079,4080962846,172450625,2569994100,980381355,4109958455,2819808352,2716589560,2568741196,3681446669,3329971472,1835478071,660984891,3704678404,4045999559,3422617507,3040415634,1762651403,1719377915,3470491036,2693910283,3642056355,3138596744,1364962596,2073328063,1983633131,926494387,3423689081,2150032023,4096667949,1749200295,3328846651,309677260,2016342300,1779581495,3079819751,111262694,1274766160,443224088,298511866,1025883608,3806446537,1145181785,168956806,3641502830,3584813610,1689216846,3666258015,3200248200,1692713982,2646376535,4042768518,1618508792,1610833997,3523052358,4130873264,2001055236,3610705100,2202168115,4028541809,2961195399,1006657119,2006996926,3186142756,1430667929,3210227297,1314452623,4074634658,4101304120,2273951170,1399257539,3367210612,3027628629,1190975929,2062231137,2333990788,2221543033,2438960610,1181637006,548689776,2362791313,3372408396,3104550113,3145860560,296247880,1970579870,3078560182,3769228297,1714227617,3291629107,3898220290,166772364,1251581989,493813264,448347421,195405023,2709975567,677966185,3703036547,1463355134,2715995803,1338867538,1343315457,2802222074,2684532164,233230375,2599980071,2000651841,3277868038,1638401717,4028070440,3237316320,6314154,819756386,300326615,590932579,1405279636,3267499572,3150704214,2428286686,3959192993,3461946742,1862657033,1266418056,963775037,2089974820,2263052895,1917689273,448879540,3550394620,3981727096,150775221,3627908307,1303187396,508620638,2975983352,2726630617,1817252668,1876281319,1457606340,908771278,3720792119,3617206836,2455994898,1729034894,1080033504],[976866871,3556439503,2881648439,1522871579,1555064734,1336096578,3548522304,2579274686,3574697629,3205460757,3593280638,3338716283,3079412587,564236357,2993598910,1781952180,1464380207,3163844217,3332601554,1699332808,1393555694,1183702653,3581086237,1288719814,691649499,2847557200,2895455976,3193889540,2717570544,1781354906,1676643554,2592534050,3230253752,1126444790,2770207658,2633158820,2210423226,2615765581,2414155088,3127139286,673620729,2805611233,1269405062,4015350505,3341807571,4149409754,1057255273,2012875353,2162469141,2276492801,2601117357,993977747,3918593370,2654263191,753973209,36408145,2530585658,25011837,3520020182,2088578344,530523599,2918365339,1524020338,1518925132,3760827505,3759777254,1202760957,3985898139,3906192525,674977740,4174734889,2031300136,2019492241,3983892565,4153806404,3822280332,352677332,2297720250,60907813,90501309,3286998549,1016092578,2535922412,2839152426,457141659,509813237,4120667899,652014361,1966332200,2975202805,55981186,2327461051,676427537,3255491064,2882294119,3433927263,1307055953,942726286,933058658,2468411793,3933900994,4215176142,1361170020,2001714738,2830558078,3274259782,1222529897,1679025792,2729314320,3714953764,1770335741,151462246,3013232138,1682292957,1483529935,471910574,1539241949,458788160,3436315007,1807016891,3718408830,978976581,1043663428,3165965781,1927990952,4200891579,2372276910,3208408903,3533431907,1412390302,2931980059,4132332400,1947078029,3881505623,4168226417,2941484381,1077988104,1320477388,886195818,18198404,3786409e3,2509781533,112762804,3463356488,1866414978,891333506,18488651,661792760,1628790961,3885187036,3141171499,876946877,2693282273,1372485963,791857591,2686433993,3759982718,3167212022,3472953795,2716379847,445679433,3561995674,3504004811,3574258232,54117162,3331405415,2381918588,3769707343,4154350007,1140177722,4074052095,668550556,3214352940,367459370,261225585,2610173221,4209349473,3468074219,3265815641,314222801,3066103646,3808782860,282218597,3406013506,3773591054,379116347,1285071038,846784868,2669647154,3771962079,3550491691,2305946142,453669953,1268987020,3317592352,3279303384,3744833421,2610507566,3859509063,266596637,3847019092,517658769,3462560207,3443424879,370717030,4247526661,2224018117,4143653529,4112773975,2788324899,2477274417,1456262402,2901442914,1517677493,1846949527,2295493580,3734397586,2176403920,1280348187,1908823572,3871786941,846861322,1172426758,3287448474,3383383037,1655181056,3139813346,901632758,1897031941,2986607138,3066810236,3447102507,1393639104,373351379,950779232,625454576,3124240540,4148612726,2007998917,544563296,2244738638,2330496472,2058025392,1291430526,424198748,50039436,29584100,3605783033,2429876329,2791104160,1057563949,3255363231,3075367218,3463963227,1469046755,985887462]];var c={pbox:[],sbox:[]};function B(l,A){let d=A>>24&255,p=A>>16&255,E=A>>8&255,F=A&255,C=l.sbox[0][d]+l.sbox[1][p];return C=C^l.sbox[2][E],C=C+l.sbox[3][F],C}function h(l,A,d){let p=A,E=d,F;for(let C=0;C<o;++C)p=p^l.pbox[C],E=B(l,p)^E,F=p,p=E,E=F;return F=p,p=E,E=F,E=E^l.pbox[o],p=p^l.pbox[o+1],{left:p,right:E}}function u(l,A,d){let p=A,E=d,F;for(let C=o+1;C>1;--C)p=p^l.pbox[C],E=B(l,p)^E,F=p,p=E,E=F;return F=p,p=E,E=F,E=E^l.pbox[1],p=p^l.pbox[0],{left:p,right:E}}function v(l,A,d){for(let _=0;_<4;_++){l.sbox[_]=[];for(let m=0;m<256;m++)l.sbox[_][m]=a[_][m]}let p=0;for(let _=0;_<o+2;_++)l.pbox[_]=f[_]^A[p],p++,p>=d&&(p=0);let E=0,F=0,C=0;for(let _=0;_<o+2;_+=2)C=h(l,E,F),E=C.left,F=C.right,l.pbox[_]=E,l.pbox[_+1]=F;for(let _=0;_<4;_++)for(let m=0;m<256;m+=2)C=h(l,E,F),E=C.left,F=C.right,l.sbox[_][m]=E,l.sbox[_][m+1]=F;return!0}var g=x.Blowfish=s.extend({_doReset:function(){if(this._keyPriorReset!==this._key){var l=this._keyPriorReset=this._key,A=l.words,d=l.sigBytes/4;v(c,A,d)}},encryptBlock:function(l,A){var d=h(c,l[A],l[A+1]);l[A]=d.left,l[A+1]=d.right},decryptBlock:function(l,A){var d=u(c,l[A],l[A+1]);l[A]=d.left,l[A+1]=d.right},blockSize:64/32,keySize:128/32,ivSize:64/32});r.Blowfish=s._createHelper(g)}(),t.Blowfish})}(Nt)),Nt.exports}(function(i,e){(function(t,r,n){i.exports=r(V(),M0(),Dr(),Ar(),y0(),yr(),F0(),se(),ht(),Fr(),fe(),_r(),mr(),br(),Et(),Sr(),g0(),t0(),wr(),Tr(),kr(),Rr(),Ir(),Pr(),Or(),Hr(),qr(),Nr(),zr(),Wr(),Lr(),Mr(),Ur(),Vr(),Kr())})(W,function(t){return t})})(Xt);var Gr=Xt.exports;const zt=gr(Gr);var $r="0123456789abcdefghijklmnopqrstuvwxyz";function v0(i){return $r.charAt(i)}function jr(i,e){return i&e}function U0(i,e){return i|e}function Pe(i,e){return i^e}function Oe(i,e){return i&~e}function Zr(i){if(i==0)return-1;var e=0;return i&65535||(i>>=16,e+=16),i&255||(i>>=8,e+=8),i&15||(i>>=4,e+=4),i&3||(i>>=2,e+=2),i&1||++e,e}function Xr(i){for(var e=0;i!=0;)i&=i-1,++e;return e}var m0="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",He="=";function V0(i){var e,t,r="";for(e=0;e+3<=i.length;e+=3)t=parseInt(i.substring(e,e+3),16),r+=m0.charAt(t>>6)+m0.charAt(t&63);for(e+1==i.length?(t=parseInt(i.substring(e,e+1),16),r+=m0.charAt(t<<2)):e+2==i.length&&(t=parseInt(i.substring(e,e+2),16),r+=m0.charAt(t>>2)+m0.charAt((t&3)<<4));(r.length&3)>0;)r+=He;return r}function qe(i){var e="",t,r=0,n=0;for(t=0;t<i.length&&i.charAt(t)!=He;++t){var s=m0.indexOf(i.charAt(t));s<0||(r==0?(e+=v0(s>>2),n=s&3,r=1):r==1?(e+=v0(n<<2|s>>4),n=s&15,r=2):r==2?(e+=v0(n),e+=v0(s>>2),n=s&3,r=3):(e+=v0(n<<2|s>>4),e+=v0(s&15),r=0))}return r==1&&(e+=v0(n<<2)),e}var b0,Qr={decode:function(i){var e;if(b0===void 0){var t="0123456789ABCDEF",r=` \f
+\r	 \u2028\u2029`;for(b0={},e=0;e<16;++e)b0[t.charAt(e)]=e;for(t=t.toLowerCase(),e=10;e<16;++e)b0[t.charAt(e)]=e;for(e=0;e<r.length;++e)b0[r.charAt(e)]=-1}var n=[],s=0,x=0;for(e=0;e<i.length;++e){var o=i.charAt(e);if(o=="=")break;if(o=b0[o],o!=-1){if(o===void 0)throw new Error("Illegal character at offset "+e);s|=o,++x>=2?(n[n.length]=s,s=0,x=0):s<<=4}}if(x)throw new Error("Hex encoding incomplete: 4 bits missing");return n}},_0,Wt={decode:function(i){var e;if(_0===void 0){var t="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/",r=`= \f
+\r	 \u2028\u2029`;for(_0=Object.create(null),e=0;e<64;++e)_0[t.charAt(e)]=e;for(_0["-"]=62,_0._=63,e=0;e<r.length;++e)_0[r.charAt(e)]=-1}var n=[],s=0,x=0;for(e=0;e<i.length;++e){var o=i.charAt(e);if(o=="=")break;if(o=_0[o],o!=-1){if(o===void 0)throw new Error("Illegal character at offset "+e);s|=o,++x>=4?(n[n.length]=s>>16,n[n.length]=s>>8&255,n[n.length]=s&255,s=0,x=0):s<<=6}}switch(x){case 1:throw new Error("Base64 encoding incomplete: at least 2 bits missing");case 2:n[n.length]=s>>10;break;case 3:n[n.length]=s>>16,n[n.length]=s>>8&255;break}return n},re:/-----BEGIN [^-]+-----([A-Za-z0-9+\/=\s]+)-----END [^-]+-----|begin-base64[^\n]+\n([A-Za-z0-9+\/=\s]+)====/,unarmor:function(i){var e=Wt.re.exec(i);if(e)if(e[1])i=e[1];else if(e[2])i=e[2];else throw new Error("RegExp out of sync");return Wt.decode(i)}},S0=1e13,P0=function(){function i(e){this.buf=[+e||0]}return i.prototype.mulAdd=function(e,t){var r=this.buf,n=r.length,s,x;for(s=0;s<n;++s)x=r[s]*e+t,x<S0?t=0:(t=0|x/S0,x-=t*S0),r[s]=x;t>0&&(r[s]=t)},i.prototype.sub=function(e){var t=this.buf,r=t.length,n,s;for(n=0;n<r;++n)s=t[n]-e,s<0?(s+=S0,e=1):e=0,t[n]=s;for(;t[t.length-1]===0;)t.pop()},i.prototype.toString=function(e){if((e||10)!=10)throw new Error("only base 10 is supported");for(var t=this.buf,r=t[t.length-1].toString(),n=t.length-2;n>=0;--n)r+=(S0+t[n]).toString().substring(1);return r},i.prototype.valueOf=function(){for(var e=this.buf,t=0,r=e.length-1;r>=0;--r)t=t*S0+e[r];return t},i.prototype.simplify=function(){var e=this.buf;return e.length==1?e[0]:this},i}(),Ne="…",Yr=/^(\d\d)(0[1-9]|1[0-2])(0[1-9]|[12]\d|3[01])([01]\d|2[0-3])(?:([0-5]\d)(?:([0-5]\d)(?:[.,](\d{1,3}))?)?)?(Z|[-+](?:[0]\d|1[0-2])([0-5]\d)?)?$/,Jr=/^(\d\d\d\d)(0[1-9]|1[0-2])(0[1-9]|[12]\d|3[01])([01]\d|2[0-3])(?:([0-5]\d)(?:([0-5]\d)(?:[.,](\d{1,3}))?)?)?(Z|[-+](?:[0]\d|1[0-2])([0-5]\d)?)?$/;function w0(i,e){return i.length>e&&(i=i.substring(0,e)+Ne),i}var Lt=function(){function i(e,t){this.hexDigits="0123456789ABCDEF",e instanceof i?(this.enc=e.enc,this.pos=e.pos):(this.enc=e,this.pos=t)}return i.prototype.get=function(e){if(e===void 0&&(e=this.pos++),e>=this.enc.length)throw new Error("Requesting byte offset ".concat(e," on a stream of length ").concat(this.enc.length));return typeof this.enc=="string"?this.enc.charCodeAt(e):this.enc[e]},i.prototype.hexByte=function(e){return this.hexDigits.charAt(e>>4&15)+this.hexDigits.charAt(e&15)},i.prototype.hexDump=function(e,t,r){for(var n="",s=e;s<t;++s)if(n+=this.hexByte(this.get(s)),r!==!0)switch(s&15){case 7:n+="  ";break;case 15:n+=`
+`;break;default:n+=" "}return n},i.prototype.isASCII=function(e,t){for(var r=e;r<t;++r){var n=this.get(r);if(n<32||n>176)return!1}return!0},i.prototype.parseStringISO=function(e,t){for(var r="",n=e;n<t;++n)r+=String.fromCharCode(this.get(n));return r},i.prototype.parseStringUTF=function(e,t){for(var r="",n=e;n<t;){var s=this.get(n++);s<128?r+=String.fromCharCode(s):s>191&&s<224?r+=String.fromCharCode((s&31)<<6|this.get(n++)&63):r+=String.fromCharCode((s&15)<<12|(this.get(n++)&63)<<6|this.get(n++)&63)}return r},i.prototype.parseStringBMP=function(e,t){for(var r="",n,s,x=e;x<t;)n=this.get(x++),s=this.get(x++),r+=String.fromCharCode(n<<8|s);return r},i.prototype.parseTime=function(e,t,r){var n=this.parseStringISO(e,t),s=(r?Yr:Jr).exec(n);return s?(r&&(s[1]=+s[1],s[1]+=+s[1]<70?2e3:1900),n=s[1]+"-"+s[2]+"-"+s[3]+" "+s[4],s[5]&&(n+=":"+s[5],s[6]&&(n+=":"+s[6],s[7]&&(n+="."+s[7]))),s[8]&&(n+=" UTC",s[8]!="Z"&&(n+=s[8],s[9]&&(n+=":"+s[9]))),n):"Unrecognized time: "+n},i.prototype.parseInteger=function(e,t){for(var r=this.get(e),n=r>127,s=n?255:0,x,o="";r==s&&++e<t;)r=this.get(e);if(x=t-e,x===0)return n?-1:0;if(x>4){for(o=r,x<<=3;!((+o^s)&128);)o=+o<<1,--x;o="("+x+` bit)
 `}n&&(r=r-256);for(var f=new P0(r),a=e+1;a<t;++a)f.mulAdd(256,this.get(a));return o+f.toString()},i.prototype.parseBitString=function(e,t,r){for(var n=this.get(e),s=(t-e-1<<3)-n,x="("+s+` bit)
-`,o="",f=e+1;f<t;++f){for(var a=this.get(f),c=f==t-1?n:0,E=7;E>=c;--E)o+=a>>E&1?"1":"0";if(o.length>r)return x+S0(o,r)}return x+o},i.prototype.parseOctetString=function(e,t,r){if(this.isASCII(e,t))return S0(this.parseStringISO(e,t),r);var n=t-e,s="("+n+` byte)
-`;r/=2,n>r&&(t=e+r);for(var x=e;x<t;++x)s+=this.hexByte(this.get(x));return n>r&&(s+=qe),s},i.prototype.parseOID=function(e,t,r){for(var n="",s=new P0,x=0,o=e;o<t;++o){var f=this.get(o);if(s.mulAdd(128,f&127),x+=7,!(f&128)){if(n==="")if(s=s.simplify(),s instanceof P0)s.sub(80),n="2."+s.toString();else{var a=s<80?s<40?0:1:2;n=a+"."+(s-a*40)}else n+="."+s.toString();if(n.length>r)return S0(n,r);s=new P0,x=0}}return x>0&&(n+=".incomplete"),n},i}(),Qr=function(){function i(e,t,r,n,s){if(!(n instanceof Ne))throw new Error("Invalid tag value.");this.stream=e,this.header=t,this.length=r,this.tag=n,this.sub=s}return i.prototype.typeName=function(){switch(this.tag.tagClass){case 0:switch(this.tag.tagNumber){case 0:return"EOC";case 1:return"BOOLEAN";case 2:return"INTEGER";case 3:return"BIT_STRING";case 4:return"OCTET_STRING";case 5:return"NULL";case 6:return"OBJECT_IDENTIFIER";case 7:return"ObjectDescriptor";case 8:return"EXTERNAL";case 9:return"REAL";case 10:return"ENUMERATED";case 11:return"EMBEDDED_PDV";case 12:return"UTF8String";case 16:return"SEQUENCE";case 17:return"SET";case 18:return"NumericString";case 19:return"PrintableString";case 20:return"TeletexString";case 21:return"VideotexString";case 22:return"IA5String";case 23:return"UTCTime";case 24:return"GeneralizedTime";case 25:return"GraphicString";case 26:return"VisibleString";case 27:return"GeneralString";case 28:return"UniversalString";case 30:return"BMPString"}return"Universal_"+this.tag.tagNumber.toString();case 1:return"Application_"+this.tag.tagNumber.toString();case 2:return"["+this.tag.tagNumber.toString()+"]";case 3:return"Private_"+this.tag.tagNumber.toString()}},i.prototype.content=function(e){if(this.tag===void 0)return null;e===void 0&&(e=1/0);var t=this.posContent(),r=Math.abs(this.length);if(!this.tag.isUniversal())return this.sub!==null?"("+this.sub.length+" elem)":this.stream.parseOctetString(t,t+r,e);switch(this.tag.tagNumber){case 1:return this.stream.get(t)===0?"false":"true";case 2:return this.stream.parseInteger(t,t+r);case 3:return this.sub?"("+this.sub.length+" elem)":this.stream.parseBitString(t,t+r,e);case 4:return this.sub?"("+this.sub.length+" elem)":this.stream.parseOctetString(t,t+r,e);case 6:return this.stream.parseOID(t,t+r,e);case 16:case 17:return this.sub!==null?"("+this.sub.length+" elem)":"(no elem)";case 12:return S0(this.stream.parseStringUTF(t,t+r),e);case 18:case 19:case 20:case 21:case 22:case 26:return S0(this.stream.parseStringISO(t,t+r),e);case 30:return S0(this.stream.parseStringBMP(t,t+r),e);case 23:case 24:return this.stream.parseTime(t,t+r,this.tag.tagNumber==23)}return null},i.prototype.toString=function(){return this.typeName()+"@"+this.stream.pos+"[header:"+this.header+",length:"+this.length+",sub:"+(this.sub===null?"null":this.sub.length)+"]"},i.prototype.toPrettyString=function(e){e===void 0&&(e="");var t=e+this.typeName()+" @"+this.stream.pos;if(this.length>=0&&(t+="+"),t+=this.length,this.tag.tagConstructed?t+=" (constructed)":this.tag.isUniversal()&&(this.tag.tagNumber==3||this.tag.tagNumber==4)&&this.sub!==null&&(t+=" (encapsulates)"),t+=`
-`,this.sub!==null){e+="  ";for(var r=0,n=this.sub.length;r<n;++r)t+=this.sub[r].toPrettyString(e)}return t},i.prototype.posStart=function(){return this.stream.pos},i.prototype.posContent=function(){return this.stream.pos+this.header},i.prototype.posEnd=function(){return this.stream.pos+this.header+Math.abs(this.length)},i.prototype.toHexString=function(){return this.stream.hexDump(this.posStart(),this.posEnd(),!0)},i.decodeLength=function(e){var t=e.get(),r=t&127;if(r==t)return r;if(r>6)throw new Error("Length over 48 bits not supported at position "+(e.pos-1));if(r===0)return null;t=0;for(var n=0;n<r;++n)t=t*256+e.get();return t},i.prototype.getHexStringValue=function(){var e=this.toHexString(),t=this.header*2,r=this.length*2;return e.substr(t,r)},i.decode=function(e){var t;e instanceof Wt?t=e:t=new Wt(e,0);var r=new Wt(t),n=new Ne(t),s=i.decodeLength(t),x=t.pos,o=x-r.pos,f=null,a=function(){var E=[];if(s!==null){for(var h=x+s;t.pos<h;)E[E.length]=i.decode(t);if(t.pos!=h)throw new Error("Content size is not correct for container starting at offset "+x)}else try{for(;;){var u=i.decode(t);if(u.tag.isEOC())break;E[E.length]=u}s=x-t.pos}catch(v){throw new Error("Exception while decoding undefined length content: "+v)}return E};if(n.tagConstructed)f=a();else if(n.isUniversal()&&(n.tagNumber==3||n.tagNumber==4))try{if(n.tagNumber==3&&t.get()!=0)throw new Error("BIT STRINGs with unused bits cannot encapsulate.");f=a();for(var c=0;c<f.length;++c)if(f[c].tag.isEOC())throw new Error("EOC is not supposed to be actual content.")}catch{f=null}if(f===null){if(s===null)throw new Error("We can't skip over an invalid tag with undefined length at offset "+x);t.pos=x+Math.abs(s)}return new i(r,o,s,n,f)},i}(),Ne=function(){function i(e){var t=e.get();if(this.tagClass=t>>6,this.tagConstructed=(t&32)!==0,this.tagNumber=t&31,this.tagNumber==31){var r=new P0;do t=e.get(),r.mulAdd(128,t&127);while(t&128);this.tagNumber=r.simplify()}}return i.prototype.isUniversal=function(){return this.tagClass===0},i.prototype.isEOC=function(){return this.tagClass===0&&this.tagNumber===0},i}(),g0,Yr=0xdeadbeefcafe,ze=(Yr&16777215)==15715070,e0=[2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73,79,83,89,97,101,103,107,109,113,127,131,137,139,149,151,157,163,167,173,179,181,191,193,197,199,211,223,227,229,233,239,241,251,257,263,269,271,277,281,283,293,307,311,313,317,331,337,347,349,353,359,367,373,379,383,389,397,401,409,419,421,431,433,439,443,449,457,461,463,467,479,487,491,499,503,509,521,523,541,547,557,563,569,571,577,587,593,599,601,607,613,617,619,631,641,643,647,653,659,661,673,677,683,691,701,709,719,727,733,739,743,751,757,761,769,773,787,797,809,811,821,823,827,829,839,853,857,859,863,877,881,883,887,907,911,919,929,937,941,947,953,967,971,977,983,991,997],Jr=(1<<26)/e0[e0.length-1],z=function(){function i(e,t,r){e!=null&&(typeof e=="number"?this.fromNumber(e,t,r):t==null&&typeof e!="string"?this.fromString(e,256):this.fromString(e,t))}return i.prototype.toString=function(e){if(this.s<0)return"-"+this.negate().toString(e);var t;if(e==16)t=4;else if(e==8)t=3;else if(e==2)t=1;else if(e==32)t=5;else if(e==4)t=2;else return this.toRadix(e);var r=(1<<t)-1,n,s=!1,x="",o=this.t,f=this.DB-o*this.DB%t;if(o-- >0)for(f<this.DB&&(n=this[o]>>f)>0&&(s=!0,x=d0(n));o>=0;)f<t?(n=(this[o]&(1<<f)-1)<<t-f,n|=this[--o]>>(f+=this.DB-t)):(n=this[o]>>(f-=t)&r,f<=0&&(f+=this.DB,--o)),n>0&&(s=!0),s&&(x+=d0(n));return s?x:"0"},i.prototype.negate=function(){var e=U();return i.ZERO.subTo(this,e),e},i.prototype.abs=function(){return this.s<0?this.negate():this},i.prototype.compareTo=function(e){var t=this.s-e.s;if(t!=0)return t;var r=this.t;if(t=r-e.t,t!=0)return this.s<0?-t:t;for(;--r>=0;)if((t=this[r]-e[r])!=0)return t;return 0},i.prototype.bitLength=function(){return this.t<=0?0:this.DB*(this.t-1)+K0(this[this.t-1]^this.s&this.DM)},i.prototype.mod=function(e){var t=U();return this.abs().divRemTo(e,null,t),this.s<0&&t.compareTo(i.ZERO)>0&&e.subTo(t,t),t},i.prototype.modPowInt=function(e,t){var r;return e<256||t.isEven()?r=new We(t):r=new Le(t),this.exp(e,r)},i.prototype.clone=function(){var e=U();return this.copyTo(e),e},i.prototype.intValue=function(){if(this.s<0){if(this.t==1)return this[0]-this.DV;if(this.t==0)return-1}else{if(this.t==1)return this[0];if(this.t==0)return 0}return(this[1]&(1<<32-this.DB)-1)<<this.DB|this[0]},i.prototype.byteValue=function(){return this.t==0?this.s:this[0]<<24>>24},i.prototype.shortValue=function(){return this.t==0?this.s:this[0]<<16>>16},i.prototype.signum=function(){return this.s<0?-1:this.t<=0||this.t==1&&this[0]<=0?0:1},i.prototype.toByteArray=function(){var e=this.t,t=[];t[0]=this.s;var r=this.DB-e*this.DB%8,n,s=0;if(e-- >0)for(r<this.DB&&(n=this[e]>>r)!=(this.s&this.DM)>>r&&(t[s++]=n|this.s<<this.DB-r);e>=0;)r<8?(n=(this[e]&(1<<r)-1)<<8-r,n|=this[--e]>>(r+=this.DB-8)):(n=this[e]>>(r-=8)&255,r<=0&&(r+=this.DB,--e)),n&128&&(n|=-256),s==0&&(this.s&128)!=(n&128)&&++s,(s>0||n!=this.s)&&(t[s++]=n);return t},i.prototype.equals=function(e){return this.compareTo(e)==0},i.prototype.min=function(e){return this.compareTo(e)<0?this:e},i.prototype.max=function(e){return this.compareTo(e)>0?this:e},i.prototype.and=function(e){var t=U();return this.bitwiseTo(e,Kr,t),t},i.prototype.or=function(e){var t=U();return this.bitwiseTo(e,M0,t),t},i.prototype.xor=function(e){var t=U();return this.bitwiseTo(e,Pe,t),t},i.prototype.andNot=function(e){var t=U();return this.bitwiseTo(e,Oe,t),t},i.prototype.not=function(){for(var e=U(),t=0;t<this.t;++t)e[t]=this.DM&~this[t];return e.t=this.t,e.s=~this.s,e},i.prototype.shiftLeft=function(e){var t=U();return e<0?this.rShiftTo(-e,t):this.lShiftTo(e,t),t},i.prototype.shiftRight=function(e){var t=U();return e<0?this.lShiftTo(-e,t):this.rShiftTo(e,t),t},i.prototype.getLowestSetBit=function(){for(var e=0;e<this.t;++e)if(this[e]!=0)return e*this.DB+$r(this[e]);return this.s<0?this.t*this.DB:-1},i.prototype.bitCount=function(){for(var e=0,t=this.s&this.DM,r=0;r<this.t;++r)e+=jr(this[r]^t);return e},i.prototype.testBit=function(e){var t=Math.floor(e/this.DB);return t>=this.t?this.s!=0:(this[t]&1<<e%this.DB)!=0},i.prototype.setBit=function(e){return this.changeBit(e,M0)},i.prototype.clearBit=function(e){return this.changeBit(e,Oe)},i.prototype.flipBit=function(e){return this.changeBit(e,Pe)},i.prototype.add=function(e){var t=U();return this.addTo(e,t),t},i.prototype.subtract=function(e){var t=U();return this.subTo(e,t),t},i.prototype.multiply=function(e){var t=U();return this.multiplyTo(e,t),t},i.prototype.divide=function(e){var t=U();return this.divRemTo(e,t,null),t},i.prototype.remainder=function(e){var t=U();return this.divRemTo(e,null,t),t},i.prototype.divideAndRemainder=function(e){var t=U(),r=U();return this.divRemTo(e,t,r),[t,r]},i.prototype.modPow=function(e,t){var r=e.bitLength(),n,s=E0(1),x;if(r<=0)return s;r<18?n=1:r<48?n=3:r<144?n=4:r<768?n=5:n=6,r<8?x=new We(t):t.isEven()?x=new ei(t):x=new Le(t);var o=[],f=3,a=n-1,c=(1<<n)-1;if(o[1]=x.convert(this),n>1){var E=U();for(x.sqrTo(o[1],E);f<=c;)o[f]=U(),x.mulTo(E,o[f-2],o[f]),f+=2}var h=e.t-1,u,v=!0,p=U(),l;for(r=K0(e[h])-1;h>=0;){for(r>=a?u=e[h]>>r-a&c:(u=(e[h]&(1<<r+1)-1)<<a-r,h>0&&(u|=e[h-1]>>this.DB+r-a)),f=n;!(u&1);)u>>=1,--f;if((r-=f)<0&&(r+=this.DB,--h),v)o[u].copyTo(s),v=!1;else{for(;f>1;)x.sqrTo(s,p),x.sqrTo(p,s),f-=2;f>0?x.sqrTo(s,p):(l=s,s=p,p=l),x.mulTo(p,o[u],s)}for(;h>=0&&!(e[h]&1<<r);)x.sqrTo(s,p),l=s,s=p,p=l,--r<0&&(r=this.DB-1,--h)}return x.revert(s)},i.prototype.modInverse=function(e){var t=e.isEven();if(this.isEven()&&t||e.signum()==0)return i.ZERO;for(var r=e.clone(),n=this.clone(),s=E0(1),x=E0(0),o=E0(0),f=E0(1);r.signum()!=0;){for(;r.isEven();)r.rShiftTo(1,r),t?((!s.isEven()||!x.isEven())&&(s.addTo(this,s),x.subTo(e,x)),s.rShiftTo(1,s)):x.isEven()||x.subTo(e,x),x.rShiftTo(1,x);for(;n.isEven();)n.rShiftTo(1,n),t?((!o.isEven()||!f.isEven())&&(o.addTo(this,o),f.subTo(e,f)),o.rShiftTo(1,o)):f.isEven()||f.subTo(e,f),f.rShiftTo(1,f);r.compareTo(n)>=0?(r.subTo(n,r),t&&s.subTo(o,s),x.subTo(f,x)):(n.subTo(r,n),t&&o.subTo(s,o),f.subTo(x,f))}if(n.compareTo(i.ONE)!=0)return i.ZERO;if(f.compareTo(e)>=0)return f.subtract(e);if(f.signum()<0)f.addTo(e,f);else return f;return f.signum()<0?f.add(e):f},i.prototype.pow=function(e){return this.exp(e,new ti)},i.prototype.gcd=function(e){var t=this.s<0?this.negate():this.clone(),r=e.s<0?e.negate():e.clone();if(t.compareTo(r)<0){var n=t;t=r,r=n}var s=t.getLowestSetBit(),x=r.getLowestSetBit();if(x<0)return t;for(s<x&&(x=s),x>0&&(t.rShiftTo(x,t),r.rShiftTo(x,r));t.signum()>0;)(s=t.getLowestSetBit())>0&&t.rShiftTo(s,t),(s=r.getLowestSetBit())>0&&r.rShiftTo(s,r),t.compareTo(r)>=0?(t.subTo(r,t),t.rShiftTo(1,t)):(r.subTo(t,r),r.rShiftTo(1,r));return x>0&&r.lShiftTo(x,r),r},i.prototype.isProbablePrime=function(e){var t,r=this.abs();if(r.t==1&&r[0]<=e0[e0.length-1]){for(t=0;t<e0.length;++t)if(r[0]==e0[t])return!0;return!1}if(r.isEven())return!1;for(t=1;t<e0.length;){for(var n=e0[t],s=t+1;s<e0.length&&n<Jr;)n*=e0[s++];for(n=r.modInt(n);t<s;)if(n%e0[t++]==0)return!1}return r.millerRabin(e)},i.prototype.copyTo=function(e){for(var t=this.t-1;t>=0;--t)e[t]=this[t];e.t=this.t,e.s=this.s},i.prototype.fromInt=function(e){this.t=1,this.s=e<0?-1:0,e>0?this[0]=e:e<-1?this[0]=e+this.DV:this.t=0},i.prototype.fromString=function(e,t){var r;if(t==16)r=4;else if(t==8)r=3;else if(t==256)r=8;else if(t==2)r=1;else if(t==32)r=5;else if(t==4)r=2;else{this.fromRadix(e,t);return}this.t=0,this.s=0;for(var n=e.length,s=!1,x=0;--n>=0;){var o=r==8?+e[n]&255:Ue(e,n);if(o<0){e.charAt(n)=="-"&&(s=!0);continue}s=!1,x==0?this[this.t++]=o:x+r>this.DB?(this[this.t-1]|=(o&(1<<this.DB-x)-1)<<x,this[this.t++]=o>>this.DB-x):this[this.t-1]|=o<<x,x+=r,x>=this.DB&&(x-=this.DB)}r==8&&+e[0]&128&&(this.s=-1,x>0&&(this[this.t-1]|=(1<<this.DB-x)-1<<x)),this.clamp(),s&&i.ZERO.subTo(this,this)},i.prototype.clamp=function(){for(var e=this.s&this.DM;this.t>0&&this[this.t-1]==e;)--this.t},i.prototype.dlShiftTo=function(e,t){var r;for(r=this.t-1;r>=0;--r)t[r+e]=this[r];for(r=e-1;r>=0;--r)t[r]=0;t.t=this.t+e,t.s=this.s},i.prototype.drShiftTo=function(e,t){for(var r=e;r<this.t;++r)t[r-e]=this[r];t.t=Math.max(this.t-e,0),t.s=this.s},i.prototype.lShiftTo=function(e,t){for(var r=e%this.DB,n=this.DB-r,s=(1<<n)-1,x=Math.floor(e/this.DB),o=this.s<<r&this.DM,f=this.t-1;f>=0;--f)t[f+x+1]=this[f]>>n|o,o=(this[f]&s)<<r;for(var f=x-1;f>=0;--f)t[f]=0;t[x]=o,t.t=this.t+x+1,t.s=this.s,t.clamp()},i.prototype.rShiftTo=function(e,t){t.s=this.s;var r=Math.floor(e/this.DB);if(r>=this.t){t.t=0;return}var n=e%this.DB,s=this.DB-n,x=(1<<n)-1;t[0]=this[r]>>n;for(var o=r+1;o<this.t;++o)t[o-r-1]|=(this[o]&x)<<s,t[o-r]=this[o]>>n;n>0&&(t[this.t-r-1]|=(this.s&x)<<s),t.t=this.t-r,t.clamp()},i.prototype.subTo=function(e,t){for(var r=0,n=0,s=Math.min(e.t,this.t);r<s;)n+=this[r]-e[r],t[r++]=n&this.DM,n>>=this.DB;if(e.t<this.t){for(n-=e.s;r<this.t;)n+=this[r],t[r++]=n&this.DM,n>>=this.DB;n+=this.s}else{for(n+=this.s;r<e.t;)n-=e[r],t[r++]=n&this.DM,n>>=this.DB;n-=e.s}t.s=n<0?-1:0,n<-1?t[r++]=this.DV+n:n>0&&(t[r++]=n),t.t=r,t.clamp()},i.prototype.multiplyTo=function(e,t){var r=this.abs(),n=e.abs(),s=r.t;for(t.t=s+n.t;--s>=0;)t[s]=0;for(s=0;s<n.t;++s)t[s+r.t]=r.am(0,n[s],t,s,0,r.t);t.s=0,t.clamp(),this.s!=e.s&&i.ZERO.subTo(t,t)},i.prototype.squareTo=function(e){for(var t=this.abs(),r=e.t=2*t.t;--r>=0;)e[r]=0;for(r=0;r<t.t-1;++r){var n=t.am(r,t[r],e,2*r,0,1);(e[r+t.t]+=t.am(r+1,2*t[r],e,2*r+1,n,t.t-r-1))>=t.DV&&(e[r+t.t]-=t.DV,e[r+t.t+1]=1)}e.t>0&&(e[e.t-1]+=t.am(r,t[r],e,2*r,0,1)),e.s=0,e.clamp()},i.prototype.divRemTo=function(e,t,r){var n=e.abs();if(!(n.t<=0)){var s=this.abs();if(s.t<n.t){t!=null&&t.fromInt(0),r!=null&&this.copyTo(r);return}r==null&&(r=U());var x=U(),o=this.s,f=e.s,a=this.DB-K0(n[n.t-1]);a>0?(n.lShiftTo(a,x),s.lShiftTo(a,r)):(n.copyTo(x),s.copyTo(r));var c=x.t,E=x[c-1];if(E!=0){var h=E*(1<<this.F1)+(c>1?x[c-2]>>this.F2:0),u=this.FV/h,v=(1<<this.F1)/h,p=1<<this.F2,l=r.t,C=l-c,d=t??U();for(x.dlShiftTo(C,d),r.compareTo(d)>=0&&(r[r.t++]=1,r.subTo(d,r)),i.ONE.dlShiftTo(c,d),d.subTo(x,x);x.t<c;)x[x.t++]=0;for(;--C>=0;){var g=r[--l]==E?this.DM:Math.floor(r[l]*u+(r[l-1]+p)*v);if((r[l]+=x.am(0,g,r,C,0,c))<g)for(x.dlShiftTo(C,d),r.subTo(d,r);r[l]<--g;)r.subTo(d,r)}t!=null&&(r.drShiftTo(c,t),o!=f&&i.ZERO.subTo(t,t)),r.t=c,r.clamp(),a>0&&r.rShiftTo(a,r),o<0&&i.ZERO.subTo(r,r)}}},i.prototype.invDigit=function(){if(this.t<1)return 0;var e=this[0];if(!(e&1))return 0;var t=e&3;return t=t*(2-(e&15)*t)&15,t=t*(2-(e&255)*t)&255,t=t*(2-((e&65535)*t&65535))&65535,t=t*(2-e*t%this.DV)%this.DV,t>0?this.DV-t:-t},i.prototype.isEven=function(){return(this.t>0?this[0]&1:this.s)==0},i.prototype.exp=function(e,t){if(e>4294967295||e<1)return i.ONE;var r=U(),n=U(),s=t.convert(this),x=K0(e)-1;for(s.copyTo(r);--x>=0;)if(t.sqrTo(r,n),(e&1<<x)>0)t.mulTo(n,s,r);else{var o=r;r=n,n=o}return t.revert(r)},i.prototype.chunkSize=function(e){return Math.floor(Math.LN2*this.DB/Math.log(e))},i.prototype.toRadix=function(e){if(e==null&&(e=10),this.signum()==0||e<2||e>36)return"0";var t=this.chunkSize(e),r=Math.pow(e,t),n=E0(r),s=U(),x=U(),o="";for(this.divRemTo(n,s,x);s.signum()>0;)o=(r+x.intValue()).toString(e).substr(1)+o,s.divRemTo(n,s,x);return x.intValue().toString(e)+o},i.prototype.fromRadix=function(e,t){this.fromInt(0),t==null&&(t=10);for(var r=this.chunkSize(t),n=Math.pow(t,r),s=!1,x=0,o=0,f=0;f<e.length;++f){var a=Ue(e,f);if(a<0){e.charAt(f)=="-"&&this.signum()==0&&(s=!0);continue}o=t*o+a,++x>=r&&(this.dMultiply(n),this.dAddOffset(o,0),x=0,o=0)}x>0&&(this.dMultiply(Math.pow(t,x)),this.dAddOffset(o,0)),s&&i.ZERO.subTo(this,this)},i.prototype.fromNumber=function(e,t,r){if(typeof t=="number")if(e<2)this.fromInt(1);else for(this.fromNumber(e,r),this.testBit(e-1)||this.bitwiseTo(i.ONE.shiftLeft(e-1),M0,this),this.isEven()&&this.dAddOffset(1,0);!this.isProbablePrime(t);)this.dAddOffset(2,0),this.bitLength()>e&&this.subTo(i.ONE.shiftLeft(e-1),this);else{var n=[],s=e&7;n.length=(e>>3)+1,t.nextBytes(n),s>0?n[0]&=(1<<s)-1:n[0]=0,this.fromString(n,256)}},i.prototype.bitwiseTo=function(e,t,r){var n,s,x=Math.min(e.t,this.t);for(n=0;n<x;++n)r[n]=t(this[n],e[n]);if(e.t<this.t){for(s=e.s&this.DM,n=x;n<this.t;++n)r[n]=t(this[n],s);r.t=this.t}else{for(s=this.s&this.DM,n=x;n<e.t;++n)r[n]=t(s,e[n]);r.t=e.t}r.s=t(this.s,e.s),r.clamp()},i.prototype.changeBit=function(e,t){var r=i.ONE.shiftLeft(e);return this.bitwiseTo(r,t,r),r},i.prototype.addTo=function(e,t){for(var r=0,n=0,s=Math.min(e.t,this.t);r<s;)n+=this[r]+e[r],t[r++]=n&this.DM,n>>=this.DB;if(e.t<this.t){for(n+=e.s;r<this.t;)n+=this[r],t[r++]=n&this.DM,n>>=this.DB;n+=this.s}else{for(n+=this.s;r<e.t;)n+=e[r],t[r++]=n&this.DM,n>>=this.DB;n+=e.s}t.s=n<0?-1:0,n>0?t[r++]=n:n<-1&&(t[r++]=this.DV+n),t.t=r,t.clamp()},i.prototype.dMultiply=function(e){this[this.t]=this.am(0,e-1,this,0,0,this.t),++this.t,this.clamp()},i.prototype.dAddOffset=function(e,t){if(e!=0){for(;this.t<=t;)this[this.t++]=0;for(this[t]+=e;this[t]>=this.DV;)this[t]-=this.DV,++t>=this.t&&(this[this.t++]=0),++this[t]}},i.prototype.multiplyLowerTo=function(e,t,r){var n=Math.min(this.t+e.t,t);for(r.s=0,r.t=n;n>0;)r[--n]=0;for(var s=r.t-this.t;n<s;++n)r[n+this.t]=this.am(0,e[n],r,n,0,this.t);for(var s=Math.min(e.t,t);n<s;++n)this.am(0,e[n],r,n,0,t-n);r.clamp()},i.prototype.multiplyUpperTo=function(e,t,r){--t;var n=r.t=this.t+e.t-t;for(r.s=0;--n>=0;)r[n]=0;for(n=Math.max(t-this.t,0);n<e.t;++n)r[this.t+n-t]=this.am(t-n,e[n],r,0,0,this.t+n-t);r.clamp(),r.drShiftTo(1,r)},i.prototype.modInt=function(e){if(e<=0)return 0;var t=this.DV%e,r=this.s<0?e-1:0;if(this.t>0)if(t==0)r=this[0]%e;else for(var n=this.t-1;n>=0;--n)r=(t*r+this[n])%e;return r},i.prototype.millerRabin=function(e){var t=this.subtract(i.ONE),r=t.getLowestSetBit();if(r<=0)return!1;var n=t.shiftRight(r);e=e+1>>1,e>e0.length&&(e=e0.length);for(var s=U(),x=0;x<e;++x){s.fromInt(e0[Math.floor(Math.random()*e0.length)]);var o=s.modPow(n,this);if(o.compareTo(i.ONE)!=0&&o.compareTo(t)!=0){for(var f=1;f++<r&&o.compareTo(t)!=0;)if(o=o.modPowInt(2,this),o.compareTo(i.ONE)==0)return!1;if(o.compareTo(t)!=0)return!1}}return!0},i.prototype.square=function(){var e=U();return this.squareTo(e),e},i.prototype.gcda=function(e,t){var r=this.s<0?this.negate():this.clone(),n=e.s<0?e.negate():e.clone();if(r.compareTo(n)<0){var s=r;r=n,n=s}var x=r.getLowestSetBit(),o=n.getLowestSetBit();if(o<0){t(r);return}x<o&&(o=x),o>0&&(r.rShiftTo(o,r),n.rShiftTo(o,n));var f=function(){(x=r.getLowestSetBit())>0&&r.rShiftTo(x,r),(x=n.getLowestSetBit())>0&&n.rShiftTo(x,n),r.compareTo(n)>=0?(r.subTo(n,r),r.rShiftTo(1,r)):(n.subTo(r,n),n.rShiftTo(1,n)),r.signum()>0?setTimeout(f,0):(o>0&&n.lShiftTo(o,n),setTimeout(function(){t(n)},0))};setTimeout(f,10)},i.prototype.fromNumberAsync=function(e,t,r,n){if(typeof t=="number")if(e<2)this.fromInt(1);else{this.fromNumber(e,r),this.testBit(e-1)||this.bitwiseTo(i.ONE.shiftLeft(e-1),M0,this),this.isEven()&&this.dAddOffset(1,0);var s=this,x=function(){s.dAddOffset(2,0),s.bitLength()>e&&s.subTo(i.ONE.shiftLeft(e-1),s),s.isProbablePrime(t)?setTimeout(function(){n()},0):setTimeout(x,0)};setTimeout(x,0)}else{var o=[],f=e&7;o.length=(e>>3)+1,t.nextBytes(o),f>0?o[0]&=(1<<f)-1:o[0]=0,this.fromString(o,256)}},i}(),ti=function(){function i(){}return i.prototype.convert=function(e){return e},i.prototype.revert=function(e){return e},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r)},i.prototype.sqrTo=function(e,t){e.squareTo(t)},i}(),We=function(){function i(e){this.m=e}return i.prototype.convert=function(e){return e.s<0||e.compareTo(this.m)>=0?e.mod(this.m):e},i.prototype.revert=function(e){return e},i.prototype.reduce=function(e){e.divRemTo(this.m,null,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}(),Le=function(){function i(e){this.m=e,this.mp=e.invDigit(),this.mpl=this.mp&32767,this.mph=this.mp>>15,this.um=(1<<e.DB-15)-1,this.mt2=2*e.t}return i.prototype.convert=function(e){var t=U();return e.abs().dlShiftTo(this.m.t,t),t.divRemTo(this.m,null,t),e.s<0&&t.compareTo(z.ZERO)>0&&this.m.subTo(t,t),t},i.prototype.revert=function(e){var t=U();return e.copyTo(t),this.reduce(t),t},i.prototype.reduce=function(e){for(;e.t<=this.mt2;)e[e.t++]=0;for(var t=0;t<this.m.t;++t){var r=e[t]&32767,n=r*this.mpl+((r*this.mph+(e[t]>>15)*this.mpl&this.um)<<15)&e.DM;for(r=t+this.m.t,e[r]+=this.m.am(0,n,e,t,0,this.m.t);e[r]>=e.DV;)e[r]-=e.DV,e[++r]++}e.clamp(),e.drShiftTo(this.m.t,e),e.compareTo(this.m)>=0&&e.subTo(this.m,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}(),ei=function(){function i(e){this.m=e,this.r2=U(),this.q3=U(),z.ONE.dlShiftTo(2*e.t,this.r2),this.mu=this.r2.divide(e)}return i.prototype.convert=function(e){if(e.s<0||e.t>2*this.m.t)return e.mod(this.m);if(e.compareTo(this.m)<0)return e;var t=U();return e.copyTo(t),this.reduce(t),t},i.prototype.revert=function(e){return e},i.prototype.reduce=function(e){for(e.drShiftTo(this.m.t-1,this.r2),e.t>this.m.t+1&&(e.t=this.m.t+1,e.clamp()),this.mu.multiplyUpperTo(this.r2,this.m.t+1,this.q3),this.m.multiplyLowerTo(this.q3,this.m.t+1,this.r2);e.compareTo(this.r2)<0;)e.dAddOffset(1,this.m.t+1);for(e.subTo(this.r2,e);e.compareTo(this.m)>=0;)e.subTo(this.m,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}();function U(){return new z(null)}function Q(i,e){return new z(i,e)}var Me=typeof navigator<"u";Me&&ze&&navigator.appName=="Microsoft Internet Explorer"?(z.prototype.am=function(e,t,r,n,s,x){for(var o=t&32767,f=t>>15;--x>=0;){var a=this[e]&32767,c=this[e++]>>15,E=f*a+c*o;a=o*a+((E&32767)<<15)+r[n]+(s&1073741823),s=(a>>>30)+(E>>>15)+f*c+(s>>>30),r[n++]=a&1073741823}return s},g0=30):Me&&ze&&navigator.appName!="Netscape"?(z.prototype.am=function(e,t,r,n,s,x){for(;--x>=0;){var o=t*this[e++]+r[n]+s;s=Math.floor(o/67108864),r[n++]=o&67108863}return s},g0=26):(z.prototype.am=function(e,t,r,n,s,x){for(var o=t&16383,f=t>>14;--x>=0;){var a=this[e]&16383,c=this[e++]>>14,E=f*a+c*o;a=o*a+((E&16383)<<14)+r[n]+s,s=(a>>28)+(E>>14)+f*c,r[n++]=a&268435455}return s},g0=28),z.prototype.DB=g0,z.prototype.DM=(1<<g0)-1,z.prototype.DV=1<<g0;var Lt=52;z.prototype.FV=Math.pow(2,Lt),z.prototype.F1=Lt-g0,z.prototype.F2=2*g0-Lt;var V0=[],w0,f0;for(w0=48,f0=0;f0<=9;++f0)V0[w0++]=f0;for(w0=97,f0=10;f0<36;++f0)V0[w0++]=f0;for(w0=65,f0=10;f0<36;++f0)V0[w0++]=f0;function Ue(i,e){var t=V0[i.charCodeAt(e)];return t??-1}function E0(i){var e=U();return e.fromInt(i),e}function K0(i){var e=1,t;return(t=i>>>16)!=0&&(i=t,e+=16),(t=i>>8)!=0&&(i=t,e+=8),(t=i>>4)!=0&&(i=t,e+=4),(t=i>>2)!=0&&(i=t,e+=2),(t=i>>1)!=0&&(i=t,e+=1),e}z.ZERO=E0(0),z.ONE=E0(1);var ri=function(){function i(){this.i=0,this.j=0,this.S=[]}return i.prototype.init=function(e){var t,r,n;for(t=0;t<256;++t)this.S[t]=t;for(r=0,t=0;t<256;++t)r=r+this.S[t]+e[t%e.length]&255,n=this.S[t],this.S[t]=this.S[r],this.S[r]=n;this.i=0,this.j=0},i.prototype.next=function(){var e;return this.i=this.i+1&255,this.j=this.j+this.S[this.i]&255,e=this.S[this.i],this.S[this.i]=this.S[this.j],this.S[this.j]=e,this.S[e+this.S[this.i]&255]},i}();function ii(){return new ri}var Ve=256,$0,B0=null,h0;if(B0==null){B0=[],h0=0;var j0=void 0;if(typeof window<"u"&&window.crypto&&window.crypto.getRandomValues){var Mt=new Uint32Array(256);for(window.crypto.getRandomValues(Mt),j0=0;j0<Mt.length;++j0)B0[h0++]=Mt[j0]&255}var G0=0,Z0=function(i){if(G0=G0||0,G0>=256||h0>=Ve){window.removeEventListener?window.removeEventListener("mousemove",Z0,!1):window.detachEvent&&window.detachEvent("onmousemove",Z0);return}try{var e=i.x+i.y;B0[h0++]=e&255,G0+=1}catch{}};typeof window<"u"&&(window.addEventListener?window.addEventListener("mousemove",Z0,!1):window.attachEvent&&window.attachEvent("onmousemove",Z0))}function ni(){if($0==null){for($0=ii();h0<Ve;){var i=Math.floor(65536*Math.random());B0[h0++]=i&255}for($0.init(B0),h0=0;h0<B0.length;++h0)B0[h0]=0;h0=0}return $0.next()}var Ut=function(){function i(){}return i.prototype.nextBytes=function(e){for(var t=0;t<e.length;++t)e[t]=ni()},i}();function si(i,e){if(e<i.length+22)return console.error("Message too long for RSA"),null;for(var t=e-i.length-6,r="",n=0;n<t;n+=2)r+="ff";var s="0001"+r+"00"+i;return Q(s,16)}function oi(i,e){if(e<i.length+11)return console.error("Message too long for RSA"),null;for(var t=[],r=i.length-1;r>=0&&e>0;){var n=i.charCodeAt(r--);n<128?t[--e]=n:n>127&&n<2048?(t[--e]=n&63|128,t[--e]=n>>6|192):(t[--e]=n&63|128,t[--e]=n>>6&63|128,t[--e]=n>>12|224)}t[--e]=0;for(var s=new Ut,x=[];e>2;){for(x[0]=0;x[0]==0;)s.nextBytes(x);t[--e]=x[0]}return t[--e]=2,t[--e]=0,new z(t)}var ai=function(){function i(){this.n=null,this.e=0,this.d=null,this.p=null,this.q=null,this.dmp1=null,this.dmq1=null,this.coeff=null}return i.prototype.doPublic=function(e){return e.modPowInt(this.e,this.n)},i.prototype.doPrivate=function(e){if(this.p==null||this.q==null)return e.modPow(this.d,this.n);for(var t=e.mod(this.p).modPow(this.dmp1,this.p),r=e.mod(this.q).modPow(this.dmq1,this.q);t.compareTo(r)<0;)t=t.add(this.p);return t.subtract(r).multiply(this.coeff).mod(this.p).multiply(this.q).add(r)},i.prototype.setPublic=function(e,t){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16)):console.error("Invalid RSA public key")},i.prototype.encrypt=function(e){var t=this.n.bitLength()+7>>3,r=oi(e,t);if(r==null)return null;var n=this.doPublic(r);if(n==null)return null;for(var s=n.toString(16),x=s.length,o=0;o<t*2-x;o++)s="0"+s;return s},i.prototype.setPrivate=function(e,t,r){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16),this.d=Q(r,16)):console.error("Invalid RSA private key")},i.prototype.setPrivateEx=function(e,t,r,n,s,x,o,f){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16),this.d=Q(r,16),this.p=Q(n,16),this.q=Q(s,16),this.dmp1=Q(x,16),this.dmq1=Q(o,16),this.coeff=Q(f,16)):console.error("Invalid RSA private key")},i.prototype.generate=function(e,t){var r=new Ut,n=e>>1;this.e=parseInt(t,16);for(var s=new z(t,16);;){for(;this.p=new z(e-n,1,r),!(this.p.subtract(z.ONE).gcd(s).compareTo(z.ONE)==0&&this.p.isProbablePrime(10)););for(;this.q=new z(n,1,r),!(this.q.subtract(z.ONE).gcd(s).compareTo(z.ONE)==0&&this.q.isProbablePrime(10)););if(this.p.compareTo(this.q)<=0){var x=this.p;this.p=this.q,this.q=x}var o=this.p.subtract(z.ONE),f=this.q.subtract(z.ONE),a=o.multiply(f);if(a.gcd(s).compareTo(z.ONE)==0){this.n=this.p.multiply(this.q),this.d=s.modInverse(a),this.dmp1=this.d.mod(o),this.dmq1=this.d.mod(f),this.coeff=this.q.modInverse(this.p);break}}},i.prototype.decrypt=function(e){var t=Q(e,16),r=this.doPrivate(t);return r==null?null:xi(r,this.n.bitLength()+7>>3)},i.prototype.generateAsync=function(e,t,r){var n=new Ut,s=e>>1;this.e=parseInt(t,16);var x=new z(t,16),o=this,f=function(){var a=function(){if(o.p.compareTo(o.q)<=0){var h=o.p;o.p=o.q,o.q=h}var u=o.p.subtract(z.ONE),v=o.q.subtract(z.ONE),p=u.multiply(v);p.gcd(x).compareTo(z.ONE)==0?(o.n=o.p.multiply(o.q),o.d=x.modInverse(p),o.dmp1=o.d.mod(u),o.dmq1=o.d.mod(v),o.coeff=o.q.modInverse(o.p),setTimeout(function(){r()},0)):setTimeout(f,0)},c=function(){o.q=U(),o.q.fromNumberAsync(s,1,n,function(){o.q.subtract(z.ONE).gcda(x,function(h){h.compareTo(z.ONE)==0&&o.q.isProbablePrime(10)?setTimeout(a,0):setTimeout(c,0)})})},E=function(){o.p=U(),o.p.fromNumberAsync(e-s,1,n,function(){o.p.subtract(z.ONE).gcda(x,function(h){h.compareTo(z.ONE)==0&&o.p.isProbablePrime(10)?setTimeout(c,0):setTimeout(E,0)})})};setTimeout(E,0)};setTimeout(f,0)},i.prototype.sign=function(e,t,r){var n=fi(r),s=n+t(e).toString(),x=si(s,this.n.bitLength()/4);if(x==null)return null;var o=this.doPrivate(x);if(o==null)return null;var f=o.toString(16);return f.length&1?"0"+f:f},i.prototype.verify=function(e,t,r){var n=Q(t,16),s=this.doPublic(n);if(s==null)return null;var x=s.toString(16).replace(/^1f+00/,""),o=ci(x);return o==r(e).toString()},i}();function xi(i,e){for(var t=i.toByteArray(),r=0;r<t.length&&t[r]==0;)++r;if(t.length-r!=e-1||t[r]!=2)return null;for(++r;t[r]!=0;)if(++r>=t.length)return null;for(var n="";++r<t.length;){var s=t[r]&255;s<128?n+=String.fromCharCode(s):s>191&&s<224?(n+=String.fromCharCode((s&31)<<6|t[r+1]&63),++r):(n+=String.fromCharCode((s&15)<<12|(t[r+1]&63)<<6|t[r+2]&63),r+=2)}return n}var X0={md2:"3020300c06082a864886f70d020205000410",md5:"3020300c06082a864886f70d020505000410",sha1:"3021300906052b0e03021a05000414",sha224:"302d300d06096086480165030402040500041c",sha256:"3031300d060960864801650304020105000420",sha384:"3041300d060960864801650304020205000430",sha512:"3051300d060960864801650304020305000440",ripemd160:"3021300906052b2403020105000414"};function fi(i){return X0[i]||""}function ci(i){for(var e in X0)if(X0.hasOwnProperty(e)){var t=X0[e],r=t.length;if(i.substr(0,r)==t)return i.substr(r)}return i}/*!
+`,o="",f=e+1;f<t;++f){for(var a=this.get(f),c=f==t-1?n:0,B=7;B>=c;--B)o+=a>>B&1?"1":"0";if(o.length>r)return x+w0(o,r)}return x+o},i.prototype.parseOctetString=function(e,t,r){if(this.isASCII(e,t))return w0(this.parseStringISO(e,t),r);var n=t-e,s="("+n+` byte)
+`;r/=2,n>r&&(t=e+r);for(var x=e;x<t;++x)s+=this.hexByte(this.get(x));return n>r&&(s+=Ne),s},i.prototype.parseOID=function(e,t,r){for(var n="",s=new P0,x=0,o=e;o<t;++o){var f=this.get(o);if(s.mulAdd(128,f&127),x+=7,!(f&128)){if(n==="")if(s=s.simplify(),s instanceof P0)s.sub(80),n="2."+s.toString();else{var a=s<80?s<40?0:1:2;n=a+"."+(s-a*40)}else n+="."+s.toString();if(n.length>r)return w0(n,r);s=new P0,x=0}}return x>0&&(n+=".incomplete"),n},i}(),ti=function(){function i(e,t,r,n,s){if(!(n instanceof ze))throw new Error("Invalid tag value.");this.stream=e,this.header=t,this.length=r,this.tag=n,this.sub=s}return i.prototype.typeName=function(){switch(this.tag.tagClass){case 0:switch(this.tag.tagNumber){case 0:return"EOC";case 1:return"BOOLEAN";case 2:return"INTEGER";case 3:return"BIT_STRING";case 4:return"OCTET_STRING";case 5:return"NULL";case 6:return"OBJECT_IDENTIFIER";case 7:return"ObjectDescriptor";case 8:return"EXTERNAL";case 9:return"REAL";case 10:return"ENUMERATED";case 11:return"EMBEDDED_PDV";case 12:return"UTF8String";case 16:return"SEQUENCE";case 17:return"SET";case 18:return"NumericString";case 19:return"PrintableString";case 20:return"TeletexString";case 21:return"VideotexString";case 22:return"IA5String";case 23:return"UTCTime";case 24:return"GeneralizedTime";case 25:return"GraphicString";case 26:return"VisibleString";case 27:return"GeneralString";case 28:return"UniversalString";case 30:return"BMPString"}return"Universal_"+this.tag.tagNumber.toString();case 1:return"Application_"+this.tag.tagNumber.toString();case 2:return"["+this.tag.tagNumber.toString()+"]";case 3:return"Private_"+this.tag.tagNumber.toString()}},i.prototype.content=function(e){if(this.tag===void 0)return null;e===void 0&&(e=1/0);var t=this.posContent(),r=Math.abs(this.length);if(!this.tag.isUniversal())return this.sub!==null?"("+this.sub.length+" elem)":this.stream.parseOctetString(t,t+r,e);switch(this.tag.tagNumber){case 1:return this.stream.get(t)===0?"false":"true";case 2:return this.stream.parseInteger(t,t+r);case 3:return this.sub?"("+this.sub.length+" elem)":this.stream.parseBitString(t,t+r,e);case 4:return this.sub?"("+this.sub.length+" elem)":this.stream.parseOctetString(t,t+r,e);case 6:return this.stream.parseOID(t,t+r,e);case 16:case 17:return this.sub!==null?"("+this.sub.length+" elem)":"(no elem)";case 12:return w0(this.stream.parseStringUTF(t,t+r),e);case 18:case 19:case 20:case 21:case 22:case 26:return w0(this.stream.parseStringISO(t,t+r),e);case 30:return w0(this.stream.parseStringBMP(t,t+r),e);case 23:case 24:return this.stream.parseTime(t,t+r,this.tag.tagNumber==23)}return null},i.prototype.toString=function(){return this.typeName()+"@"+this.stream.pos+"[header:"+this.header+",length:"+this.length+",sub:"+(this.sub===null?"null":this.sub.length)+"]"},i.prototype.toPrettyString=function(e){e===void 0&&(e="");var t=e+this.typeName()+" @"+this.stream.pos;if(this.length>=0&&(t+="+"),t+=this.length,this.tag.tagConstructed?t+=" (constructed)":this.tag.isUniversal()&&(this.tag.tagNumber==3||this.tag.tagNumber==4)&&this.sub!==null&&(t+=" (encapsulates)"),t+=`
+`,this.sub!==null){e+="  ";for(var r=0,n=this.sub.length;r<n;++r)t+=this.sub[r].toPrettyString(e)}return t},i.prototype.posStart=function(){return this.stream.pos},i.prototype.posContent=function(){return this.stream.pos+this.header},i.prototype.posEnd=function(){return this.stream.pos+this.header+Math.abs(this.length)},i.prototype.toHexString=function(){return this.stream.hexDump(this.posStart(),this.posEnd(),!0)},i.decodeLength=function(e){var t=e.get(),r=t&127;if(r==t)return r;if(r>6)throw new Error("Length over 48 bits not supported at position "+(e.pos-1));if(r===0)return null;t=0;for(var n=0;n<r;++n)t=t*256+e.get();return t},i.prototype.getHexStringValue=function(){var e=this.toHexString(),t=this.header*2,r=this.length*2;return e.substr(t,r)},i.decode=function(e){var t;e instanceof Lt?t=e:t=new Lt(e,0);var r=new Lt(t),n=new ze(t),s=i.decodeLength(t),x=t.pos,o=x-r.pos,f=null,a=function(){var B=[];if(s!==null){for(var h=x+s;t.pos<h;)B[B.length]=i.decode(t);if(t.pos!=h)throw new Error("Content size is not correct for container starting at offset "+x)}else try{for(;;){var u=i.decode(t);if(u.tag.isEOC())break;B[B.length]=u}s=x-t.pos}catch(v){throw new Error("Exception while decoding undefined length content: "+v)}return B};if(n.tagConstructed)f=a();else if(n.isUniversal()&&(n.tagNumber==3||n.tagNumber==4))try{if(n.tagNumber==3&&t.get()!=0)throw new Error("BIT STRINGs with unused bits cannot encapsulate.");f=a();for(var c=0;c<f.length;++c)if(f[c].tag.isEOC())throw new Error("EOC is not supposed to be actual content.")}catch{f=null}if(f===null){if(s===null)throw new Error("We can't skip over an invalid tag with undefined length at offset "+x);t.pos=x+Math.abs(s)}return new i(r,o,s,n,f)},i}(),ze=function(){function i(e){var t=e.get();if(this.tagClass=t>>6,this.tagConstructed=(t&32)!==0,this.tagNumber=t&31,this.tagNumber==31){var r=new P0;do t=e.get(),r.mulAdd(128,t&127);while(t&128);this.tagNumber=r.simplify()}}return i.prototype.isUniversal=function(){return this.tagClass===0},i.prototype.isEOC=function(){return this.tagClass===0&&this.tagNumber===0},i}(),E0,ei=0xdeadbeefcafe,We=(ei&16777215)==15715070,e0=[2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73,79,83,89,97,101,103,107,109,113,127,131,137,139,149,151,157,163,167,173,179,181,191,193,197,199,211,223,227,229,233,239,241,251,257,263,269,271,277,281,283,293,307,311,313,317,331,337,347,349,353,359,367,373,379,383,389,397,401,409,419,421,431,433,439,443,449,457,461,463,467,479,487,491,499,503,509,521,523,541,547,557,563,569,571,577,587,593,599,601,607,613,617,619,631,641,643,647,653,659,661,673,677,683,691,701,709,719,727,733,739,743,751,757,761,769,773,787,797,809,811,821,823,827,829,839,853,857,859,863,877,881,883,887,907,911,919,929,937,941,947,953,967,971,977,983,991,997],ri=(1<<26)/e0[e0.length-1],z=function(){function i(e,t,r){e!=null&&(typeof e=="number"?this.fromNumber(e,t,r):t==null&&typeof e!="string"?this.fromString(e,256):this.fromString(e,t))}return i.prototype.toString=function(e){if(this.s<0)return"-"+this.negate().toString(e);var t;if(e==16)t=4;else if(e==8)t=3;else if(e==2)t=1;else if(e==32)t=5;else if(e==4)t=2;else return this.toRadix(e);var r=(1<<t)-1,n,s=!1,x="",o=this.t,f=this.DB-o*this.DB%t;if(o-- >0)for(f<this.DB&&(n=this[o]>>f)>0&&(s=!0,x=v0(n));o>=0;)f<t?(n=(this[o]&(1<<f)-1)<<t-f,n|=this[--o]>>(f+=this.DB-t)):(n=this[o]>>(f-=t)&r,f<=0&&(f+=this.DB,--o)),n>0&&(s=!0),s&&(x+=v0(n));return s?x:"0"},i.prototype.negate=function(){var e=U();return i.ZERO.subTo(this,e),e},i.prototype.abs=function(){return this.s<0?this.negate():this},i.prototype.compareTo=function(e){var t=this.s-e.s;if(t!=0)return t;var r=this.t;if(t=r-e.t,t!=0)return this.s<0?-t:t;for(;--r>=0;)if((t=this[r]-e[r])!=0)return t;return 0},i.prototype.bitLength=function(){return this.t<=0?0:this.DB*(this.t-1)+G0(this[this.t-1]^this.s&this.DM)},i.prototype.mod=function(e){var t=U();return this.abs().divRemTo(e,null,t),this.s<0&&t.compareTo(i.ZERO)>0&&e.subTo(t,t),t},i.prototype.modPowInt=function(e,t){var r;return e<256||t.isEven()?r=new Le(t):r=new Me(t),this.exp(e,r)},i.prototype.clone=function(){var e=U();return this.copyTo(e),e},i.prototype.intValue=function(){if(this.s<0){if(this.t==1)return this[0]-this.DV;if(this.t==0)return-1}else{if(this.t==1)return this[0];if(this.t==0)return 0}return(this[1]&(1<<32-this.DB)-1)<<this.DB|this[0]},i.prototype.byteValue=function(){return this.t==0?this.s:this[0]<<24>>24},i.prototype.shortValue=function(){return this.t==0?this.s:this[0]<<16>>16},i.prototype.signum=function(){return this.s<0?-1:this.t<=0||this.t==1&&this[0]<=0?0:1},i.prototype.toByteArray=function(){var e=this.t,t=[];t[0]=this.s;var r=this.DB-e*this.DB%8,n,s=0;if(e-- >0)for(r<this.DB&&(n=this[e]>>r)!=(this.s&this.DM)>>r&&(t[s++]=n|this.s<<this.DB-r);e>=0;)r<8?(n=(this[e]&(1<<r)-1)<<8-r,n|=this[--e]>>(r+=this.DB-8)):(n=this[e]>>(r-=8)&255,r<=0&&(r+=this.DB,--e)),n&128&&(n|=-256),s==0&&(this.s&128)!=(n&128)&&++s,(s>0||n!=this.s)&&(t[s++]=n);return t},i.prototype.equals=function(e){return this.compareTo(e)==0},i.prototype.min=function(e){return this.compareTo(e)<0?this:e},i.prototype.max=function(e){return this.compareTo(e)>0?this:e},i.prototype.and=function(e){var t=U();return this.bitwiseTo(e,jr,t),t},i.prototype.or=function(e){var t=U();return this.bitwiseTo(e,U0,t),t},i.prototype.xor=function(e){var t=U();return this.bitwiseTo(e,Pe,t),t},i.prototype.andNot=function(e){var t=U();return this.bitwiseTo(e,Oe,t),t},i.prototype.not=function(){for(var e=U(),t=0;t<this.t;++t)e[t]=this.DM&~this[t];return e.t=this.t,e.s=~this.s,e},i.prototype.shiftLeft=function(e){var t=U();return e<0?this.rShiftTo(-e,t):this.lShiftTo(e,t),t},i.prototype.shiftRight=function(e){var t=U();return e<0?this.lShiftTo(-e,t):this.rShiftTo(e,t),t},i.prototype.getLowestSetBit=function(){for(var e=0;e<this.t;++e)if(this[e]!=0)return e*this.DB+Zr(this[e]);return this.s<0?this.t*this.DB:-1},i.prototype.bitCount=function(){for(var e=0,t=this.s&this.DM,r=0;r<this.t;++r)e+=Xr(this[r]^t);return e},i.prototype.testBit=function(e){var t=Math.floor(e/this.DB);return t>=this.t?this.s!=0:(this[t]&1<<e%this.DB)!=0},i.prototype.setBit=function(e){return this.changeBit(e,U0)},i.prototype.clearBit=function(e){return this.changeBit(e,Oe)},i.prototype.flipBit=function(e){return this.changeBit(e,Pe)},i.prototype.add=function(e){var t=U();return this.addTo(e,t),t},i.prototype.subtract=function(e){var t=U();return this.subTo(e,t),t},i.prototype.multiply=function(e){var t=U();return this.multiplyTo(e,t),t},i.prototype.divide=function(e){var t=U();return this.divRemTo(e,t,null),t},i.prototype.remainder=function(e){var t=U();return this.divRemTo(e,null,t),t},i.prototype.divideAndRemainder=function(e){var t=U(),r=U();return this.divRemTo(e,t,r),[t,r]},i.prototype.modPow=function(e,t){var r=e.bitLength(),n,s=B0(1),x;if(r<=0)return s;r<18?n=1:r<48?n=3:r<144?n=4:r<768?n=5:n=6,r<8?x=new Le(t):t.isEven()?x=new ni(t):x=new Me(t);var o=[],f=3,a=n-1,c=(1<<n)-1;if(o[1]=x.convert(this),n>1){var B=U();for(x.sqrTo(o[1],B);f<=c;)o[f]=U(),x.mulTo(B,o[f-2],o[f]),f+=2}var h=e.t-1,u,v=!0,g=U(),l;for(r=G0(e[h])-1;h>=0;){for(r>=a?u=e[h]>>r-a&c:(u=(e[h]&(1<<r+1)-1)<<a-r,h>0&&(u|=e[h-1]>>this.DB+r-a)),f=n;!(u&1);)u>>=1,--f;if((r-=f)<0&&(r+=this.DB,--h),v)o[u].copyTo(s),v=!1;else{for(;f>1;)x.sqrTo(s,g),x.sqrTo(g,s),f-=2;f>0?x.sqrTo(s,g):(l=s,s=g,g=l),x.mulTo(g,o[u],s)}for(;h>=0&&!(e[h]&1<<r);)x.sqrTo(s,g),l=s,s=g,g=l,--r<0&&(r=this.DB-1,--h)}return x.revert(s)},i.prototype.modInverse=function(e){var t=e.isEven();if(this.isEven()&&t||e.signum()==0)return i.ZERO;for(var r=e.clone(),n=this.clone(),s=B0(1),x=B0(0),o=B0(0),f=B0(1);r.signum()!=0;){for(;r.isEven();)r.rShiftTo(1,r),t?((!s.isEven()||!x.isEven())&&(s.addTo(this,s),x.subTo(e,x)),s.rShiftTo(1,s)):x.isEven()||x.subTo(e,x),x.rShiftTo(1,x);for(;n.isEven();)n.rShiftTo(1,n),t?((!o.isEven()||!f.isEven())&&(o.addTo(this,o),f.subTo(e,f)),o.rShiftTo(1,o)):f.isEven()||f.subTo(e,f),f.rShiftTo(1,f);r.compareTo(n)>=0?(r.subTo(n,r),t&&s.subTo(o,s),x.subTo(f,x)):(n.subTo(r,n),t&&o.subTo(s,o),f.subTo(x,f))}if(n.compareTo(i.ONE)!=0)return i.ZERO;if(f.compareTo(e)>=0)return f.subtract(e);if(f.signum()<0)f.addTo(e,f);else return f;return f.signum()<0?f.add(e):f},i.prototype.pow=function(e){return this.exp(e,new ii)},i.prototype.gcd=function(e){var t=this.s<0?this.negate():this.clone(),r=e.s<0?e.negate():e.clone();if(t.compareTo(r)<0){var n=t;t=r,r=n}var s=t.getLowestSetBit(),x=r.getLowestSetBit();if(x<0)return t;for(s<x&&(x=s),x>0&&(t.rShiftTo(x,t),r.rShiftTo(x,r));t.signum()>0;)(s=t.getLowestSetBit())>0&&t.rShiftTo(s,t),(s=r.getLowestSetBit())>0&&r.rShiftTo(s,r),t.compareTo(r)>=0?(t.subTo(r,t),t.rShiftTo(1,t)):(r.subTo(t,r),r.rShiftTo(1,r));return x>0&&r.lShiftTo(x,r),r},i.prototype.isProbablePrime=function(e){var t,r=this.abs();if(r.t==1&&r[0]<=e0[e0.length-1]){for(t=0;t<e0.length;++t)if(r[0]==e0[t])return!0;return!1}if(r.isEven())return!1;for(t=1;t<e0.length;){for(var n=e0[t],s=t+1;s<e0.length&&n<ri;)n*=e0[s++];for(n=r.modInt(n);t<s;)if(n%e0[t++]==0)return!1}return r.millerRabin(e)},i.prototype.copyTo=function(e){for(var t=this.t-1;t>=0;--t)e[t]=this[t];e.t=this.t,e.s=this.s},i.prototype.fromInt=function(e){this.t=1,this.s=e<0?-1:0,e>0?this[0]=e:e<-1?this[0]=e+this.DV:this.t=0},i.prototype.fromString=function(e,t){var r;if(t==16)r=4;else if(t==8)r=3;else if(t==256)r=8;else if(t==2)r=1;else if(t==32)r=5;else if(t==4)r=2;else{this.fromRadix(e,t);return}this.t=0,this.s=0;for(var n=e.length,s=!1,x=0;--n>=0;){var o=r==8?+e[n]&255:Ve(e,n);if(o<0){e.charAt(n)=="-"&&(s=!0);continue}s=!1,x==0?this[this.t++]=o:x+r>this.DB?(this[this.t-1]|=(o&(1<<this.DB-x)-1)<<x,this[this.t++]=o>>this.DB-x):this[this.t-1]|=o<<x,x+=r,x>=this.DB&&(x-=this.DB)}r==8&&+e[0]&128&&(this.s=-1,x>0&&(this[this.t-1]|=(1<<this.DB-x)-1<<x)),this.clamp(),s&&i.ZERO.subTo(this,this)},i.prototype.clamp=function(){for(var e=this.s&this.DM;this.t>0&&this[this.t-1]==e;)--this.t},i.prototype.dlShiftTo=function(e,t){var r;for(r=this.t-1;r>=0;--r)t[r+e]=this[r];for(r=e-1;r>=0;--r)t[r]=0;t.t=this.t+e,t.s=this.s},i.prototype.drShiftTo=function(e,t){for(var r=e;r<this.t;++r)t[r-e]=this[r];t.t=Math.max(this.t-e,0),t.s=this.s},i.prototype.lShiftTo=function(e,t){for(var r=e%this.DB,n=this.DB-r,s=(1<<n)-1,x=Math.floor(e/this.DB),o=this.s<<r&this.DM,f=this.t-1;f>=0;--f)t[f+x+1]=this[f]>>n|o,o=(this[f]&s)<<r;for(var f=x-1;f>=0;--f)t[f]=0;t[x]=o,t.t=this.t+x+1,t.s=this.s,t.clamp()},i.prototype.rShiftTo=function(e,t){t.s=this.s;var r=Math.floor(e/this.DB);if(r>=this.t){t.t=0;return}var n=e%this.DB,s=this.DB-n,x=(1<<n)-1;t[0]=this[r]>>n;for(var o=r+1;o<this.t;++o)t[o-r-1]|=(this[o]&x)<<s,t[o-r]=this[o]>>n;n>0&&(t[this.t-r-1]|=(this.s&x)<<s),t.t=this.t-r,t.clamp()},i.prototype.subTo=function(e,t){for(var r=0,n=0,s=Math.min(e.t,this.t);r<s;)n+=this[r]-e[r],t[r++]=n&this.DM,n>>=this.DB;if(e.t<this.t){for(n-=e.s;r<this.t;)n+=this[r],t[r++]=n&this.DM,n>>=this.DB;n+=this.s}else{for(n+=this.s;r<e.t;)n-=e[r],t[r++]=n&this.DM,n>>=this.DB;n-=e.s}t.s=n<0?-1:0,n<-1?t[r++]=this.DV+n:n>0&&(t[r++]=n),t.t=r,t.clamp()},i.prototype.multiplyTo=function(e,t){var r=this.abs(),n=e.abs(),s=r.t;for(t.t=s+n.t;--s>=0;)t[s]=0;for(s=0;s<n.t;++s)t[s+r.t]=r.am(0,n[s],t,s,0,r.t);t.s=0,t.clamp(),this.s!=e.s&&i.ZERO.subTo(t,t)},i.prototype.squareTo=function(e){for(var t=this.abs(),r=e.t=2*t.t;--r>=0;)e[r]=0;for(r=0;r<t.t-1;++r){var n=t.am(r,t[r],e,2*r,0,1);(e[r+t.t]+=t.am(r+1,2*t[r],e,2*r+1,n,t.t-r-1))>=t.DV&&(e[r+t.t]-=t.DV,e[r+t.t+1]=1)}e.t>0&&(e[e.t-1]+=t.am(r,t[r],e,2*r,0,1)),e.s=0,e.clamp()},i.prototype.divRemTo=function(e,t,r){var n=e.abs();if(!(n.t<=0)){var s=this.abs();if(s.t<n.t){t!=null&&t.fromInt(0),r!=null&&this.copyTo(r);return}r==null&&(r=U());var x=U(),o=this.s,f=e.s,a=this.DB-G0(n[n.t-1]);a>0?(n.lShiftTo(a,x),s.lShiftTo(a,r)):(n.copyTo(x),s.copyTo(r));var c=x.t,B=x[c-1];if(B!=0){var h=B*(1<<this.F1)+(c>1?x[c-2]>>this.F2:0),u=this.FV/h,v=(1<<this.F1)/h,g=1<<this.F2,l=r.t,A=l-c,d=t??U();for(x.dlShiftTo(A,d),r.compareTo(d)>=0&&(r[r.t++]=1,r.subTo(d,r)),i.ONE.dlShiftTo(c,d),d.subTo(x,x);x.t<c;)x[x.t++]=0;for(;--A>=0;){var p=r[--l]==B?this.DM:Math.floor(r[l]*u+(r[l-1]+g)*v);if((r[l]+=x.am(0,p,r,A,0,c))<p)for(x.dlShiftTo(A,d),r.subTo(d,r);r[l]<--p;)r.subTo(d,r)}t!=null&&(r.drShiftTo(c,t),o!=f&&i.ZERO.subTo(t,t)),r.t=c,r.clamp(),a>0&&r.rShiftTo(a,r),o<0&&i.ZERO.subTo(r,r)}}},i.prototype.invDigit=function(){if(this.t<1)return 0;var e=this[0];if(!(e&1))return 0;var t=e&3;return t=t*(2-(e&15)*t)&15,t=t*(2-(e&255)*t)&255,t=t*(2-((e&65535)*t&65535))&65535,t=t*(2-e*t%this.DV)%this.DV,t>0?this.DV-t:-t},i.prototype.isEven=function(){return(this.t>0?this[0]&1:this.s)==0},i.prototype.exp=function(e,t){if(e>4294967295||e<1)return i.ONE;var r=U(),n=U(),s=t.convert(this),x=G0(e)-1;for(s.copyTo(r);--x>=0;)if(t.sqrTo(r,n),(e&1<<x)>0)t.mulTo(n,s,r);else{var o=r;r=n,n=o}return t.revert(r)},i.prototype.chunkSize=function(e){return Math.floor(Math.LN2*this.DB/Math.log(e))},i.prototype.toRadix=function(e){if(e==null&&(e=10),this.signum()==0||e<2||e>36)return"0";var t=this.chunkSize(e),r=Math.pow(e,t),n=B0(r),s=U(),x=U(),o="";for(this.divRemTo(n,s,x);s.signum()>0;)o=(r+x.intValue()).toString(e).substr(1)+o,s.divRemTo(n,s,x);return x.intValue().toString(e)+o},i.prototype.fromRadix=function(e,t){this.fromInt(0),t==null&&(t=10);for(var r=this.chunkSize(t),n=Math.pow(t,r),s=!1,x=0,o=0,f=0;f<e.length;++f){var a=Ve(e,f);if(a<0){e.charAt(f)=="-"&&this.signum()==0&&(s=!0);continue}o=t*o+a,++x>=r&&(this.dMultiply(n),this.dAddOffset(o,0),x=0,o=0)}x>0&&(this.dMultiply(Math.pow(t,x)),this.dAddOffset(o,0)),s&&i.ZERO.subTo(this,this)},i.prototype.fromNumber=function(e,t,r){if(typeof t=="number")if(e<2)this.fromInt(1);else for(this.fromNumber(e,r),this.testBit(e-1)||this.bitwiseTo(i.ONE.shiftLeft(e-1),U0,this),this.isEven()&&this.dAddOffset(1,0);!this.isProbablePrime(t);)this.dAddOffset(2,0),this.bitLength()>e&&this.subTo(i.ONE.shiftLeft(e-1),this);else{var n=[],s=e&7;n.length=(e>>3)+1,t.nextBytes(n),s>0?n[0]&=(1<<s)-1:n[0]=0,this.fromString(n,256)}},i.prototype.bitwiseTo=function(e,t,r){var n,s,x=Math.min(e.t,this.t);for(n=0;n<x;++n)r[n]=t(this[n],e[n]);if(e.t<this.t){for(s=e.s&this.DM,n=x;n<this.t;++n)r[n]=t(this[n],s);r.t=this.t}else{for(s=this.s&this.DM,n=x;n<e.t;++n)r[n]=t(s,e[n]);r.t=e.t}r.s=t(this.s,e.s),r.clamp()},i.prototype.changeBit=function(e,t){var r=i.ONE.shiftLeft(e);return this.bitwiseTo(r,t,r),r},i.prototype.addTo=function(e,t){for(var r=0,n=0,s=Math.min(e.t,this.t);r<s;)n+=this[r]+e[r],t[r++]=n&this.DM,n>>=this.DB;if(e.t<this.t){for(n+=e.s;r<this.t;)n+=this[r],t[r++]=n&this.DM,n>>=this.DB;n+=this.s}else{for(n+=this.s;r<e.t;)n+=e[r],t[r++]=n&this.DM,n>>=this.DB;n+=e.s}t.s=n<0?-1:0,n>0?t[r++]=n:n<-1&&(t[r++]=this.DV+n),t.t=r,t.clamp()},i.prototype.dMultiply=function(e){this[this.t]=this.am(0,e-1,this,0,0,this.t),++this.t,this.clamp()},i.prototype.dAddOffset=function(e,t){if(e!=0){for(;this.t<=t;)this[this.t++]=0;for(this[t]+=e;this[t]>=this.DV;)this[t]-=this.DV,++t>=this.t&&(this[this.t++]=0),++this[t]}},i.prototype.multiplyLowerTo=function(e,t,r){var n=Math.min(this.t+e.t,t);for(r.s=0,r.t=n;n>0;)r[--n]=0;for(var s=r.t-this.t;n<s;++n)r[n+this.t]=this.am(0,e[n],r,n,0,this.t);for(var s=Math.min(e.t,t);n<s;++n)this.am(0,e[n],r,n,0,t-n);r.clamp()},i.prototype.multiplyUpperTo=function(e,t,r){--t;var n=r.t=this.t+e.t-t;for(r.s=0;--n>=0;)r[n]=0;for(n=Math.max(t-this.t,0);n<e.t;++n)r[this.t+n-t]=this.am(t-n,e[n],r,0,0,this.t+n-t);r.clamp(),r.drShiftTo(1,r)},i.prototype.modInt=function(e){if(e<=0)return 0;var t=this.DV%e,r=this.s<0?e-1:0;if(this.t>0)if(t==0)r=this[0]%e;else for(var n=this.t-1;n>=0;--n)r=(t*r+this[n])%e;return r},i.prototype.millerRabin=function(e){var t=this.subtract(i.ONE),r=t.getLowestSetBit();if(r<=0)return!1;var n=t.shiftRight(r);e=e+1>>1,e>e0.length&&(e=e0.length);for(var s=U(),x=0;x<e;++x){s.fromInt(e0[Math.floor(Math.random()*e0.length)]);var o=s.modPow(n,this);if(o.compareTo(i.ONE)!=0&&o.compareTo(t)!=0){for(var f=1;f++<r&&o.compareTo(t)!=0;)if(o=o.modPowInt(2,this),o.compareTo(i.ONE)==0)return!1;if(o.compareTo(t)!=0)return!1}}return!0},i.prototype.square=function(){var e=U();return this.squareTo(e),e},i.prototype.gcda=function(e,t){var r=this.s<0?this.negate():this.clone(),n=e.s<0?e.negate():e.clone();if(r.compareTo(n)<0){var s=r;r=n,n=s}var x=r.getLowestSetBit(),o=n.getLowestSetBit();if(o<0){t(r);return}x<o&&(o=x),o>0&&(r.rShiftTo(o,r),n.rShiftTo(o,n));var f=function(){(x=r.getLowestSetBit())>0&&r.rShiftTo(x,r),(x=n.getLowestSetBit())>0&&n.rShiftTo(x,n),r.compareTo(n)>=0?(r.subTo(n,r),r.rShiftTo(1,r)):(n.subTo(r,n),n.rShiftTo(1,n)),r.signum()>0?setTimeout(f,0):(o>0&&n.lShiftTo(o,n),setTimeout(function(){t(n)},0))};setTimeout(f,10)},i.prototype.fromNumberAsync=function(e,t,r,n){if(typeof t=="number")if(e<2)this.fromInt(1);else{this.fromNumber(e,r),this.testBit(e-1)||this.bitwiseTo(i.ONE.shiftLeft(e-1),U0,this),this.isEven()&&this.dAddOffset(1,0);var s=this,x=function(){s.dAddOffset(2,0),s.bitLength()>e&&s.subTo(i.ONE.shiftLeft(e-1),s),s.isProbablePrime(t)?setTimeout(function(){n()},0):setTimeout(x,0)};setTimeout(x,0)}else{var o=[],f=e&7;o.length=(e>>3)+1,t.nextBytes(o),f>0?o[0]&=(1<<f)-1:o[0]=0,this.fromString(o,256)}},i}(),ii=function(){function i(){}return i.prototype.convert=function(e){return e},i.prototype.revert=function(e){return e},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r)},i.prototype.sqrTo=function(e,t){e.squareTo(t)},i}(),Le=function(){function i(e){this.m=e}return i.prototype.convert=function(e){return e.s<0||e.compareTo(this.m)>=0?e.mod(this.m):e},i.prototype.revert=function(e){return e},i.prototype.reduce=function(e){e.divRemTo(this.m,null,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}(),Me=function(){function i(e){this.m=e,this.mp=e.invDigit(),this.mpl=this.mp&32767,this.mph=this.mp>>15,this.um=(1<<e.DB-15)-1,this.mt2=2*e.t}return i.prototype.convert=function(e){var t=U();return e.abs().dlShiftTo(this.m.t,t),t.divRemTo(this.m,null,t),e.s<0&&t.compareTo(z.ZERO)>0&&this.m.subTo(t,t),t},i.prototype.revert=function(e){var t=U();return e.copyTo(t),this.reduce(t),t},i.prototype.reduce=function(e){for(;e.t<=this.mt2;)e[e.t++]=0;for(var t=0;t<this.m.t;++t){var r=e[t]&32767,n=r*this.mpl+((r*this.mph+(e[t]>>15)*this.mpl&this.um)<<15)&e.DM;for(r=t+this.m.t,e[r]+=this.m.am(0,n,e,t,0,this.m.t);e[r]>=e.DV;)e[r]-=e.DV,e[++r]++}e.clamp(),e.drShiftTo(this.m.t,e),e.compareTo(this.m)>=0&&e.subTo(this.m,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}(),ni=function(){function i(e){this.m=e,this.r2=U(),this.q3=U(),z.ONE.dlShiftTo(2*e.t,this.r2),this.mu=this.r2.divide(e)}return i.prototype.convert=function(e){if(e.s<0||e.t>2*this.m.t)return e.mod(this.m);if(e.compareTo(this.m)<0)return e;var t=U();return e.copyTo(t),this.reduce(t),t},i.prototype.revert=function(e){return e},i.prototype.reduce=function(e){for(e.drShiftTo(this.m.t-1,this.r2),e.t>this.m.t+1&&(e.t=this.m.t+1,e.clamp()),this.mu.multiplyUpperTo(this.r2,this.m.t+1,this.q3),this.m.multiplyLowerTo(this.q3,this.m.t+1,this.r2);e.compareTo(this.r2)<0;)e.dAddOffset(1,this.m.t+1);for(e.subTo(this.r2,e);e.compareTo(this.m)>=0;)e.subTo(this.m,e)},i.prototype.mulTo=function(e,t,r){e.multiplyTo(t,r),this.reduce(r)},i.prototype.sqrTo=function(e,t){e.squareTo(t),this.reduce(t)},i}();function U(){return new z(null)}function Q(i,e){return new z(i,e)}var Ue=typeof navigator<"u";Ue&&We&&navigator.appName=="Microsoft Internet Explorer"?(z.prototype.am=function(e,t,r,n,s,x){for(var o=t&32767,f=t>>15;--x>=0;){var a=this[e]&32767,c=this[e++]>>15,B=f*a+c*o;a=o*a+((B&32767)<<15)+r[n]+(s&1073741823),s=(a>>>30)+(B>>>15)+f*c+(s>>>30),r[n++]=a&1073741823}return s},E0=30):Ue&&We&&navigator.appName!="Netscape"?(z.prototype.am=function(e,t,r,n,s,x){for(;--x>=0;){var o=t*this[e++]+r[n]+s;s=Math.floor(o/67108864),r[n++]=o&67108863}return s},E0=26):(z.prototype.am=function(e,t,r,n,s,x){for(var o=t&16383,f=t>>14;--x>=0;){var a=this[e]&16383,c=this[e++]>>14,B=f*a+c*o;a=o*a+((B&16383)<<14)+r[n]+s,s=(a>>28)+(B>>14)+f*c,r[n++]=a&268435455}return s},E0=28),z.prototype.DB=E0,z.prototype.DM=(1<<E0)-1,z.prototype.DV=1<<E0;var Mt=52;z.prototype.FV=Math.pow(2,Mt),z.prototype.F1=Mt-E0,z.prototype.F2=2*E0-Mt;var K0=[],T0,f0;for(T0=48,f0=0;f0<=9;++f0)K0[T0++]=f0;for(T0=97,f0=10;f0<36;++f0)K0[T0++]=f0;for(T0=65,f0=10;f0<36;++f0)K0[T0++]=f0;function Ve(i,e){var t=K0[i.charCodeAt(e)];return t??-1}function B0(i){var e=U();return e.fromInt(i),e}function G0(i){var e=1,t;return(t=i>>>16)!=0&&(i=t,e+=16),(t=i>>8)!=0&&(i=t,e+=8),(t=i>>4)!=0&&(i=t,e+=4),(t=i>>2)!=0&&(i=t,e+=2),(t=i>>1)!=0&&(i=t,e+=1),e}z.ZERO=B0(0),z.ONE=B0(1);var si=function(){function i(){this.i=0,this.j=0,this.S=[]}return i.prototype.init=function(e){var t,r,n;for(t=0;t<256;++t)this.S[t]=t;for(r=0,t=0;t<256;++t)r=r+this.S[t]+e[t%e.length]&255,n=this.S[t],this.S[t]=this.S[r],this.S[r]=n;this.i=0,this.j=0},i.prototype.next=function(){var e;return this.i=this.i+1&255,this.j=this.j+this.S[this.i]&255,e=this.S[this.i],this.S[this.i]=this.S[this.j],this.S[this.j]=e,this.S[e+this.S[this.i]&255]},i}();function oi(){return new si}var Ke=256,$0,C0=null,h0;if(C0==null){C0=[],h0=0;var j0=void 0;if(typeof window<"u"&&window.crypto&&window.crypto.getRandomValues){var Ut=new Uint32Array(256);for(window.crypto.getRandomValues(Ut),j0=0;j0<Ut.length;++j0)C0[h0++]=Ut[j0]&255}var Z0=0,X0=function(i){if(Z0=Z0||0,Z0>=256||h0>=Ke){window.removeEventListener?window.removeEventListener("mousemove",X0,!1):window.detachEvent&&window.detachEvent("onmousemove",X0);return}try{var e=i.x+i.y;C0[h0++]=e&255,Z0+=1}catch{}};typeof window<"u"&&(window.addEventListener?window.addEventListener("mousemove",X0,!1):window.attachEvent&&window.attachEvent("onmousemove",X0))}function ai(){if($0==null){for($0=oi();h0<Ke;){var i=Math.floor(65536*Math.random());C0[h0++]=i&255}for($0.init(C0),h0=0;h0<C0.length;++h0)C0[h0]=0;h0=0}return $0.next()}var Vt=function(){function i(){}return i.prototype.nextBytes=function(e){for(var t=0;t<e.length;++t)e[t]=ai()},i}();function xi(i,e){if(e<i.length+22)return console.error("Message too long for RSA"),null;for(var t=e-i.length-6,r="",n=0;n<t;n+=2)r+="ff";var s="0001"+r+"00"+i;return Q(s,16)}function fi(i,e){if(e<i.length+11)return console.error("Message too long for RSA"),null;for(var t=[],r=i.length-1;r>=0&&e>0;){var n=i.charCodeAt(r--);n<128?t[--e]=n:n>127&&n<2048?(t[--e]=n&63|128,t[--e]=n>>6|192):(t[--e]=n&63|128,t[--e]=n>>6&63|128,t[--e]=n>>12|224)}t[--e]=0;for(var s=new Vt,x=[];e>2;){for(x[0]=0;x[0]==0;)s.nextBytes(x);t[--e]=x[0]}return t[--e]=2,t[--e]=0,new z(t)}var ci=function(){function i(){this.n=null,this.e=0,this.d=null,this.p=null,this.q=null,this.dmp1=null,this.dmq1=null,this.coeff=null}return i.prototype.doPublic=function(e){return e.modPowInt(this.e,this.n)},i.prototype.doPrivate=function(e){if(this.p==null||this.q==null)return e.modPow(this.d,this.n);for(var t=e.mod(this.p).modPow(this.dmp1,this.p),r=e.mod(this.q).modPow(this.dmq1,this.q);t.compareTo(r)<0;)t=t.add(this.p);return t.subtract(r).multiply(this.coeff).mod(this.p).multiply(this.q).add(r)},i.prototype.setPublic=function(e,t){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16)):console.error("Invalid RSA public key")},i.prototype.encrypt=function(e){var t=this.n.bitLength()+7>>3,r=fi(e,t);if(r==null)return null;var n=this.doPublic(r);if(n==null)return null;for(var s=n.toString(16),x=s.length,o=0;o<t*2-x;o++)s="0"+s;return s},i.prototype.setPrivate=function(e,t,r){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16),this.d=Q(r,16)):console.error("Invalid RSA private key")},i.prototype.setPrivateEx=function(e,t,r,n,s,x,o,f){e!=null&&t!=null&&e.length>0&&t.length>0?(this.n=Q(e,16),this.e=parseInt(t,16),this.d=Q(r,16),this.p=Q(n,16),this.q=Q(s,16),this.dmp1=Q(x,16),this.dmq1=Q(o,16),this.coeff=Q(f,16)):console.error("Invalid RSA private key")},i.prototype.generate=function(e,t){var r=new Vt,n=e>>1;this.e=parseInt(t,16);for(var s=new z(t,16);;){for(;this.p=new z(e-n,1,r),!(this.p.subtract(z.ONE).gcd(s).compareTo(z.ONE)==0&&this.p.isProbablePrime(10)););for(;this.q=new z(n,1,r),!(this.q.subtract(z.ONE).gcd(s).compareTo(z.ONE)==0&&this.q.isProbablePrime(10)););if(this.p.compareTo(this.q)<=0){var x=this.p;this.p=this.q,this.q=x}var o=this.p.subtract(z.ONE),f=this.q.subtract(z.ONE),a=o.multiply(f);if(a.gcd(s).compareTo(z.ONE)==0){this.n=this.p.multiply(this.q),this.d=s.modInverse(a),this.dmp1=this.d.mod(o),this.dmq1=this.d.mod(f),this.coeff=this.q.modInverse(this.p);break}}},i.prototype.decrypt=function(e){var t=Q(e,16),r=this.doPrivate(t);return r==null?null:hi(r,this.n.bitLength()+7>>3)},i.prototype.generateAsync=function(e,t,r){var n=new Vt,s=e>>1;this.e=parseInt(t,16);var x=new z(t,16),o=this,f=function(){var a=function(){if(o.p.compareTo(o.q)<=0){var h=o.p;o.p=o.q,o.q=h}var u=o.p.subtract(z.ONE),v=o.q.subtract(z.ONE),g=u.multiply(v);g.gcd(x).compareTo(z.ONE)==0?(o.n=o.p.multiply(o.q),o.d=x.modInverse(g),o.dmp1=o.d.mod(u),o.dmq1=o.d.mod(v),o.coeff=o.q.modInverse(o.p),setTimeout(function(){r()},0)):setTimeout(f,0)},c=function(){o.q=U(),o.q.fromNumberAsync(s,1,n,function(){o.q.subtract(z.ONE).gcda(x,function(h){h.compareTo(z.ONE)==0&&o.q.isProbablePrime(10)?setTimeout(a,0):setTimeout(c,0)})})},B=function(){o.p=U(),o.p.fromNumberAsync(e-s,1,n,function(){o.p.subtract(z.ONE).gcda(x,function(h){h.compareTo(z.ONE)==0&&o.p.isProbablePrime(10)?setTimeout(c,0):setTimeout(B,0)})})};setTimeout(B,0)};setTimeout(f,0)},i.prototype.sign=function(e,t,r){var n=ui(r),s=n+t(e).toString(),x=xi(s,this.n.bitLength()/4);if(x==null)return null;var o=this.doPrivate(x);if(o==null)return null;var f=o.toString(16);return f.length&1?"0"+f:f},i.prototype.verify=function(e,t,r){var n=Q(t,16),s=this.doPublic(n);if(s==null)return null;var x=s.toString(16).replace(/^1f+00/,""),o=li(x);return o==r(e).toString()},i}();function hi(i,e){for(var t=i.toByteArray(),r=0;r<t.length&&t[r]==0;)++r;if(t.length-r!=e-1||t[r]!=2)return null;for(++r;t[r]!=0;)if(++r>=t.length)return null;for(var n="";++r<t.length;){var s=t[r]&255;s<128?n+=String.fromCharCode(s):s>191&&s<224?(n+=String.fromCharCode((s&31)<<6|t[r+1]&63),++r):(n+=String.fromCharCode((s&15)<<12|(t[r+1]&63)<<6|t[r+2]&63),r+=2)}return n}var Q0={md2:"3020300c06082a864886f70d020205000410",md5:"3020300c06082a864886f70d020505000410",sha1:"3021300906052b0e03021a05000414",sha224:"302d300d06096086480165030402040500041c",sha256:"3031300d060960864801650304020105000420",sha384:"3041300d060960864801650304020205000430",sha512:"3051300d060960864801650304020305000440",ripemd160:"3021300906052b2403020105000414"};function ui(i){return Q0[i]||""}function li(i){for(var e in Q0)if(Q0.hasOwnProperty(e)){var t=Q0[e],r=t.length;if(i.substr(0,r)==t)return i.substr(r)}return i}/*!
 Copyright (c) 2011, Yahoo! Inc. All rights reserved.
 Code licensed under the BSD License:
 http://developer.yahoo.com/yui/license.html
@@ -30,10 +30,10 @@ version: 2.9.0
  * @version asn1 1.0.13 (2017-Jun-02)
  * @since jsrsasign 2.1
  * @license <a href="https://kjur.github.io/jsrsasign/license/">MIT License</a>
- */var m={};(typeof m.asn1>"u"||!m.asn1)&&(m.asn1={}),m.asn1.ASN1Util=new function(){this.integerToByteHex=function(i){var e=i.toString(16);return e.length%2==1&&(e="0"+e),e},this.bigIntToMinTwosComplementsHex=function(i){var e=i.toString(16);if(e.substr(0,1)!="-")e.length%2==1?e="0"+e:e.match(/^[0-7]/)||(e="00"+e);else{var t=e.substr(1),r=t.length;r%2==1?r+=1:e.match(/^[0-7]/)||(r+=2);for(var n="",s=0;s<r;s++)n+="f";var x=new z(n,16),o=x.xor(i).add(z.ONE);e=o.toString(16).replace(/^-/,"")}return e},this.getPEMStringFromHex=function(i,e){return hextopem(i,e)},this.newObject=function(i){var e=m,t=e.asn1,r=t.DERBoolean,n=t.DERInteger,s=t.DERBitString,x=t.DEROctetString,o=t.DERNull,f=t.DERObjectIdentifier,a=t.DEREnumerated,c=t.DERUTF8String,E=t.DERNumericString,h=t.DERPrintableString,u=t.DERTeletexString,v=t.DERIA5String,p=t.DERUTCTime,l=t.DERGeneralizedTime,C=t.DERSequence,d=t.DERSet,g=t.DERTaggedObject,B=t.ASN1Util.newObject,F=Object.keys(i);if(F.length!=1)throw"key of param shall be only one.";var A=F[0];if(":bool:int:bitstr:octstr:null:oid:enum:utf8str:numstr:prnstr:telstr:ia5str:utctime:gentime:seq:set:tag:".indexOf(":"+A+":")==-1)throw"undefined key: "+A;if(A=="bool")return new r(i[A]);if(A=="int")return new n(i[A]);if(A=="bitstr")return new s(i[A]);if(A=="octstr")return new x(i[A]);if(A=="null")return new o(i[A]);if(A=="oid")return new f(i[A]);if(A=="enum")return new a(i[A]);if(A=="utf8str")return new c(i[A]);if(A=="numstr")return new E(i[A]);if(A=="prnstr")return new h(i[A]);if(A=="telstr")return new u(i[A]);if(A=="ia5str")return new v(i[A]);if(A=="utctime")return new p(i[A]);if(A=="gentime")return new l(i[A]);if(A=="seq"){for(var _=i[A],b=[],P=0;P<_.length;P++){var D=B(_[P]);b.push(D)}return new C({array:b})}if(A=="set"){for(var _=i[A],b=[],P=0;P<_.length;P++){var D=B(_[P]);b.push(D)}return new d({array:b})}if(A=="tag"){var y=i[A];if(Object.prototype.toString.call(y)==="[object Array]"&&y.length==3){var T=B(y[2]);return new g({tag:y[0],explicit:y[1],obj:T})}else{var w={};if(y.explicit!==void 0&&(w.explicit=y.explicit),y.tag!==void 0&&(w.tag=y.tag),y.obj===void 0)throw"obj shall be specified for 'tag'.";return w.obj=B(y.obj),new g(w)}}},this.jsonToASN1HEX=function(i){var e=this.newObject(i);return e.getEncodedHex()}},m.asn1.ASN1Util.oidHexToInt=function(i){for(var n="",e=parseInt(i.substr(0,2),16),t=Math.floor(e/40),r=e%40,n=t+"."+r,s="",x=2;x<i.length;x+=2){var o=parseInt(i.substr(x,2),16),f=("00000000"+o.toString(2)).slice(-8);if(s=s+f.substr(1,7),f.substr(0,1)=="0"){var a=new z(s,2);n=n+"."+a.toString(10),s=""}}return n},m.asn1.ASN1Util.oidIntToHex=function(i){var e=function(o){var f=o.toString(16);return f.length==1&&(f="0"+f),f},t=function(o){var f="",a=new z(o,10),c=a.toString(2),E=7-c.length%7;E==7&&(E=0);for(var h="",u=0;u<E;u++)h+="0";c=h+c;for(var u=0;u<c.length-1;u+=7){var v=c.substr(u,7);u!=c.length-7&&(v="1"+v),f+=e(parseInt(v,2))}return f};if(!i.match(/^[0-9.]+$/))throw"malformed oid string: "+i;var r="",n=i.split("."),s=parseInt(n[0])*40+parseInt(n[1]);r+=e(s),n.splice(0,2);for(var x=0;x<n.length;x++)r+=t(n[x]);return r},m.asn1.ASN1Object=function(){var i="";this.getLengthHexFromValue=function(){if(typeof this.hV>"u"||this.hV==null)throw"this.hV is null or undefined.";if(this.hV.length%2==1)throw"value hex must be even length: n="+i.length+",v="+this.hV;var e=this.hV.length/2,t=e.toString(16);if(t.length%2==1&&(t="0"+t),e<128)return t;var r=t.length/2;if(r>15)throw"ASN.1 length too long to represent by 8x: n = "+e.toString(16);var n=128+r;return n.toString(16)+t},this.getEncodedHex=function(){return(this.hTLV==null||this.isModified)&&(this.hV=this.getFreshValueHex(),this.hL=this.getLengthHexFromValue(),this.hTLV=this.hT+this.hL+this.hV,this.isModified=!1),this.hTLV},this.getValueHex=function(){return this.getEncodedHex(),this.hV},this.getFreshValueHex=function(){return""}},m.asn1.DERAbstractString=function(i){m.asn1.DERAbstractString.superclass.constructor.call(this),this.getString=function(){return this.s},this.setString=function(e){this.hTLV=null,this.isModified=!0,this.s=e,this.hV=stohex(this.s)},this.setStringHex=function(e){this.hTLV=null,this.isModified=!0,this.s=null,this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i=="string"?this.setString(i):typeof i.str<"u"?this.setString(i.str):typeof i.hex<"u"&&this.setStringHex(i.hex))},Y.lang.extend(m.asn1.DERAbstractString,m.asn1.ASN1Object),m.asn1.DERAbstractTime=function(i){m.asn1.DERAbstractTime.superclass.constructor.call(this),this.localDateToUTC=function(e){utc=e.getTime()+e.getTimezoneOffset()*6e4;var t=new Date(utc);return t},this.formatDate=function(e,t,r){var n=this.zeroPadding,s=this.localDateToUTC(e),x=String(s.getFullYear());t=="utc"&&(x=x.substr(2,2));var o=n(String(s.getMonth()+1),2),f=n(String(s.getDate()),2),a=n(String(s.getHours()),2),c=n(String(s.getMinutes()),2),E=n(String(s.getSeconds()),2),h=x+o+f+a+c+E;if(r===!0){var u=s.getMilliseconds();if(u!=0){var v=n(String(u),3);v=v.replace(/[0]+$/,""),h=h+"."+v}}return h+"Z"},this.zeroPadding=function(e,t){return e.length>=t?e:new Array(t-e.length+1).join("0")+e},this.getString=function(){return this.s},this.setString=function(e){this.hTLV=null,this.isModified=!0,this.s=e,this.hV=stohex(e)},this.setByDateValue=function(e,t,r,n,s,x){var o=new Date(Date.UTC(e,t-1,r,n,s,x,0));this.setByDate(o)},this.getFreshValueHex=function(){return this.hV}},Y.lang.extend(m.asn1.DERAbstractTime,m.asn1.ASN1Object),m.asn1.DERAbstractStructured=function(i){m.asn1.DERAbstractString.superclass.constructor.call(this),this.setByASN1ObjectArray=function(e){this.hTLV=null,this.isModified=!0,this.asn1Array=e},this.appendASN1Object=function(e){this.hTLV=null,this.isModified=!0,this.asn1Array.push(e)},this.asn1Array=new Array,typeof i<"u"&&typeof i.array<"u"&&(this.asn1Array=i.array)},Y.lang.extend(m.asn1.DERAbstractStructured,m.asn1.ASN1Object),m.asn1.DERBoolean=function(){m.asn1.DERBoolean.superclass.constructor.call(this),this.hT="01",this.hTLV="0101ff"},Y.lang.extend(m.asn1.DERBoolean,m.asn1.ASN1Object),m.asn1.DERInteger=function(i){m.asn1.DERInteger.superclass.constructor.call(this),this.hT="02",this.setByBigInteger=function(e){this.hTLV=null,this.isModified=!0,this.hV=m.asn1.ASN1Util.bigIntToMinTwosComplementsHex(e)},this.setByInteger=function(e){var t=new z(String(e),10);this.setByBigInteger(t)},this.setValueHex=function(e){this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.bigint<"u"?this.setByBigInteger(i.bigint):typeof i.int<"u"?this.setByInteger(i.int):typeof i=="number"?this.setByInteger(i):typeof i.hex<"u"&&this.setValueHex(i.hex))},Y.lang.extend(m.asn1.DERInteger,m.asn1.ASN1Object),m.asn1.DERBitString=function(i){if(i!==void 0&&typeof i.obj<"u"){var e=m.asn1.ASN1Util.newObject(i.obj);i.hex="00"+e.getEncodedHex()}m.asn1.DERBitString.superclass.constructor.call(this),this.hT="03",this.setHexValueIncludingUnusedBits=function(t){this.hTLV=null,this.isModified=!0,this.hV=t},this.setUnusedBitsAndHexValue=function(t,r){if(t<0||7<t)throw"unused bits shall be from 0 to 7: u = "+t;var n="0"+t;this.hTLV=null,this.isModified=!0,this.hV=n+r},this.setByBinaryString=function(t){t=t.replace(/0+$/,"");var r=8-t.length%8;r==8&&(r=0);for(var n=0;n<=r;n++)t+="0";for(var s="",n=0;n<t.length-1;n+=8){var x=t.substr(n,8),o=parseInt(x,2).toString(16);o.length==1&&(o="0"+o),s+=o}this.hTLV=null,this.isModified=!0,this.hV="0"+r+s},this.setByBooleanArray=function(t){for(var r="",n=0;n<t.length;n++)t[n]==!0?r+="1":r+="0";this.setByBinaryString(r)},this.newFalseArray=function(t){for(var r=new Array(t),n=0;n<t;n++)r[n]=!1;return r},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i=="string"&&i.toLowerCase().match(/^[0-9a-f]+$/)?this.setHexValueIncludingUnusedBits(i):typeof i.hex<"u"?this.setHexValueIncludingUnusedBits(i.hex):typeof i.bin<"u"?this.setByBinaryString(i.bin):typeof i.array<"u"&&this.setByBooleanArray(i.array))},Y.lang.extend(m.asn1.DERBitString,m.asn1.ASN1Object),m.asn1.DEROctetString=function(i){if(i!==void 0&&typeof i.obj<"u"){var e=m.asn1.ASN1Util.newObject(i.obj);i.hex=e.getEncodedHex()}m.asn1.DEROctetString.superclass.constructor.call(this,i),this.hT="04"},Y.lang.extend(m.asn1.DEROctetString,m.asn1.DERAbstractString),m.asn1.DERNull=function(){m.asn1.DERNull.superclass.constructor.call(this),this.hT="05",this.hTLV="0500"},Y.lang.extend(m.asn1.DERNull,m.asn1.ASN1Object),m.asn1.DERObjectIdentifier=function(i){var e=function(r){var n=r.toString(16);return n.length==1&&(n="0"+n),n},t=function(r){var n="",s=new z(r,10),x=s.toString(2),o=7-x.length%7;o==7&&(o=0);for(var f="",a=0;a<o;a++)f+="0";x=f+x;for(var a=0;a<x.length-1;a+=7){var c=x.substr(a,7);a!=x.length-7&&(c="1"+c),n+=e(parseInt(c,2))}return n};m.asn1.DERObjectIdentifier.superclass.constructor.call(this),this.hT="06",this.setValueHex=function(r){this.hTLV=null,this.isModified=!0,this.s=null,this.hV=r},this.setValueOidString=function(r){if(!r.match(/^[0-9.]+$/))throw"malformed oid string: "+r;var n="",s=r.split("."),x=parseInt(s[0])*40+parseInt(s[1]);n+=e(x),s.splice(0,2);for(var o=0;o<s.length;o++)n+=t(s[o]);this.hTLV=null,this.isModified=!0,this.s=null,this.hV=n},this.setValueName=function(r){var n=m.asn1.x509.OID.name2oid(r);if(n!=="")this.setValueOidString(n);else throw"DERObjectIdentifier oidName undefined: "+r},this.getFreshValueHex=function(){return this.hV},i!==void 0&&(typeof i=="string"?i.match(/^[0-2].[0-9.]+$/)?this.setValueOidString(i):this.setValueName(i):i.oid!==void 0?this.setValueOidString(i.oid):i.hex!==void 0?this.setValueHex(i.hex):i.name!==void 0&&this.setValueName(i.name))},Y.lang.extend(m.asn1.DERObjectIdentifier,m.asn1.ASN1Object),m.asn1.DEREnumerated=function(i){m.asn1.DEREnumerated.superclass.constructor.call(this),this.hT="0a",this.setByBigInteger=function(e){this.hTLV=null,this.isModified=!0,this.hV=m.asn1.ASN1Util.bigIntToMinTwosComplementsHex(e)},this.setByInteger=function(e){var t=new z(String(e),10);this.setByBigInteger(t)},this.setValueHex=function(e){this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.int<"u"?this.setByInteger(i.int):typeof i=="number"?this.setByInteger(i):typeof i.hex<"u"&&this.setValueHex(i.hex))},Y.lang.extend(m.asn1.DEREnumerated,m.asn1.ASN1Object),m.asn1.DERUTF8String=function(i){m.asn1.DERUTF8String.superclass.constructor.call(this,i),this.hT="0c"},Y.lang.extend(m.asn1.DERUTF8String,m.asn1.DERAbstractString),m.asn1.DERNumericString=function(i){m.asn1.DERNumericString.superclass.constructor.call(this,i),this.hT="12"},Y.lang.extend(m.asn1.DERNumericString,m.asn1.DERAbstractString),m.asn1.DERPrintableString=function(i){m.asn1.DERPrintableString.superclass.constructor.call(this,i),this.hT="13"},Y.lang.extend(m.asn1.DERPrintableString,m.asn1.DERAbstractString),m.asn1.DERTeletexString=function(i){m.asn1.DERTeletexString.superclass.constructor.call(this,i),this.hT="14"},Y.lang.extend(m.asn1.DERTeletexString,m.asn1.DERAbstractString),m.asn1.DERIA5String=function(i){m.asn1.DERIA5String.superclass.constructor.call(this,i),this.hT="16"},Y.lang.extend(m.asn1.DERIA5String,m.asn1.DERAbstractString),m.asn1.DERUTCTime=function(i){m.asn1.DERUTCTime.superclass.constructor.call(this,i),this.hT="17",this.setByDate=function(e){this.hTLV=null,this.isModified=!0,this.date=e,this.s=this.formatDate(this.date,"utc"),this.hV=stohex(this.s)},this.getFreshValueHex=function(){return typeof this.date>"u"&&typeof this.s>"u"&&(this.date=new Date,this.s=this.formatDate(this.date,"utc"),this.hV=stohex(this.s)),this.hV},i!==void 0&&(i.str!==void 0?this.setString(i.str):typeof i=="string"&&i.match(/^[0-9]{12}Z$/)?this.setString(i):i.hex!==void 0?this.setStringHex(i.hex):i.date!==void 0&&this.setByDate(i.date))},Y.lang.extend(m.asn1.DERUTCTime,m.asn1.DERAbstractTime),m.asn1.DERGeneralizedTime=function(i){m.asn1.DERGeneralizedTime.superclass.constructor.call(this,i),this.hT="18",this.withMillis=!1,this.setByDate=function(e){this.hTLV=null,this.isModified=!0,this.date=e,this.s=this.formatDate(this.date,"gen",this.withMillis),this.hV=stohex(this.s)},this.getFreshValueHex=function(){return this.date===void 0&&this.s===void 0&&(this.date=new Date,this.s=this.formatDate(this.date,"gen",this.withMillis),this.hV=stohex(this.s)),this.hV},i!==void 0&&(i.str!==void 0?this.setString(i.str):typeof i=="string"&&i.match(/^[0-9]{14}Z$/)?this.setString(i):i.hex!==void 0?this.setStringHex(i.hex):i.date!==void 0&&this.setByDate(i.date),i.millis===!0&&(this.withMillis=!0))},Y.lang.extend(m.asn1.DERGeneralizedTime,m.asn1.DERAbstractTime),m.asn1.DERSequence=function(i){m.asn1.DERSequence.superclass.constructor.call(this,i),this.hT="30",this.getFreshValueHex=function(){for(var e="",t=0;t<this.asn1Array.length;t++){var r=this.asn1Array[t];e+=r.getEncodedHex()}return this.hV=e,this.hV}},Y.lang.extend(m.asn1.DERSequence,m.asn1.DERAbstractStructured),m.asn1.DERSet=function(i){m.asn1.DERSet.superclass.constructor.call(this,i),this.hT="31",this.sortFlag=!0,this.getFreshValueHex=function(){for(var e=new Array,t=0;t<this.asn1Array.length;t++){var r=this.asn1Array[t];e.push(r.getEncodedHex())}return this.sortFlag==!0&&e.sort(),this.hV=e.join(""),this.hV},typeof i<"u"&&typeof i.sortflag<"u"&&i.sortflag==!1&&(this.sortFlag=!1)},Y.lang.extend(m.asn1.DERSet,m.asn1.DERAbstractStructured),m.asn1.DERTaggedObject=function(i){m.asn1.DERTaggedObject.superclass.constructor.call(this),this.hT="a0",this.hV="",this.isExplicit=!0,this.asn1Object=null,this.setASN1Object=function(e,t,r){this.hT=t,this.isExplicit=e,this.asn1Object=r,this.isExplicit?(this.hV=this.asn1Object.getEncodedHex(),this.hTLV=null,this.isModified=!0):(this.hV=null,this.hTLV=r.getEncodedHex(),this.hTLV=this.hTLV.replace(/^../,t),this.isModified=!1)},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.tag<"u"&&(this.hT=i.tag),typeof i.explicit<"u"&&(this.isExplicit=i.explicit),typeof i.obj<"u"&&(this.asn1Object=i.obj,this.setASN1Object(this.isExplicit,this.hT,this.asn1Object)))},Y.lang.extend(m.asn1.DERTaggedObject,m.asn1.ASN1Object);var hi=function(){var i=function(e,t){return i=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(r,n){r.__proto__=n}||function(r,n){for(var s in n)Object.prototype.hasOwnProperty.call(n,s)&&(r[s]=n[s])},i(e,t)};return function(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");i(e,t);function r(){this.constructor=e}e.prototype=t===null?Object.create(t):(r.prototype=t.prototype,new r)}}(),Ke=function(i){hi(e,i);function e(t){var r=i.call(this)||this;return t&&(typeof t=="string"?r.parseKey(t):(e.hasPrivateKeyProperty(t)||e.hasPublicKeyProperty(t))&&r.parsePropertiesFrom(t)),r}return e.prototype.parseKey=function(t){try{var r=0,n=0,s=/^\s*(?:[0-9A-Fa-f][0-9A-Fa-f]\s*)+$/,x=s.test(t)?Gr.decode(t):zt.unarmor(t),o=Qr.decode(x);if(o.sub.length===3&&(o=o.sub[2].sub[0]),o.sub.length===9){r=o.sub[1].getHexStringValue(),this.n=Q(r,16),n=o.sub[2].getHexStringValue(),this.e=parseInt(n,16);var f=o.sub[3].getHexStringValue();this.d=Q(f,16);var a=o.sub[4].getHexStringValue();this.p=Q(a,16);var c=o.sub[5].getHexStringValue();this.q=Q(c,16);var E=o.sub[6].getHexStringValue();this.dmp1=Q(E,16);var h=o.sub[7].getHexStringValue();this.dmq1=Q(h,16);var u=o.sub[8].getHexStringValue();this.coeff=Q(u,16)}else if(o.sub.length===2)if(o.sub[0].sub){var v=o.sub[1],p=v.sub[0];r=p.sub[0].getHexStringValue(),this.n=Q(r,16),n=p.sub[1].getHexStringValue(),this.e=parseInt(n,16)}else r=o.sub[0].getHexStringValue(),this.n=Q(r,16),n=o.sub[1].getHexStringValue(),this.e=parseInt(n,16);else return!1;return!0}catch{return!1}},e.prototype.getPrivateBaseKey=function(){var t={array:[new m.asn1.DERInteger({int:0}),new m.asn1.DERInteger({bigint:this.n}),new m.asn1.DERInteger({int:this.e}),new m.asn1.DERInteger({bigint:this.d}),new m.asn1.DERInteger({bigint:this.p}),new m.asn1.DERInteger({bigint:this.q}),new m.asn1.DERInteger({bigint:this.dmp1}),new m.asn1.DERInteger({bigint:this.dmq1}),new m.asn1.DERInteger({bigint:this.coeff})]},r=new m.asn1.DERSequence(t);return r.getEncodedHex()},e.prototype.getPrivateBaseKeyB64=function(){return U0(this.getPrivateBaseKey())},e.prototype.getPublicBaseKey=function(){var t=new m.asn1.DERSequence({array:[new m.asn1.DERObjectIdentifier({oid:"1.2.840.113549.1.1.1"}),new m.asn1.DERNull]}),r=new m.asn1.DERSequence({array:[new m.asn1.DERInteger({bigint:this.n}),new m.asn1.DERInteger({int:this.e})]}),n=new m.asn1.DERBitString({hex:"00"+r.getEncodedHex()}),s=new m.asn1.DERSequence({array:[t,n]});return s.getEncodedHex()},e.prototype.getPublicBaseKeyB64=function(){return U0(this.getPublicBaseKey())},e.wordwrap=function(t,r){if(r=r||64,!t)return t;var n="(.{1,"+r+`})( +|$
+ */var b={};(typeof b.asn1>"u"||!b.asn1)&&(b.asn1={}),b.asn1.ASN1Util=new function(){this.integerToByteHex=function(i){var e=i.toString(16);return e.length%2==1&&(e="0"+e),e},this.bigIntToMinTwosComplementsHex=function(i){var e=i.toString(16);if(e.substr(0,1)!="-")e.length%2==1?e="0"+e:e.match(/^[0-7]/)||(e="00"+e);else{var t=e.substr(1),r=t.length;r%2==1?r+=1:e.match(/^[0-7]/)||(r+=2);for(var n="",s=0;s<r;s++)n+="f";var x=new z(n,16),o=x.xor(i).add(z.ONE);e=o.toString(16).replace(/^-/,"")}return e},this.getPEMStringFromHex=function(i,e){return hextopem(i,e)},this.newObject=function(i){var e=b,t=e.asn1,r=t.DERBoolean,n=t.DERInteger,s=t.DERBitString,x=t.DEROctetString,o=t.DERNull,f=t.DERObjectIdentifier,a=t.DEREnumerated,c=t.DERUTF8String,B=t.DERNumericString,h=t.DERPrintableString,u=t.DERTeletexString,v=t.DERIA5String,g=t.DERUTCTime,l=t.DERGeneralizedTime,A=t.DERSequence,d=t.DERSet,p=t.DERTaggedObject,E=t.ASN1Util.newObject,F=Object.keys(i);if(F.length!=1)throw"key of param shall be only one.";var C=F[0];if(":bool:int:bitstr:octstr:null:oid:enum:utf8str:numstr:prnstr:telstr:ia5str:utctime:gentime:seq:set:tag:".indexOf(":"+C+":")==-1)throw"undefined key: "+C;if(C=="bool")return new r(i[C]);if(C=="int")return new n(i[C]);if(C=="bitstr")return new s(i[C]);if(C=="octstr")return new x(i[C]);if(C=="null")return new o(i[C]);if(C=="oid")return new f(i[C]);if(C=="enum")return new a(i[C]);if(C=="utf8str")return new c(i[C]);if(C=="numstr")return new B(i[C]);if(C=="prnstr")return new h(i[C]);if(C=="telstr")return new u(i[C]);if(C=="ia5str")return new v(i[C]);if(C=="utctime")return new g(i[C]);if(C=="gentime")return new l(i[C]);if(C=="seq"){for(var _=i[C],m=[],I=0;I<_.length;I++){var D=E(_[I]);m.push(D)}return new A({array:m})}if(C=="set"){for(var _=i[C],m=[],I=0;I<_.length;I++){var D=E(_[I]);m.push(D)}return new d({array:m})}if(C=="tag"){var y=i[C];if(Object.prototype.toString.call(y)==="[object Array]"&&y.length==3){var T=E(y[2]);return new p({tag:y[0],explicit:y[1],obj:T})}else{var w={};if(y.explicit!==void 0&&(w.explicit=y.explicit),y.tag!==void 0&&(w.tag=y.tag),y.obj===void 0)throw"obj shall be specified for 'tag'.";return w.obj=E(y.obj),new p(w)}}},this.jsonToASN1HEX=function(i){var e=this.newObject(i);return e.getEncodedHex()}},b.asn1.ASN1Util.oidHexToInt=function(i){for(var n="",e=parseInt(i.substr(0,2),16),t=Math.floor(e/40),r=e%40,n=t+"."+r,s="",x=2;x<i.length;x+=2){var o=parseInt(i.substr(x,2),16),f=("00000000"+o.toString(2)).slice(-8);if(s=s+f.substr(1,7),f.substr(0,1)=="0"){var a=new z(s,2);n=n+"."+a.toString(10),s=""}}return n},b.asn1.ASN1Util.oidIntToHex=function(i){var e=function(o){var f=o.toString(16);return f.length==1&&(f="0"+f),f},t=function(o){var f="",a=new z(o,10),c=a.toString(2),B=7-c.length%7;B==7&&(B=0);for(var h="",u=0;u<B;u++)h+="0";c=h+c;for(var u=0;u<c.length-1;u+=7){var v=c.substr(u,7);u!=c.length-7&&(v="1"+v),f+=e(parseInt(v,2))}return f};if(!i.match(/^[0-9.]+$/))throw"malformed oid string: "+i;var r="",n=i.split("."),s=parseInt(n[0])*40+parseInt(n[1]);r+=e(s),n.splice(0,2);for(var x=0;x<n.length;x++)r+=t(n[x]);return r},b.asn1.ASN1Object=function(){var i="";this.getLengthHexFromValue=function(){if(typeof this.hV>"u"||this.hV==null)throw"this.hV is null or undefined.";if(this.hV.length%2==1)throw"value hex must be even length: n="+i.length+",v="+this.hV;var e=this.hV.length/2,t=e.toString(16);if(t.length%2==1&&(t="0"+t),e<128)return t;var r=t.length/2;if(r>15)throw"ASN.1 length too long to represent by 8x: n = "+e.toString(16);var n=128+r;return n.toString(16)+t},this.getEncodedHex=function(){return(this.hTLV==null||this.isModified)&&(this.hV=this.getFreshValueHex(),this.hL=this.getLengthHexFromValue(),this.hTLV=this.hT+this.hL+this.hV,this.isModified=!1),this.hTLV},this.getValueHex=function(){return this.getEncodedHex(),this.hV},this.getFreshValueHex=function(){return""}},b.asn1.DERAbstractString=function(i){b.asn1.DERAbstractString.superclass.constructor.call(this),this.getString=function(){return this.s},this.setString=function(e){this.hTLV=null,this.isModified=!0,this.s=e,this.hV=stohex(this.s)},this.setStringHex=function(e){this.hTLV=null,this.isModified=!0,this.s=null,this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i=="string"?this.setString(i):typeof i.str<"u"?this.setString(i.str):typeof i.hex<"u"&&this.setStringHex(i.hex))},Y.lang.extend(b.asn1.DERAbstractString,b.asn1.ASN1Object),b.asn1.DERAbstractTime=function(i){b.asn1.DERAbstractTime.superclass.constructor.call(this),this.localDateToUTC=function(e){utc=e.getTime()+e.getTimezoneOffset()*6e4;var t=new Date(utc);return t},this.formatDate=function(e,t,r){var n=this.zeroPadding,s=this.localDateToUTC(e),x=String(s.getFullYear());t=="utc"&&(x=x.substr(2,2));var o=n(String(s.getMonth()+1),2),f=n(String(s.getDate()),2),a=n(String(s.getHours()),2),c=n(String(s.getMinutes()),2),B=n(String(s.getSeconds()),2),h=x+o+f+a+c+B;if(r===!0){var u=s.getMilliseconds();if(u!=0){var v=n(String(u),3);v=v.replace(/[0]+$/,""),h=h+"."+v}}return h+"Z"},this.zeroPadding=function(e,t){return e.length>=t?e:new Array(t-e.length+1).join("0")+e},this.getString=function(){return this.s},this.setString=function(e){this.hTLV=null,this.isModified=!0,this.s=e,this.hV=stohex(e)},this.setByDateValue=function(e,t,r,n,s,x){var o=new Date(Date.UTC(e,t-1,r,n,s,x,0));this.setByDate(o)},this.getFreshValueHex=function(){return this.hV}},Y.lang.extend(b.asn1.DERAbstractTime,b.asn1.ASN1Object),b.asn1.DERAbstractStructured=function(i){b.asn1.DERAbstractString.superclass.constructor.call(this),this.setByASN1ObjectArray=function(e){this.hTLV=null,this.isModified=!0,this.asn1Array=e},this.appendASN1Object=function(e){this.hTLV=null,this.isModified=!0,this.asn1Array.push(e)},this.asn1Array=new Array,typeof i<"u"&&typeof i.array<"u"&&(this.asn1Array=i.array)},Y.lang.extend(b.asn1.DERAbstractStructured,b.asn1.ASN1Object),b.asn1.DERBoolean=function(){b.asn1.DERBoolean.superclass.constructor.call(this),this.hT="01",this.hTLV="0101ff"},Y.lang.extend(b.asn1.DERBoolean,b.asn1.ASN1Object),b.asn1.DERInteger=function(i){b.asn1.DERInteger.superclass.constructor.call(this),this.hT="02",this.setByBigInteger=function(e){this.hTLV=null,this.isModified=!0,this.hV=b.asn1.ASN1Util.bigIntToMinTwosComplementsHex(e)},this.setByInteger=function(e){var t=new z(String(e),10);this.setByBigInteger(t)},this.setValueHex=function(e){this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.bigint<"u"?this.setByBigInteger(i.bigint):typeof i.int<"u"?this.setByInteger(i.int):typeof i=="number"?this.setByInteger(i):typeof i.hex<"u"&&this.setValueHex(i.hex))},Y.lang.extend(b.asn1.DERInteger,b.asn1.ASN1Object),b.asn1.DERBitString=function(i){if(i!==void 0&&typeof i.obj<"u"){var e=b.asn1.ASN1Util.newObject(i.obj);i.hex="00"+e.getEncodedHex()}b.asn1.DERBitString.superclass.constructor.call(this),this.hT="03",this.setHexValueIncludingUnusedBits=function(t){this.hTLV=null,this.isModified=!0,this.hV=t},this.setUnusedBitsAndHexValue=function(t,r){if(t<0||7<t)throw"unused bits shall be from 0 to 7: u = "+t;var n="0"+t;this.hTLV=null,this.isModified=!0,this.hV=n+r},this.setByBinaryString=function(t){t=t.replace(/0+$/,"");var r=8-t.length%8;r==8&&(r=0);for(var n=0;n<=r;n++)t+="0";for(var s="",n=0;n<t.length-1;n+=8){var x=t.substr(n,8),o=parseInt(x,2).toString(16);o.length==1&&(o="0"+o),s+=o}this.hTLV=null,this.isModified=!0,this.hV="0"+r+s},this.setByBooleanArray=function(t){for(var r="",n=0;n<t.length;n++)t[n]==!0?r+="1":r+="0";this.setByBinaryString(r)},this.newFalseArray=function(t){for(var r=new Array(t),n=0;n<t;n++)r[n]=!1;return r},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i=="string"&&i.toLowerCase().match(/^[0-9a-f]+$/)?this.setHexValueIncludingUnusedBits(i):typeof i.hex<"u"?this.setHexValueIncludingUnusedBits(i.hex):typeof i.bin<"u"?this.setByBinaryString(i.bin):typeof i.array<"u"&&this.setByBooleanArray(i.array))},Y.lang.extend(b.asn1.DERBitString,b.asn1.ASN1Object),b.asn1.DEROctetString=function(i){if(i!==void 0&&typeof i.obj<"u"){var e=b.asn1.ASN1Util.newObject(i.obj);i.hex=e.getEncodedHex()}b.asn1.DEROctetString.superclass.constructor.call(this,i),this.hT="04"},Y.lang.extend(b.asn1.DEROctetString,b.asn1.DERAbstractString),b.asn1.DERNull=function(){b.asn1.DERNull.superclass.constructor.call(this),this.hT="05",this.hTLV="0500"},Y.lang.extend(b.asn1.DERNull,b.asn1.ASN1Object),b.asn1.DERObjectIdentifier=function(i){var e=function(r){var n=r.toString(16);return n.length==1&&(n="0"+n),n},t=function(r){var n="",s=new z(r,10),x=s.toString(2),o=7-x.length%7;o==7&&(o=0);for(var f="",a=0;a<o;a++)f+="0";x=f+x;for(var a=0;a<x.length-1;a+=7){var c=x.substr(a,7);a!=x.length-7&&(c="1"+c),n+=e(parseInt(c,2))}return n};b.asn1.DERObjectIdentifier.superclass.constructor.call(this),this.hT="06",this.setValueHex=function(r){this.hTLV=null,this.isModified=!0,this.s=null,this.hV=r},this.setValueOidString=function(r){if(!r.match(/^[0-9.]+$/))throw"malformed oid string: "+r;var n="",s=r.split("."),x=parseInt(s[0])*40+parseInt(s[1]);n+=e(x),s.splice(0,2);for(var o=0;o<s.length;o++)n+=t(s[o]);this.hTLV=null,this.isModified=!0,this.s=null,this.hV=n},this.setValueName=function(r){var n=b.asn1.x509.OID.name2oid(r);if(n!=="")this.setValueOidString(n);else throw"DERObjectIdentifier oidName undefined: "+r},this.getFreshValueHex=function(){return this.hV},i!==void 0&&(typeof i=="string"?i.match(/^[0-2].[0-9.]+$/)?this.setValueOidString(i):this.setValueName(i):i.oid!==void 0?this.setValueOidString(i.oid):i.hex!==void 0?this.setValueHex(i.hex):i.name!==void 0&&this.setValueName(i.name))},Y.lang.extend(b.asn1.DERObjectIdentifier,b.asn1.ASN1Object),b.asn1.DEREnumerated=function(i){b.asn1.DEREnumerated.superclass.constructor.call(this),this.hT="0a",this.setByBigInteger=function(e){this.hTLV=null,this.isModified=!0,this.hV=b.asn1.ASN1Util.bigIntToMinTwosComplementsHex(e)},this.setByInteger=function(e){var t=new z(String(e),10);this.setByBigInteger(t)},this.setValueHex=function(e){this.hV=e},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.int<"u"?this.setByInteger(i.int):typeof i=="number"?this.setByInteger(i):typeof i.hex<"u"&&this.setValueHex(i.hex))},Y.lang.extend(b.asn1.DEREnumerated,b.asn1.ASN1Object),b.asn1.DERUTF8String=function(i){b.asn1.DERUTF8String.superclass.constructor.call(this,i),this.hT="0c"},Y.lang.extend(b.asn1.DERUTF8String,b.asn1.DERAbstractString),b.asn1.DERNumericString=function(i){b.asn1.DERNumericString.superclass.constructor.call(this,i),this.hT="12"},Y.lang.extend(b.asn1.DERNumericString,b.asn1.DERAbstractString),b.asn1.DERPrintableString=function(i){b.asn1.DERPrintableString.superclass.constructor.call(this,i),this.hT="13"},Y.lang.extend(b.asn1.DERPrintableString,b.asn1.DERAbstractString),b.asn1.DERTeletexString=function(i){b.asn1.DERTeletexString.superclass.constructor.call(this,i),this.hT="14"},Y.lang.extend(b.asn1.DERTeletexString,b.asn1.DERAbstractString),b.asn1.DERIA5String=function(i){b.asn1.DERIA5String.superclass.constructor.call(this,i),this.hT="16"},Y.lang.extend(b.asn1.DERIA5String,b.asn1.DERAbstractString),b.asn1.DERUTCTime=function(i){b.asn1.DERUTCTime.superclass.constructor.call(this,i),this.hT="17",this.setByDate=function(e){this.hTLV=null,this.isModified=!0,this.date=e,this.s=this.formatDate(this.date,"utc"),this.hV=stohex(this.s)},this.getFreshValueHex=function(){return typeof this.date>"u"&&typeof this.s>"u"&&(this.date=new Date,this.s=this.formatDate(this.date,"utc"),this.hV=stohex(this.s)),this.hV},i!==void 0&&(i.str!==void 0?this.setString(i.str):typeof i=="string"&&i.match(/^[0-9]{12}Z$/)?this.setString(i):i.hex!==void 0?this.setStringHex(i.hex):i.date!==void 0&&this.setByDate(i.date))},Y.lang.extend(b.asn1.DERUTCTime,b.asn1.DERAbstractTime),b.asn1.DERGeneralizedTime=function(i){b.asn1.DERGeneralizedTime.superclass.constructor.call(this,i),this.hT="18",this.withMillis=!1,this.setByDate=function(e){this.hTLV=null,this.isModified=!0,this.date=e,this.s=this.formatDate(this.date,"gen",this.withMillis),this.hV=stohex(this.s)},this.getFreshValueHex=function(){return this.date===void 0&&this.s===void 0&&(this.date=new Date,this.s=this.formatDate(this.date,"gen",this.withMillis),this.hV=stohex(this.s)),this.hV},i!==void 0&&(i.str!==void 0?this.setString(i.str):typeof i=="string"&&i.match(/^[0-9]{14}Z$/)?this.setString(i):i.hex!==void 0?this.setStringHex(i.hex):i.date!==void 0&&this.setByDate(i.date),i.millis===!0&&(this.withMillis=!0))},Y.lang.extend(b.asn1.DERGeneralizedTime,b.asn1.DERAbstractTime),b.asn1.DERSequence=function(i){b.asn1.DERSequence.superclass.constructor.call(this,i),this.hT="30",this.getFreshValueHex=function(){for(var e="",t=0;t<this.asn1Array.length;t++){var r=this.asn1Array[t];e+=r.getEncodedHex()}return this.hV=e,this.hV}},Y.lang.extend(b.asn1.DERSequence,b.asn1.DERAbstractStructured),b.asn1.DERSet=function(i){b.asn1.DERSet.superclass.constructor.call(this,i),this.hT="31",this.sortFlag=!0,this.getFreshValueHex=function(){for(var e=new Array,t=0;t<this.asn1Array.length;t++){var r=this.asn1Array[t];e.push(r.getEncodedHex())}return this.sortFlag==!0&&e.sort(),this.hV=e.join(""),this.hV},typeof i<"u"&&typeof i.sortflag<"u"&&i.sortflag==!1&&(this.sortFlag=!1)},Y.lang.extend(b.asn1.DERSet,b.asn1.DERAbstractStructured),b.asn1.DERTaggedObject=function(i){b.asn1.DERTaggedObject.superclass.constructor.call(this),this.hT="a0",this.hV="",this.isExplicit=!0,this.asn1Object=null,this.setASN1Object=function(e,t,r){this.hT=t,this.isExplicit=e,this.asn1Object=r,this.isExplicit?(this.hV=this.asn1Object.getEncodedHex(),this.hTLV=null,this.isModified=!0):(this.hV=null,this.hTLV=r.getEncodedHex(),this.hTLV=this.hTLV.replace(/^../,t),this.isModified=!1)},this.getFreshValueHex=function(){return this.hV},typeof i<"u"&&(typeof i.tag<"u"&&(this.hT=i.tag),typeof i.explicit<"u"&&(this.isExplicit=i.explicit),typeof i.obj<"u"&&(this.asn1Object=i.obj,this.setASN1Object(this.isExplicit,this.hT,this.asn1Object)))},Y.lang.extend(b.asn1.DERTaggedObject,b.asn1.ASN1Object);var di=function(){var i=function(e,t){return i=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(r,n){r.__proto__=n}||function(r,n){for(var s in n)Object.prototype.hasOwnProperty.call(n,s)&&(r[s]=n[s])},i(e,t)};return function(e,t){if(typeof t!="function"&&t!==null)throw new TypeError("Class extends value "+String(t)+" is not a constructor or null");i(e,t);function r(){this.constructor=e}e.prototype=t===null?Object.create(t):(r.prototype=t.prototype,new r)}}(),Ge=function(i){di(e,i);function e(t){var r=i.call(this)||this;return t&&(typeof t=="string"?r.parseKey(t):(e.hasPrivateKeyProperty(t)||e.hasPublicKeyProperty(t))&&r.parsePropertiesFrom(t)),r}return e.prototype.parseKey=function(t){try{var r=0,n=0,s=/^\s*(?:[0-9A-Fa-f][0-9A-Fa-f]\s*)+$/,x=s.test(t)?Qr.decode(t):Wt.unarmor(t),o=ti.decode(x);if(o.sub.length===3&&(o=o.sub[2].sub[0]),o.sub.length===9){r=o.sub[1].getHexStringValue(),this.n=Q(r,16),n=o.sub[2].getHexStringValue(),this.e=parseInt(n,16);var f=o.sub[3].getHexStringValue();this.d=Q(f,16);var a=o.sub[4].getHexStringValue();this.p=Q(a,16);var c=o.sub[5].getHexStringValue();this.q=Q(c,16);var B=o.sub[6].getHexStringValue();this.dmp1=Q(B,16);var h=o.sub[7].getHexStringValue();this.dmq1=Q(h,16);var u=o.sub[8].getHexStringValue();this.coeff=Q(u,16)}else if(o.sub.length===2)if(o.sub[0].sub){var v=o.sub[1],g=v.sub[0];r=g.sub[0].getHexStringValue(),this.n=Q(r,16),n=g.sub[1].getHexStringValue(),this.e=parseInt(n,16)}else r=o.sub[0].getHexStringValue(),this.n=Q(r,16),n=o.sub[1].getHexStringValue(),this.e=parseInt(n,16);else return!1;return!0}catch{return!1}},e.prototype.getPrivateBaseKey=function(){var t={array:[new b.asn1.DERInteger({int:0}),new b.asn1.DERInteger({bigint:this.n}),new b.asn1.DERInteger({int:this.e}),new b.asn1.DERInteger({bigint:this.d}),new b.asn1.DERInteger({bigint:this.p}),new b.asn1.DERInteger({bigint:this.q}),new b.asn1.DERInteger({bigint:this.dmp1}),new b.asn1.DERInteger({bigint:this.dmq1}),new b.asn1.DERInteger({bigint:this.coeff})]},r=new b.asn1.DERSequence(t);return r.getEncodedHex()},e.prototype.getPrivateBaseKeyB64=function(){return V0(this.getPrivateBaseKey())},e.prototype.getPublicBaseKey=function(){var t=new b.asn1.DERSequence({array:[new b.asn1.DERObjectIdentifier({oid:"1.2.840.113549.1.1.1"}),new b.asn1.DERNull]}),r=new b.asn1.DERSequence({array:[new b.asn1.DERInteger({bigint:this.n}),new b.asn1.DERInteger({int:this.e})]}),n=new b.asn1.DERBitString({hex:"00"+r.getEncodedHex()}),s=new b.asn1.DERSequence({array:[t,n]});return s.getEncodedHex()},e.prototype.getPublicBaseKeyB64=function(){return V0(this.getPublicBaseKey())},e.wordwrap=function(t,r){if(r=r||64,!t)return t;var n="(.{1,"+r+`})( +|$
 ?)|(.{1,`+r+"})";return t.match(RegExp(n,"g")).join(`
 `)},e.prototype.getPrivateKey=function(){var t=`-----BEGIN RSA PRIVATE KEY-----
 `;return t+=e.wordwrap(this.getPrivateBaseKeyB64())+`
 `,t+="-----END RSA PRIVATE KEY-----",t},e.prototype.getPublicKey=function(){var t=`-----BEGIN PUBLIC KEY-----
 `;return t+=e.wordwrap(this.getPublicBaseKeyB64())+`
-`,t+="-----END PUBLIC KEY-----",t},e.hasPublicKeyProperty=function(t){return t=t||{},t.hasOwnProperty("n")&&t.hasOwnProperty("e")},e.hasPrivateKeyProperty=function(t){return t=t||{},t.hasOwnProperty("n")&&t.hasOwnProperty("e")&&t.hasOwnProperty("d")&&t.hasOwnProperty("p")&&t.hasOwnProperty("q")&&t.hasOwnProperty("dmp1")&&t.hasOwnProperty("dmq1")&&t.hasOwnProperty("coeff")},e.prototype.parsePropertiesFrom=function(t){this.n=t.n,this.e=t.e,t.hasOwnProperty("d")&&(this.d=t.d,this.p=t.p,this.q=t.q,this.dmp1=t.dmp1,this.dmq1=t.dmq1,this.coeff=t.coeff)},e}(ai),Vt,ui=typeof process<"u"?(Vt=process.env)===null||Vt===void 0?void 0:Vt.npm_package_version:void 0,li=function(){function i(e){e===void 0&&(e={}),e=e||{},this.default_key_size=e.default_key_size?parseInt(e.default_key_size,10):1024,this.default_public_exponent=e.default_public_exponent||"010001",this.log=e.log||!1,this.key=null}return i.prototype.setKey=function(e){this.log&&this.key&&console.warn("A key was already set, overriding existing."),this.key=new Ke(e)},i.prototype.setPrivateKey=function(e){this.setKey(e)},i.prototype.setPublicKey=function(e){this.setKey(e)},i.prototype.decrypt=function(e){try{return this.getKey().decrypt(He(e))}catch{return!1}},i.prototype.encrypt=function(e){try{return U0(this.getKey().encrypt(e))}catch{return!1}},i.prototype.sign=function(e,t,r){try{return U0(this.getKey().sign(e,t,r))}catch{return!1}},i.prototype.verify=function(e,t,r){try{return this.getKey().verify(e,He(t),r)}catch{return!1}},i.prototype.getKey=function(e){if(!this.key){if(this.key=new Ke,e&&{}.toString.call(e)==="[object Function]"){this.key.generateAsync(this.default_key_size,this.default_public_exponent,e);return}this.key.generate(this.default_key_size,this.default_public_exponent)}return this.key},i.prototype.getPrivateKey=function(){return this.getKey().getPrivateKey()},i.prototype.getPrivateKeyB64=function(){return this.getKey().getPrivateBaseKeyB64()},i.prototype.getPublicKey=function(){return this.getKey().getPublicKey()},i.prototype.getPublicKeyB64=function(){return this.getKey().getPublicBaseKeyB64()},i.version=ui,i}();const $e=(i,e)=>{const t=new li;return t.setPublicKey(i),t.encrypt(e)},di=()=>{const i=Nt.lib.WordArray.random(16),e=Nt.lib.WordArray.random(16);return{key:i,iv:e}},vi=({key:i,iv:e},t)=>Nt.AES.encrypt(t,i,{iv:e}).toString(),pi=(i,e)=>{const{key:t,iv:r}=di(),n=t.toString(),s=r.toString(),x=$e(i,n);if(x===!1)throw new Error("Failed to encrypt AES key.");const o=$e(i,s);if(o===!1)throw new Error("Failed to encrypt AES IV.");const f=vi({key:t,iv:r},e);return{encryptedKey:x,encryptedIV:o,encryptedBody:f}};class je{constructor(e){if(this._options=e,!("fetch"in window))throw new Error("Fetch is not available in this browser.");this._options.headers||(this._options.headers={})}setOptions(e){this._options=e}send(e,t,r){const n={method:t,headers:this._options.headers,mode:"cors",redirect:"follow"};return t==="POST"&&r&&(n.body=r),new Promise((s,x)=>{window.fetch(e,n).then(o=>{s(o)}).catch(o=>{x(o)})})}}class Ge{static getTransport(e,t){return this.transports[e]?new this.transports[e](t):new je(t)}}i0(Ge,"transports",{http:je});const gi=(i,e,t,r,n,s,x,o,f,a,c,E,h="telemetree_twa")=>({eventType:i,userDetails:e,eventDetails:t,telegramID:r,language:n,device:s,referrerType:x,referrer:o,timestamp:f,isAutocapture:a,wallet:c,sessionIdentifier:E,eventSource:h});var Kt=(i=>(i.ERROR="error",i.WARN="warn",i.INFO="info",i.DEBUG="debug",i))(Kt||{});class n0{static setLevel(e){this.level=e}static log(e,t,r){if(this.shouldLog(e)){const n=r?`${t} ${JSON.stringify(r)}`:t;console[e](n)}}static shouldLog(e){const t=Object.values(Kt);return t.indexOf(e)<=t.indexOf(this.level)}static error(e,t){this.log("error",e,t)}static warn(e,t){this.log("warn",e,t)}static info(e,t){this.log("info",e,t)}static debug(e,t){this.log("debug",e,t)}}i0(n0,"level","info");class Ze{constructor(e,t,r){i0(this,"transport",null);i0(this,"config",null);i0(this,"sessionIdentifier",null);i0(this,"userIdentifier",null);i0(this,"pushHandler",new ur(this));this.projectId=e,this.apiKey=t,this.data=r,this._init()}setSessionIdentifier(e){return this.sessionIdentifier=e,this}setUserId(e){return this.userIdentifier=e,this}async _init(){const e=Ge.getTransport("http",{headers:{authorization:`Bearer ${this.apiKey}`},requestTimeout:1e3});this.setSessionIdentifier(Gt());try{const t=await e.send(a0+`?project=${this.projectId}`,"GET");if(t.status===200){const r=await t.json();this.config=r}e.setOptions({headers:{"x-api-key":`${this.apiKey}`,"x-project-id":`${this.projectId}`,"content-type":"application/json"},requestTimeout:1500}),this.setTransport(e),await this.pushHandler.flush(),this.setupAutoCaptureListener()}catch(t){console.error(`Cannot load config: ${t}`)}}setupAutoCaptureListener(){var t;const e=this.getConfig();if(e!==null&&e.auto_capture===!0){const r=e.auto_capture_tags.map(n=>n.toUpperCase());(t=document.querySelector("body"))==null||t.addEventListener("click",n=>{var o,f,a,c,E,h;const x=(((o=n.composedPath)==null?void 0:o.call(n))||[]).find(u=>u instanceof HTMLElement&&r.includes(u.tagName));if(x){const u={},v=["id","href","class","name","value","type","placeholder","title","alt","src"];for(const C of v)x.hasAttribute(C)&&(u[C]=x.getAttribute(C));const p=((a=(f=n.target)==null?void 0:f.innerText)==null?void 0:a.trim())||"",l=((c=x.innerText)==null?void 0:c.trim())||"";u.text=p||l,u.tag=x.tagName.toLowerCase(),n.target!==x&&(u.clicked_child=(h=(E=n.target)==null?void 0:E.tagName)==null?void 0:h.toLowerCase(),u.clicked_child_text=p),this.track(`${v0} ${q.Click}${hr}${u.text}`,u)}})}}getConfig(){return this.config}setTransport(e){return this.transport=e,this}async track(e,t){var n,s,x,o,f,a,c;if(e===null)throw new Error("Event name is not set.");if(!((n=this.data.user)!=null&&n.id)){console.error(`Event ${e} is not tracked because user is not set.`);return}const r=gi(e,{username:((s=this.data.user)==null?void 0:s.username)||"",firstName:this.data.user.first_name,lastName:((x=this.data.user)==null?void 0:x.last_name)||"",isPremium:((o=this.data.user)==null?void 0:o.is_premium)||!1,writeAccess:((f=this.data.user)==null?void 0:f.allows_write_to_pm)||!1},{startParameter:this.data.start_param||"",path:document.location.pathname,params:t},(a=this.data.user)==null?void 0:a.id.toString(),(c=this.data.user)==null?void 0:c.language_code,this.data.platform,this.data.chat_type||"N/A",this.data.chat_instance||"0",lr(),e.startsWith(v0),t.wallet||void 0,this.sessionIdentifier||void 0);return this.pushHandler.push(r)}async processEvent(e){if(this.config===null||this.transport===null){n0.info("Event processing failed: config or transport is not set.");return}try{const{encryptedKey:t,encryptedIV:r,encryptedBody:n}=pi(this.config.public_key,JSON.stringify(e));await this.transport.send(this.config.host,"POST",JSON.stringify({key:t,iv:r,body:n})),n0.debug("Event sent successfully",{eventData:e})}catch(t){console.error(t)}}}const Q0=typeof window<"u"&&((Qe=window==null?void 0:window.Telegram)!=null&&Qe.WebApp)?window.Telegram.WebApp:null,I=typeof window<"u"&&((Ye=window==null?void 0:window.Telegram)!=null&&Ye.WebView)?window.Telegram.WebView:null,Ei=()=>Q0===null?Ai():Bi(),Bi=()=>{if(Q0===null)throw new Error("Telegram WebApp is not loaded.");return{...Q0.initDataUnsafe,platform:Q0.platform}},Ai=()=>{if(document.location.hash==="")throw new Error("Telegram WebApp is not loaded.");const e=decodeURIComponent(decodeURIComponent(document.location.hash.substring(1))),t=new URLSearchParams(e),r=t.get("tgWebAppData");if(r===null)throw new Error("Telegram WebApp is not loaded.");const n=new URLSearchParams(r);return{user:n.get("user")?JSON.parse(n.get("user")):void 0,chat_type:t.get("chat_type"),chat_instance:t.get("chat_instance")||void 0,platform:t.get("tgWebAppPlatform")||"unknown",auth_date:t.get("auth_date")?parseInt(t.get("auth_date")):0,hash:t.get("hash")||"",start_param:t.get("start_param")||t.get("startapp")||void 0}},Ci=(i,e)=>{localStorage.setItem(i,e)},Di=i=>localStorage.getItem(i),yi=1e12,Xe=1,Fi=()=>Math.floor(Math.random()*(yi-Xe)+Xe);let _i={...{apiGateway:a0,requestTimeout:1e4,defaultSystemEventPrefix:"[TS]",defaultSystemEventDataSeparator:" | "}};const bi=()=>_i;class mi{constructor(e){i0(this,"uiScope","ton-connect-ui-");i0(this,"sdkScope","ton-connect-");i0(this,"tonConnectEvents",[J.WalletConnectSuccess,J.WalletConnectError,J.ConnectionRestoringSuccess,J.ConnectionRestoringError,J.TransactionSentForSignature,J.TransactionSigned,J.TransactionSigningFailed,J.WalletDisconnect]);i0(this,"tonConnectUiEvents",[J.WalletConnectStarted,J.WalletConnectError,J.TransactionSigningFailed]);this.eventBuilder=e,this.initialize()}initialize(){this.tonConnectUiEvents.forEach(e=>{const t=`${this.uiScope}${e}`;window.addEventListener(t,r=>{n0.debug(`UI Event received: ${t}`,{detail:r.detail,raw:r}),r.detail&&this.dispatchEvent(e,r.detail).catch(n=>{n0.error("Failed to dispatch UI event",{event:e,error:n instanceof Error?n.message:String(n)})})})}),this.tonConnectEvents.forEach(e=>{const t=`${this.sdkScope}${e}`;window.addEventListener(t,r=>{n0.debug(`SDK Event received: ${t}`,{detail:r.detail,raw:r}),r.detail&&this.dispatchEvent(e,r.detail).catch(n=>{n0.error("Failed to dispatch SDK event",{event:e,error:n instanceof Error?n.message:String(n)})})})}),n0.info("TonConnect observer initialized")}async dispatchEvent(e,t){var s,x,o,f,a,c,E,h,u,v;const r=bi(),n=localStorage.getItem("telemetree-ton-is-connected")==="true";n0.debug("Raw event details:",{eventName:e,eventDetails:t});try{switch(e){case J.WalletConnectSuccess:{if(!n){localStorage.setItem("telemetree-ton-is-connected","true");const p=t,l=((s=p==null?void 0:p.account)==null?void 0:s.address)||(p==null?void 0:p.wallet_address);await this.eventBuilder.track(`${q.Wallet}`,{wallet:l,provider:((x=p==null?void 0:p.wallet)==null?void 0:x.name)||"unknown",chain:(o=p==null?void 0:p.account)==null?void 0:o.chain,wallet_type:t.wallet_type,wallet_version:t.wallet_version,sdk_version:(f=t.custom_data)==null?void 0:f.ton_connect_sdk_lib,ui_version:(a=t.custom_data)==null?void 0:a.ton_connect_ui_lib,timestamp:Date.now()})}break}case J.TransactionSentForSignature:{await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSentForSignature}`,{messages:(c=t.messages)==null?void 0:c.map(p=>({address:p.address,amount:p.amount,payload:p.payload})),from:t.from,network:t.network,validUntil:t.valid_until,wallet_type:t.wallet_type,wallet_version:t.wallet_version,chain_id:(E=t.custom_data)==null?void 0:E.chain_id,provider:(h=t.custom_data)==null?void 0:h.provider,sdk_version:(u=t.custom_data)==null?void 0:u.ton_connect_sdk_lib,ui_version:(v=t.custom_data)==null?void 0:v.ton_connect_ui_lib,timestamp:Date.now()});break}case J.TransactionSigned:{const p=t;await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSigned}`,{boc:p==null?void 0:p.boc,from:t.from,messages:t.messages,valid_until:t.valid_until,auth_type:t.auth_type,wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet_address:t.wallet_address,timestamp:Date.now()});break}case J.TransactionSigningFailed:{await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSigningFailed}`,{error_code:t.error_code,error_message:t.error_message,is_success:t.is_success,from:t.from,messages:t.messages,valid_until:t.valid_until,auth_type:t.auth_type,wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet_address:t.wallet_address,timestamp:Date.now()});break}case J.WalletDisconnect:{n&&localStorage.setItem("telemetree-ton-is-connected","false"),await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletDisconnected}`,{lastWallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringError:{await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestoreError}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringSuccess:{await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestored}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringStarted:{await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestoringStarted}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}}}catch(p){n0.error("Error dispatching event",{eventName:e,error:p instanceof Error?p.message:String(p),rawDetails:t})}}destroy(){const e=new Map;this.tonConnectUiEvents.forEach(t=>{const r=`${this.uiScope}${t}`,n=e.get(r);n&&window.removeEventListener(r,n)}),this.tonConnectEvents.forEach(t=>{const r=`${this.sdkScope}${t}`,n=e.get(r);n&&window.removeEventListener(r,n)})}}const Si="ton-connect-storage_bridge-connection",wi="ton-connect-ui_preferred-wallet";return i=>{var x;if(n0.setLevel(i.logLevel||Kt.INFO),!i.projectId)throw new Error("TWA Analytics Provider: Missing projectId");let e;if(i.isTelegramContext){const o=Ei();e=new Ze(i.projectId,i.apiKey,o)}else{let o=Di("telemetree.id");o===null&&(o=Fi().toString(),Ci("telemetree.id",o)),e=new Ze(i.projectId,i.apiKey,{auth_date:+Gt(),hash:"",user:{id:+o,first_name:o,username:"",last_name:"",language_code:"",is_premium:!1},platform:"web"})}I==null||I.onEvent("main_button_pressed",o=>{e.track(q.MainButtonPressed,{})}),I==null||I.onEvent("settings_button_pressed",o=>{e.track(q.SettingsButtonPressed,{})}),I==null||I.onEvent("back_button_pressed",o=>{e.track(q.BackButtonPressed,{})}),I==null||I.onEvent("secondary_button_pressed",(o,f)=>{e.track(q.SecondaryButtonPressed,{...f})}),I==null||I.onEvent("prepared_message_sent",o=>{e.track(q.PreparedMessageSent,{})}),I==null||I.onEvent("fullscreen_changed",(o,f)=>{e.track(q.FullScreenChanged,{...f})}),I==null||I.onEvent("home_screen_added",o=>{e.track(q.HomeScreenAdded,{})}),I==null||I.onEvent("home_screen_checked",(o,f)=>{e.track(q.HomeScreenChecked,{...f})}),I==null||I.onEvent("emoji_status_set",o=>{e.track(q.EmojiStatusSet,{})}),I==null||I.onEvent("location_checked",(o,f)=>{e.track(q.LocationChecked,{...f})}),I==null||I.onEvent("location_requested",(o,f)=>{e.track(q.LocationRequested,{...f})}),I==null||I.onEvent("accelerometer_started",o=>{e.track(q.AccelerometerStarted,{})}),I==null||I.onEvent("accelerometer_stopped",o=>{e.track(q.AccelerometerStopped,{})}),I==null||I.onEvent("accelerometer_changed",o=>{e.track(q.AccelerometerChanged,{})}),I==null||I.onEvent("device_orientation_started",o=>{e.track(q.DeviceOrientationStarted,{})}),I==null||I.onEvent("device_orientation_stopped",o=>{e.track(q.DeviceOrientationStopped,{})}),I==null||I.onEvent("device_orientation_changed",o=>{e.track(q.DeviceOrientationChanged,{})}),I==null||I.onEvent("device_orientation_failed",o=>{e.track(q.DeviceOrientationFailed,{})}),I==null||I.onEvent("gyroscope_started",o=>{e.track(q.GyroscropeStarted,{})}),I==null||I.onEvent("gyroscope_stopped",o=>{e.track(q.GyroscropeStopped,{})}),I==null||I.onEvent("gyroscope_changed",o=>{e.track(q.GyroscropeChanged,{})}),I==null||I.onEvent("gyroscope_failed",o=>{e.track(q.GyroscropeFailed,{})}),I==null||I.onEvent("invoice_closed",(o,f)=>{e.track(q.InvoiceClosed,{...f})}),I==null||I.onEvent("clipboard_text_received",(o,f)=>{e.track(q.ClipboardTextReceived,{...f})}),I==null||I.onEvent("popup_closed",(o,f)=>{e.track(q.PopupClosed,{})}),I==null||I.onEvent("write_access_requested",(o,f)=>{e.track(q.WriteAccessRequested,{})}),I==null||I.onEvent("qr_text_received",(o,f)=>{e.track(q.QRTextReceived,{...f})}),I==null||I.onEvent("phone_requested",(o,f)=>{e.track(q.PhoneRequested,{...f})}),I==null||I.onEvent("web_app_request_fullscreen",o=>{e.track(q.WebAppRequestFullscreen,{timestamp:Date.now()})}),I==null||I.onEvent("web_app_exit_fullscreen",o=>{e.track(q.WebAppExitFullscreen,{timestamp:Date.now()})});const t=(x=window==null?void 0:window.Telegram)==null?void 0:x.WebApp;if(t){if(!window.__telemetreeSessionStarted){window.__telemetreeSessionStarted=!0,e.track(q.SessionStart,{timestamp:Date.now(),platform:t.platform,version:t.version,colorScheme:t.colorScheme,viewportHeight:t.viewportHeight,viewportStableHeight:t.viewportStableHeight,isExpanded:t.isExpanded});const l=location.pathname||"/";e.track(`${q.PageView} ${l}`,{path:l,timestamp:Date.now()})}const o=t.close,f=t.switchInlineQuery,a=t.openInvoice,c=t.shareToStory,E=t.openLink,h=t.openTelegramLink;t.close=()=>{e.track(q.SessionEnd,{timestamp:Date.now(),platform:t.platform,version:t.version,colorScheme:t.colorScheme,viewportHeight:t.viewportHeight,viewportStableHeight:t.viewportStableHeight,isExpanded:t.isExpanded}),setTimeout(()=>{o.call(t)},50)},t.switchInlineQuery=(l,C)=>(e.track(`${q.SwitchInlineQuery}: ${l}`,{query:l,chat_types:C||[],timestamp:Date.now()}),f.call(t,l,C)),t.openInvoice=(l,C)=>{const d=l.split("/").pop()||"";return e.track(`${q.InvoiceOpened}: ${d}`,{url:l,slug:d,timestamp:Date.now()}),a.call(t,l,C)};let u=t.viewportHeight,v=!1;const p=l=>{const C=t.viewportHeight>u;C&&!v?(v=!0,e.track(q.WebAppRequestFullscreen,{timestamp:Date.now(),isStateStable:l.isStateStable,previousViewportHeight:u,newViewportHeight:t.viewportHeight,viewportStableHeight:t.viewportStableHeight})):!C&&v&&(v=!1,e.track(q.WebAppExitFullscreen,{timestamp:Date.now(),isStateStable:l.isStateStable,previousViewportHeight:u,newViewportHeight:t.viewportHeight,viewportStableHeight:t.viewportStableHeight})),u=t.viewportHeight};t.onEvent("viewportChanged",p),t.shareToStory=(l,C)=>(e.track(`${q.ShareToStory}: ${l}`,{media_url:l,text:C==null?void 0:C.text,timestamp:Date.now()}),c.call(t,l,C)),t.openLink=(l,C)=>(e.track(`${q.OpenLink}: ${l}`,{url:l,options:C,timestamp:Date.now()}),E.call(t,l,C)),t.openTelegramLink=(l,C)=>(e.track(`${q.OpenTgLink}: ${l}`,{url:l,options:C,timestamp:Date.now()}),h.call(t,l,C))}const r=location.pathname;e.track(`${q.PageView} ${r}`,{path:r});let n=null;setInterval(()=>{var f,a;const o=localStorage.getItem(Si);if(o)try{const E=((a=(f=JSON.parse(o).connectEvent)==null?void 0:f.payload)==null?void 0:a.items)||[];if(E&&E.length===0||n===E[0].address)return;const h=localStorage.getItem(wi),u={wallet:E[0].address,walletProvider:h||"unknown"};n=E[0].address}catch(c){console.error(c)}},1e3);let s=null;try{s=new mi(e),n0.info("TON Connect observer initialized successfully")}catch(o){n0.error("Failed to initialize TON Connect observer",{error:o instanceof Error?o.message:"Unknown error"})}return e}});
+`,t+="-----END PUBLIC KEY-----",t},e.hasPublicKeyProperty=function(t){return t=t||{},t.hasOwnProperty("n")&&t.hasOwnProperty("e")},e.hasPrivateKeyProperty=function(t){return t=t||{},t.hasOwnProperty("n")&&t.hasOwnProperty("e")&&t.hasOwnProperty("d")&&t.hasOwnProperty("p")&&t.hasOwnProperty("q")&&t.hasOwnProperty("dmp1")&&t.hasOwnProperty("dmq1")&&t.hasOwnProperty("coeff")},e.prototype.parsePropertiesFrom=function(t){this.n=t.n,this.e=t.e,t.hasOwnProperty("d")&&(this.d=t.d,this.p=t.p,this.q=t.q,this.dmp1=t.dmp1,this.dmq1=t.dmq1,this.coeff=t.coeff)},e}(ci),Kt,vi=typeof process<"u"?(Kt=process.env)===null||Kt===void 0?void 0:Kt.npm_package_version:void 0,pi=function(){function i(e){e===void 0&&(e={}),e=e||{},this.default_key_size=e.default_key_size?parseInt(e.default_key_size,10):1024,this.default_public_exponent=e.default_public_exponent||"010001",this.log=e.log||!1,this.key=null}return i.prototype.setKey=function(e){this.log&&this.key&&console.warn("A key was already set, overriding existing."),this.key=new Ge(e)},i.prototype.setPrivateKey=function(e){this.setKey(e)},i.prototype.setPublicKey=function(e){this.setKey(e)},i.prototype.decrypt=function(e){try{return this.getKey().decrypt(qe(e))}catch{return!1}},i.prototype.encrypt=function(e){try{return V0(this.getKey().encrypt(e))}catch{return!1}},i.prototype.sign=function(e,t,r){try{return V0(this.getKey().sign(e,t,r))}catch{return!1}},i.prototype.verify=function(e,t,r){try{return this.getKey().verify(e,qe(t),r)}catch{return!1}},i.prototype.getKey=function(e){if(!this.key){if(this.key=new Ge,e&&{}.toString.call(e)==="[object Function]"){this.key.generateAsync(this.default_key_size,this.default_public_exponent,e);return}this.key.generate(this.default_key_size,this.default_public_exponent)}return this.key},i.prototype.getPrivateKey=function(){return this.getKey().getPrivateKey()},i.prototype.getPrivateKeyB64=function(){return this.getKey().getPrivateBaseKeyB64()},i.prototype.getPublicKey=function(){return this.getKey().getPublicKey()},i.prototype.getPublicKeyB64=function(){return this.getKey().getPublicBaseKeyB64()},i.version=vi,i}();const $e=(i,e)=>{const t=new pi;return t.setPublicKey(i),t.encrypt(e)},gi=()=>{const i=zt.lib.WordArray.random(16),e=zt.lib.WordArray.random(16);return{key:i,iv:e}},Ei=({key:i,iv:e},t)=>zt.AES.encrypt(t,i,{iv:e}).toString(),Bi=(i,e)=>{const{key:t,iv:r}=gi(),n=t.toString(),s=r.toString(),x=$e(i,n);if(x===!1)throw new Error("Failed to encrypt AES key.");const o=$e(i,s);if(o===!1)throw new Error("Failed to encrypt AES IV.");const f=Ei({key:t,iv:r},e);return{encryptedKey:x,encryptedIV:o,encryptedBody:f}};class je{constructor(e){if(this._options=e,!("fetch"in window))throw new Error("Fetch is not available in this browser.");this._options.headers||(this._options.headers={})}setOptions(e){this._options=e}send(e,t,r){const n={method:t,headers:this._options.headers,mode:"cors",redirect:"follow"};return t==="POST"&&r&&(n.body=r),new Promise((s,x)=>{window.fetch(e,n).then(o=>{s(o)}).catch(o=>{x(o)})})}}class Ze{static getTransport(e,t){return this.transports[e]?new this.transports[e](t):new je(t)}}i0(Ze,"transports",{http:je});const Ci=(i,e,t,r,n,s,x,o,f,a,c,B,h="telemetree_twa")=>({eventType:i,userDetails:e,eventDetails:t,telegramID:r,language:n,device:s,referrerType:x,referrer:o,timestamp:f,isAutocapture:a,wallet:c,sessionIdentifier:B,eventSource:h});var Gt=(i=>(i.ERROR="error",i.WARN="warn",i.INFO="info",i.DEBUG="debug",i))(Gt||{});class n0{static setLevel(e){this.level=e}static log(e,t,r){if(this.shouldLog(e)){const n=r?`${t} ${JSON.stringify(r)}`:t;console[e](n)}}static shouldLog(e){const t=Object.values(Gt);return t.indexOf(e)<=t.indexOf(this.level)}static error(e,t){this.log("error",e,t)}static warn(e,t){this.log("warn",e,t)}static info(e,t){this.log("info",e,t)}static debug(e,t){this.log("debug",e,t)}}i0(n0,"level","info");const Y0=typeof window<"u"&&((tr=window==null?void 0:window.Telegram)!=null&&tr.WebApp)?window.Telegram.WebApp:null,O=typeof window<"u"&&((er=window==null?void 0:window.Telegram)!=null&&er.WebView)?window.Telegram.WebView:null,Di=()=>Y0===null?yi():Ai(),Ai=()=>{if(Y0===null)throw new Error("Telegram WebApp is not loaded.");return{...Y0.initDataUnsafe,platform:Y0.platform}},yi=()=>{if(document.location.hash==="")throw new Error("Telegram WebApp is not loaded.");const e=decodeURIComponent(decodeURIComponent(document.location.hash.substring(1))),t=new URLSearchParams(e),r=t.get("tgWebAppData");if(r===null)throw new Error("Telegram WebApp is not loaded.");const n=new URLSearchParams(r);return{user:n.get("user")?JSON.parse(n.get("user")):void 0,chat_type:t.get("chat_type"),chat_instance:t.get("chat_instance")||void 0,platform:t.get("tgWebAppPlatform")||"unknown",auth_date:t.get("auth_date")?parseInt(t.get("auth_date")):0,hash:t.get("hash")||"",start_param:t.get("start_param")||t.get("startapp")||void 0}};var u0=(i=>(i.LOW="low",i.MEDIUM="medium",i.HIGH="high",i))(u0||{});const Xe=(i,e)=>{if(O==null||O.onEvent("invoice_closed",(r,n)=>{i.track(q.InvoiceClosed,{...n})}),e){if(!window.__telemetreeSessionStarted){window.__telemetreeSessionStarted=!0,i.track(q.SessionStart,{timestamp:Date.now(),platform:e.platform,version:e.version,colorScheme:e.colorScheme,viewportHeight:e.viewportHeight,viewportStableHeight:e.viewportStableHeight,isExpanded:e.isExpanded});const n=location.pathname||"/";i.track(`${q.PageView} ${n}`,{path:n,timestamp:Date.now()})}const r=e.openInvoice;e.openInvoice=(n,s)=>{const x=n.split("/").pop()||"";return i.track(`${q.InvoiceOpened}: ${x}`,{url:n,slug:x,timestamp:Date.now()}),r.call(e,n,s)}}const t=location.pathname;i.track(`${q.PageView} ${t}`,{path:t})},Qe=(i,e)=>{if(Xe(i,e),e){const t=e.openLink,r=e.openTelegramLink;e.openLink=(n,s)=>(i.track(`${q.OpenLink}: ${n}`,{url:n,options:s,timestamp:Date.now()}),t.call(e,n,s)),e.openTelegramLink=(n,s)=>(i.track(`${q.OpenTgLink}: ${n}`,{url:n,options:s,timestamp:Date.now()}),r.call(e,n,s))}},Fi=(i,e)=>{if(Qe(i,e),O==null||O.onEvent("main_button_pressed",t=>{i.track(q.MainButtonPressed,{})}),O==null||O.onEvent("settings_button_pressed",t=>{i.track(q.SettingsButtonPressed,{})}),O==null||O.onEvent("back_button_pressed",t=>{i.track(q.BackButtonPressed,{})}),O==null||O.onEvent("secondary_button_pressed",(t,r)=>{i.track(q.SecondaryButtonPressed,{...r})}),O==null||O.onEvent("prepared_message_sent",t=>{i.track(q.PreparedMessageSent,{})}),O==null||O.onEvent("fullscreen_changed",(t,r)=>{i.track(q.FullScreenChanged,{...r})}),O==null||O.onEvent("home_screen_added",t=>{i.track(q.HomeScreenAdded,{})}),O==null||O.onEvent("home_screen_checked",(t,r)=>{i.track(q.HomeScreenChecked,{...r})}),O==null||O.onEvent("emoji_status_set",t=>{i.track(q.EmojiStatusSet,{})}),O==null||O.onEvent("location_checked",(t,r)=>{i.track(q.LocationChecked,{...r})}),O==null||O.onEvent("location_requested",(t,r)=>{i.track(q.LocationRequested,{...r})}),O==null||O.onEvent("accelerometer_started",t=>{i.track(q.AccelerometerStarted,{})}),O==null||O.onEvent("accelerometer_stopped",t=>{i.track(q.AccelerometerStopped,{})}),O==null||O.onEvent("accelerometer_changed",t=>{i.track(q.AccelerometerChanged,{})}),O==null||O.onEvent("device_orientation_started",t=>{i.track(q.DeviceOrientationStarted,{})}),O==null||O.onEvent("device_orientation_stopped",t=>{i.track(q.DeviceOrientationStopped,{})}),O==null||O.onEvent("device_orientation_changed",t=>{i.track(q.DeviceOrientationChanged,{})}),O==null||O.onEvent("device_orientation_failed",t=>{i.track(q.DeviceOrientationFailed,{})}),O==null||O.onEvent("gyroscope_started",t=>{i.track(q.GyroscropeStarted,{})}),O==null||O.onEvent("gyroscope_stopped",t=>{i.track(q.GyroscropeStopped,{})}),O==null||O.onEvent("gyroscope_changed",t=>{i.track(q.GyroscropeChanged,{})}),O==null||O.onEvent("gyroscope_failed",t=>{i.track(q.GyroscropeFailed,{})}),O==null||O.onEvent("clipboard_text_received",(t,r)=>{i.track(q.ClipboardTextReceived,{...r})}),O==null||O.onEvent("popup_closed",(t,r)=>{i.track(q.PopupClosed,{})}),O==null||O.onEvent("write_access_requested",(t,r)=>{i.track(q.WriteAccessRequested,{})}),O==null||O.onEvent("qr_text_received",(t,r)=>{i.track(q.QRTextReceived,{...r})}),O==null||O.onEvent("phone_requested",(t,r)=>{i.track(q.PhoneRequested,{...r})}),O==null||O.onEvent("web_app_request_fullscreen",t=>{i.track(q.WebAppRequestFullscreen,{timestamp:Date.now()})}),O==null||O.onEvent("web_app_exit_fullscreen",t=>{i.track(q.WebAppExitFullscreen,{timestamp:Date.now()})}),e){const t=e.close,r=e.switchInlineQuery,n=e.shareToStory;e.close=()=>{i.track(q.SessionEnd,{timestamp:Date.now(),platform:e.platform,version:e.version,colorScheme:e.colorScheme,viewportHeight:e.viewportHeight,viewportStableHeight:e.viewportStableHeight,isExpanded:e.isExpanded}),setTimeout(()=>{t.call(e)},50)},e.switchInlineQuery=(f,a)=>(i.track(`${q.SwitchInlineQuery}: ${f}`,{query:f,chat_types:a||[],timestamp:Date.now()}),r.call(e,f,a));let s=e.viewportHeight,x=!1;const o=f=>{const a=e.viewportHeight>s;a&&!x?(x=!0,i.track(q.WebAppRequestFullscreen,{timestamp:Date.now(),isStateStable:f.isStateStable,previousViewportHeight:s,newViewportHeight:e.viewportHeight,viewportStableHeight:e.viewportStableHeight})):!a&&x&&(x=!1,i.track(q.WebAppExitFullscreen,{timestamp:Date.now(),isStateStable:f.isStateStable,previousViewportHeight:s,newViewportHeight:e.viewportHeight,viewportStableHeight:e.viewportStableHeight})),s=e.viewportHeight};e.onEvent("viewportChanged",o),e.shareToStory=(f,a)=>(i.track(`${q.ShareToStory}: ${f}`,{media_url:f,text:a==null?void 0:a.text,timestamp:Date.now()}),n.call(e,f,a))}};class Ye{constructor(e,t,r,n=null){i0(this,"transport",null);i0(this,"config",null);i0(this,"sessionIdentifier",null);i0(this,"userIdentifier",null);i0(this,"pushHandler",new vr(this));this.projectId=e,this.apiKey=t,this.data=r,this.trackGroup=n,this._init()}setSessionIdentifier(e){return this.sessionIdentifier=e,this}getTrackGroup(){return this.trackGroup}setUserId(e){return this.userIdentifier=e,this}async _init(){const e=Ze.getTransport("http",{headers:{authorization:`Bearer ${this.apiKey}`},requestTimeout:1e3});this.setSessionIdentifier(Zt());try{const t=await e.send(a0+`?project=${this.projectId}`,"GET");if(t.status===200){const r=await t.json();this.config=r}e.setOptions({headers:{"x-api-key":`${this.apiKey}`,"x-project-id":`${this.projectId}`,"content-type":"application/json"},requestTimeout:1500}),this.setTransport(e),await this.pushHandler.flush(),this.setupAutoCaptureListener()}catch(t){console.error(`Cannot load config: ${t}`)}}setupAutoCaptureListener(){var t;const e=this.getConfig();if(e!==null&&e.auto_capture===!0&&this.trackGroup==u0.MEDIUM){const r=e.auto_capture_tags.map(n=>n.toUpperCase());(t=document.querySelector("body"))==null||t.addEventListener("click",n=>{var o,f,a,c,B,h;const x=(((o=n.composedPath)==null?void 0:o.call(n))||[]).find(u=>u instanceof HTMLElement&&r.includes(u.tagName));if(x){const u={},v=["id","href","class","name","value","type","placeholder","title","alt","src"];for(const A of v)x.hasAttribute(A)&&(u[A]=x.getAttribute(A));const g=((a=(f=n.target)==null?void 0:f.innerText)==null?void 0:a.trim())||"",l=((c=x.innerText)==null?void 0:c.trim())||"";u.text=g||l,u.tag=x.tagName.toLowerCase(),n.target!==x&&(u.clicked_child=(h=(B=n.target)==null?void 0:B.tagName)==null?void 0:h.toLowerCase(),u.clicked_child_text=g),this.track(`${p0} ${q.Click}${dr}${u.text}`,u)}})}}getConfig(){return this.config}setTransport(e){return this.transport=e,this}async track(e,t){var n,s,x,o,f,a,c;if(e===null)throw new Error("Event name is not set.");if(!((n=this.data.user)!=null&&n.id)){console.error(`Event ${e} is not tracked because user is not set.`);return}const r=Ci(e,{username:((s=this.data.user)==null?void 0:s.username)||"",firstName:this.data.user.first_name,lastName:((x=this.data.user)==null?void 0:x.last_name)||"",isPremium:((o=this.data.user)==null?void 0:o.is_premium)||!1,writeAccess:((f=this.data.user)==null?void 0:f.allows_write_to_pm)||!1},{startParameter:this.data.start_param||"",path:document.location.pathname,params:t},(a=this.data.user)==null?void 0:a.id.toString(),(c=this.data.user)==null?void 0:c.language_code,this.data.platform,this.data.chat_type||"N/A",this.data.chat_instance||"0",pr(),e.startsWith(p0),t.wallet||void 0,this.sessionIdentifier||void 0);return this.pushHandler.push(r)}async processEvent(e){if(this.config===null||this.transport===null){n0.info("Event processing failed: config or transport is not set.");return}try{const{encryptedKey:t,encryptedIV:r,encryptedBody:n}=Bi(this.config.public_key,JSON.stringify(e));await this.transport.send(this.config.host,"POST",JSON.stringify({key:t,iv:r,body:n})),n0.debug("Event sent successfully",{eventData:e})}catch(t){console.error(t)}}}const _i=(i,e)=>{localStorage.setItem(i,e)},mi=i=>localStorage.getItem(i),bi=1e12,Je=1,Si=()=>Math.floor(Math.random()*(bi-Je)+Je);let wi={...{apiGateway:a0,requestTimeout:1e4,defaultSystemEventPrefix:"[TS]",defaultSystemEventDataSeparator:" | "}};const Ti=()=>wi;class ki{constructor(e){i0(this,"uiScope","ton-connect-ui-");i0(this,"sdkScope","ton-connect-");i0(this,"tonConnectEvents",[J.WalletConnectSuccess,J.WalletConnectError,J.ConnectionRestoringSuccess,J.ConnectionRestoringError,J.TransactionSentForSignature,J.TransactionSigned,J.TransactionSigningFailed,J.WalletDisconnect]);i0(this,"tonConnectUiEvents",[J.WalletConnectStarted,J.WalletConnectError,J.TransactionSigningFailed]);this.eventBuilder=e,this.initialize()}initialize(){this.tonConnectUiEvents.forEach(e=>{const t=`${this.uiScope}${e}`;window.addEventListener(t,r=>{n0.debug(`UI Event received: ${t}`,{detail:r.detail,raw:r}),r.detail&&this.dispatchEvent(e,r.detail).catch(n=>{n0.error("Failed to dispatch UI event",{event:e,error:n instanceof Error?n.message:String(n)})})})}),this.tonConnectEvents.forEach(e=>{const t=`${this.sdkScope}${e}`;window.addEventListener(t,r=>{n0.debug(`SDK Event received: ${t}`,{detail:r.detail,raw:r}),r.detail&&this.dispatchEvent(e,r.detail).catch(n=>{n0.error("Failed to dispatch SDK event",{event:e,error:n instanceof Error?n.message:String(n)})})})}),n0.info("TonConnect observer initialized")}async dispatchEvent(e,t){var x,o,f,a,c,B,h,u,v,g;const r=Ti(),n=localStorage.getItem("telemetree-ton-is-connected")==="true",s=this.eventBuilder.getTrackGroup();n0.debug("Raw event details:",{eventName:e,eventDetails:t});try{switch(e){case J.WalletConnectSuccess:{if(!n){localStorage.setItem("telemetree-ton-is-connected","true");const l=t,A=((x=l==null?void 0:l.account)==null?void 0:x.address)||(l==null?void 0:l.wallet_address);await this.eventBuilder.track(`${q.Wallet}`,{wallet:A,provider:((o=l==null?void 0:l.wallet)==null?void 0:o.name)||"unknown",chain:(f=l==null?void 0:l.account)==null?void 0:f.chain,wallet_type:t.wallet_type,wallet_version:t.wallet_version,sdk_version:(a=t.custom_data)==null?void 0:a.ton_connect_sdk_lib,ui_version:(c=t.custom_data)==null?void 0:c.ton_connect_ui_lib,timestamp:Date.now()})}break}case J.TransactionSentForSignature:{s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSentForSignature}`,{messages:(B=t.messages)==null?void 0:B.map(l=>({address:l.address,amount:l.amount,payload:l.payload})),from:t.from,network:t.network,validUntil:t.valid_until,wallet_type:t.wallet_type,wallet_version:t.wallet_version,chain_id:(h=t.custom_data)==null?void 0:h.chain_id,provider:(u=t.custom_data)==null?void 0:u.provider,sdk_version:(v=t.custom_data)==null?void 0:v.ton_connect_sdk_lib,ui_version:(g=t.custom_data)==null?void 0:g.ton_connect_ui_lib,timestamp:Date.now()});break}case J.TransactionSigned:{const l=t;await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSigned}`,{boc:l==null?void 0:l.boc,from:t.from,messages:t.messages,valid_until:t.valid_until,auth_type:t.auth_type,wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet_address:t.wallet_address,timestamp:Date.now()});break}case J.TransactionSigningFailed:{s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.TransactionSigningFailed}`,{error_code:t.error_code,error_message:t.error_message,is_success:t.is_success,from:t.from,messages:t.messages,valid_until:t.valid_until,auth_type:t.auth_type,wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet_address:t.wallet_address,timestamp:Date.now()});break}case J.WalletDisconnect:{n&&localStorage.setItem("telemetree-ton-is-connected","false"),s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletDisconnected}`,{lastWallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringError:{s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestoreError}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringSuccess:{s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestored}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}case J.ConnectionRestoringStarted:{s===u0.HIGH&&await this.eventBuilder.track(`${r.defaultSystemEventPrefix} ${q.WalletConnectionRestoringStarted}`,{wallet_type:t.wallet_type,wallet_version:t.wallet_version,wallet:t==null?void 0:t.wallet_address,timestamp:Date.now()});break}}}catch(l){n0.error("Error dispatching event",{eventName:e,error:l instanceof Error?l.message:String(l),rawDetails:t})}}destroy(){const e=new Map;this.tonConnectUiEvents.forEach(t=>{const r=`${this.uiScope}${t}`,n=e.get(r);n&&window.removeEventListener(r,n)}),this.tonConnectEvents.forEach(t=>{const r=`${this.sdkScope}${t}`,n=e.get(r);n&&window.removeEventListener(r,n)})}}const Ri="ton-connect-storage_bridge-connection",Ii="ton-connect-ui_preferred-wallet";return i=>{var t;if(n0.setLevel(i.logLevel||Gt.INFO),!i.projectId)throw new Error("TWA Analytics Provider: Missing projectId");let e;if(i.isTelegramContext){const r=Di();e=new Ye(i.projectId,i.apiKey,r,i.trackGroup)}else{let r=mi("telemetree.id");r===null&&(r=Si().toString(),_i("telemetree.id",r)),e=new Ye(i.projectId,i.apiKey,{auth_date:+Zt(),hash:"",user:{id:+r,first_name:r,username:"",last_name:"",language_code:"",is_premium:!1},platform:"web"},i.trackGroup)}if(i.trackGroup){const r=(t=window==null?void 0:window.Telegram)==null?void 0:t.WebApp;i.trackGroup===u0.LOW&&Xe(e,r),i.trackGroup===u0.MEDIUM&&Qe(e,r),i.trackGroup===u0.HIGH&&Fi(e,r);let n=null;setInterval(()=>{var o,f;const x=localStorage.getItem(Ri);if(x)try{const c=((f=(o=JSON.parse(x).connectEvent)==null?void 0:o.payload)==null?void 0:f.items)||[];if(c&&c.length===0||n===c[0].address)return;const B=localStorage.getItem(Ii),h={wallet:c[0].address,walletProvider:B||"unknown"};n=c[0].address}catch(a){console.error(a)}},1e3);let s=null;if(i.trackGroup)try{s=new ki(e),n0.info("TON Connect observer initialized successfully")}catch(x){n0.error("Failed to initialize TON Connect observer",{error:x instanceof Error?x.message:"Unknown error"})}}return e}});
